chore: remove standalone ContactScout app, consolidate into InviteFlow scout
