+++ index.html
@@ -172,15 +172,6 @@ body {
       <div class="ac-desc">Compose and send personalized invite emails via Gmail. Import contacts, track send status, and follow up on RSVPs.</div>
       <div class="ac-cta">Open InviteFlow →</div>
     </a>
-    <a class="app-card" href="contactscout.html">
-      <div class="ac-top">
-        <span class="ac-name">ContactScout</span>
-        <span class="badge badge-blue">Claude AI</span>
-      </div>
-      <div class="ac-file">contactscout.html</div>
-      <div class="ac-desc">Discover and verify elected officials using Claude AI with web search. Export contact lists directly into InviteFlow.</div>
-      <div class="ac-cta">Open ContactScout →</div>
-    </a>
   </div>
 
   <div class="footer">
