--- src/contactscout/built/assets/index-CzzvB3Ea.js
@@ -1,48 +1,0 @@
-(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const l of document.querySelectorAll('link[rel="modulepreload"]'))r(l);new MutationObserver(l=>{for(const i of l)if(i.type==="childList")for(const o of i.addedNodes)o.tagName==="LINK"&&o.rel==="modulepreload"&&r(o)}).observe(document,{childList:!0,subtree:!0});function n(l){const i={};return l.integrity&&(i.integrity=l.integrity),l.referrerPolicy&&(i.referrerPolicy=l.referrerPolicy),l.crossOrigin==="use-credentials"?i.credentials="include":l.crossOrigin==="anonymous"?i.credentials="omit":i.credentials="same-origin",i}function r(l){if(l.ep)return;l.ep=!0;const i=n(l);fetch(l.href,i)}})();var us={exports:{}},yl={},ss={exports:{}},R={};/**
- * @license React
- * react.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var dr=Symbol.for("react.element"),Oc=Symbol.for("react.portal"),Ic=Symbol.for("react.fragment"),Rc=Symbol.for("react.strict_mode"),Dc=Symbol.for("react.profiler"),Fc=Symbol.for("react.provider"),Mc=Symbol.for("react.context"),Ac=Symbol.for("react.forward_ref"),Uc=Symbol.for("react.suspense"),$c=Symbol.for("react.memo"),Bc=Symbol.for("react.lazy"),Jo=Symbol.iterator;function Hc(e){return e===null||typeof e!="object"?null:(e=Jo&&e[Jo]||e["@@iterator"],typeof e=="function"?e:null)}var as={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},cs=Object.assign,fs={};function kn(e,t,n){this.props=e,this.context=t,this.refs=fs,this.updater=n||as}kn.prototype.isReactComponent={};kn.prototype.setState=function(e,t){if(typeof e!="object"&&typeof e!="function"&&e!=null)throw Error("setState(...): takes an object of state variables to update or a function which returns an object of state variables.");this.updater.enqueueSetState(this,e,t,"setState")};kn.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")};function ds(){}ds.prototype=kn.prototype;function eo(e,t,n){this.props=e,this.context=t,this.refs=fs,this.updater=n||as}var to=eo.prototype=new ds;to.constructor=eo;cs(to,kn.prototype);to.isPureReactComponent=!0;var Xo=Array.isArray,ps=Object.prototype.hasOwnProperty,no={current:null},ms={key:!0,ref:!0,__self:!0,__source:!0};function hs(e,t,n){var r,l={},i=null,o=null;if(t!=null)for(r in t.ref!==void 0&&(o=t.ref),t.key!==void 0&&(i=""+t.key),t)ps.call(t,r)&&!ms.hasOwnProperty(r)&&(l[r]=t[r]);var u=arguments.length-2;if(u===1)l.children=n;else if(1<u){for(var s=Array(u),d=0;d<u;d++)s[d]=arguments[d+2];l.children=s}if(e&&e.defaultProps)for(r in u=e.defaultProps,u)l[r]===void 0&&(l[r]=u[r]);return{$$typeof:dr,type:e,key:i,ref:o,props:l,_owner:no.current}}function Vc(e,t){return{$$typeof:dr,type:e.type,key:t,ref:e.ref,props:e.props,_owner:e._owner}}function ro(e){return typeof e=="object"&&e!==null&&e.$$typeof===dr}function Wc(e){var t={"=":"=0",":":"=2"};return"$"+e.replace(/[=:]/g,function(n){return t[n]})}var Zo=/\/+/g;function Il(e,t){return typeof e=="object"&&e!==null&&e.key!=null?Wc(""+e.key):t.toString(36)}function Dr(e,t,n,r,l){var i=typeof e;(i==="undefined"||i==="boolean")&&(e=null);var o=!1;if(e===null)o=!0;else switch(i){case"string":case"number":o=!0;break;case"object":switch(e.$$typeof){case dr:case Oc:o=!0}}if(o)return o=e,l=l(o),e=r===""?"."+Il(o,0):r,Xo(l)?(n="",e!=null&&(n=e.replace(Zo,"$&/")+"/"),Dr(l,t,n,"",function(d){return d})):l!=null&&(ro(l)&&(l=Vc(l,n+(!l.key||o&&o.key===l.key?"":(""+l.key).replace(Zo,"$&/")+"/")+e)),t.push(l)),1;if(o=0,r=r===""?".":r+":",Xo(e))for(var u=0;u<e.length;u++){i=e[u];var s=r+Il(i,u);o+=Dr(i,t,n,s,l)}else if(s=Hc(e),typeof s=="function")for(e=s.call(e),u=0;!(i=e.next()).done;)i=i.value,s=r+Il(i,u++),o+=Dr(i,t,n,s,l);else if(i==="object")throw t=String(e),Error("Objects are not valid as a React child (found: "+(t==="[object Object]"?"object with keys {"+Object.keys(e).join(", ")+"}":t)+"). If you meant to render a collection of children, use an array instead.");return o}function gr(e,t,n){if(e==null)return e;var r=[],l=0;return Dr(e,r,"","",function(i){return t.call(n,i,l++)}),r}function Qc(e){if(e._status===-1){var t=e._result;t=t(),t.then(function(n){(e._status===0||e._status===-1)&&(e._status=1,e._result=n)},function(n){(e._status===0||e._status===-1)&&(e._status=2,e._result=n)}),e._status===-1&&(e._status=0,e._result=t)}if(e._status===1)return e._result.default;throw e._result}var he={current:null},Fr={transition:null},Kc={ReactCurrentDispatcher:he,ReactCurrentBatchConfig:Fr,ReactCurrentOwner:no};function vs(){throw Error("act(...) is not supported in production builds of React.")}R.Children={map:gr,forEach:function(e,t,n){gr(e,function(){t.apply(this,arguments)},n)},count:function(e){var t=0;return gr(e,function(){t++}),t},toArray:function(e){return gr(e,function(t){return t})||[]},only:function(e){if(!ro(e))throw Error("React.Children.only expected to receive a single React element child.");return e}};R.Component=kn;R.Fragment=Ic;R.Profiler=Dc;R.PureComponent=eo;R.StrictMode=Rc;R.Suspense=Uc;R.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=Kc;R.act=vs;R.cloneElement=function(e,t,n){if(e==null)throw Error("React.cloneElement(...): The argument must be a React element, but you passed "+e+".");var r=cs({},e.props),l=e.key,i=e.ref,o=e._owner;if(t!=null){if(t.ref!==void 0&&(i=t.ref,o=no.current),t.key!==void 0&&(l=""+t.key),e.type&&e.type.defaultProps)var u=e.type.defaultProps;for(s in t)ps.call(t,s)&&!ms.hasOwnProperty(s)&&(r[s]=t[s]===void 0&&u!==void 0?u[s]:t[s])}var s=arguments.length-2;if(s===1)r.children=n;else if(1<s){u=Array(s);for(var d=0;d<s;d++)u[d]=arguments[d+2];r.children=u}return{$$typeof:dr,type:e.type,key:l,ref:i,props:r,_owner:o}};R.createContext=function(e){return e={$$typeof:Mc,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null,_defaultValue:null,_globalName:null},e.Provider={$$typeof:Fc,_context:e},e.Consumer=e};R.createElement=hs;R.createFactory=function(e){var t=hs.bind(null,e);return t.type=e,t};R.createRef=function(){return{current:null}};R.forwardRef=function(e){return{$$typeof:Ac,render:e}};R.isValidElement=ro;R.lazy=function(e){return{$$typeof:Bc,_payload:{_status:-1,_result:e},_init:Qc}};R.memo=function(e,t){return{$$typeof:$c,type:e,compare:t===void 0?null:t}};R.startTransition=function(e){var t=Fr.transition;Fr.transition={};try{e()}finally{Fr.transition=t}};R.unstable_act=vs;R.useCallback=function(e,t){return he.current.useCallback(e,t)};R.useContext=function(e){return he.current.useContext(e)};R.useDebugValue=function(){};R.useDeferredValue=function(e){return he.current.useDeferredValue(e)};R.useEffect=function(e,t){return he.current.useEffect(e,t)};R.useId=function(){return he.current.useId()};R.useImperativeHandle=function(e,t,n){return he.current.useImperativeHandle(e,t,n)};R.useInsertionEffect=function(e,t){return he.current.useInsertionEffect(e,t)};R.useLayoutEffect=function(e,t){return he.current.useLayoutEffect(e,t)};R.useMemo=function(e,t){return he.current.useMemo(e,t)};R.useReducer=function(e,t,n){return he.current.useReducer(e,t,n)};R.useRef=function(e){return he.current.useRef(e)};R.useState=function(e){return he.current.useState(e)};R.useSyncExternalStore=function(e,t,n){return he.current.useSyncExternalStore(e,t,n)};R.useTransition=function(){return he.current.useTransition()};R.version="18.3.1";ss.exports=R;var U=ss.exports;/**
- * @license React
- * react-jsx-runtime.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var Yc=U,Gc=Symbol.for("react.element"),Jc=Symbol.for("react.fragment"),Xc=Object.prototype.hasOwnProperty,Zc=Yc.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.ReactCurrentOwner,qc={key:!0,ref:!0,__self:!0,__source:!0};function ys(e,t,n){var r,l={},i=null,o=null;n!==void 0&&(i=""+n),t.key!==void 0&&(i=""+t.key),t.ref!==void 0&&(o=t.ref);for(r in t)Xc.call(t,r)&&!qc.hasOwnProperty(r)&&(l[r]=t[r]);if(e&&e.defaultProps)for(r in t=e.defaultProps,t)l[r]===void 0&&(l[r]=t[r]);return{$$typeof:Gc,type:e,key:i,ref:o,props:l,_owner:Zc.current}}yl.Fragment=Jc;yl.jsx=ys;yl.jsxs=ys;us.exports=yl;var f=us.exports,gs={exports:{}},je={},Ss={exports:{}},ws={};/**
- * @license React
- * scheduler.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */(function(e){function t(x,z){var O=x.length;x.push(z);e:for(;0<O;){var B=O-1>>>1,Z=x[B];if(0<l(Z,z))x[B]=z,x[O]=Z,O=B;else break e}}function n(x){return x.length===0?null:x[0]}function r(x){if(x.length===0)return null;var z=x[0],O=x.pop();if(O!==z){x[0]=O;e:for(var B=0,Z=x.length,Pt=Z>>>1;B<Pt;){var Ge=2*(B+1)-1,_n=x[Ge],Je=Ge+1,Yt=x[Je];if(0>l(_n,O))Je<Z&&0>l(Yt,_n)?(x[B]=Yt,x[Je]=O,B=Je):(x[B]=_n,x[Ge]=O,B=Ge);else if(Je<Z&&0>l(Yt,O))x[B]=Yt,x[Je]=O,B=Je;else break e}}return z}function l(x,z){var O=x.sortIndex-z.sortIndex;return O!==0?O:x.id-z.id}if(typeof performance=="object"&&typeof performance.now=="function"){var i=performance;e.unstable_now=function(){return i.now()}}else{var o=Date,u=o.now();e.unstable_now=function(){return o.now()-u}}var s=[],d=[],y=1,h=null,m=3,w=!1,k=!1,v=!1,D=typeof setTimeout=="function"?setTimeout:null,a=typeof clearTimeout=="function"?clearTimeout:null,c=typeof setImmediate<"u"?setImmediate:null;typeof navigator<"u"&&navigator.scheduling!==void 0&&navigator.scheduling.isInputPending!==void 0&&navigator.scheduling.isInputPending.bind(navigator.scheduling);function p(x){for(var z=n(d);z!==null;){if(z.callback===null)r(d);else if(z.startTime<=x)r(d),z.sortIndex=z.expirationTime,t(s,z);else break;z=n(d)}}function g(x){if(v=!1,p(x),!k)if(n(s)!==null)k=!0,ot(E);else{var z=n(d);z!==null&&Kt(g,z.startTime-x)}}function E(x,z){k=!1,v&&(v=!1,a(T),T=-1),w=!0;var O=m;try{for(p(z),h=n(s);h!==null&&(!(h.expirationTime>z)||x&&!ke());){var B=h.callback;if(typeof B=="function"){h.callback=null,m=h.priorityLevel;var Z=B(h.expirationTime<=z);z=e.unstable_now(),typeof Z=="function"?h.callback=Z:h===n(s)&&r(s),p(z)}else r(s);h=n(s)}if(h!==null)var Pt=!0;else{var Ge=n(d);Ge!==null&&Kt(g,Ge.startTime-z),Pt=!1}return Pt}finally{h=null,m=O,w=!1}}var j=!1,_=null,T=-1,K=5,I=-1;function ke(){return!(e.unstable_now()-I<K)}function He(){if(_!==null){var x=e.unstable_now();I=x;var z=!0;try{z=_(!0,x)}finally{z?jt():(j=!1,_=null)}}else j=!1}var jt;if(typeof c=="function")jt=function(){c(He)};else if(typeof MessageChannel<"u"){var Nn=new MessageChannel,Qt=Nn.port2;Nn.port1.onmessage=He,jt=function(){Qt.postMessage(null)}}else jt=function(){D(He,0)};function ot(x){_=x,j||(j=!0,jt())}function Kt(x,z){T=D(function(){x(e.unstable_now())},z)}e.unstable_IdlePriority=5,e.unstable_ImmediatePriority=1,e.unstable_LowPriority=4,e.unstable_NormalPriority=3,e.unstable_Profiling=null,e.unstable_UserBlockingPriority=2,e.unstable_cancelCallback=function(x){x.callback=null},e.unstable_continueExecution=function(){k||w||(k=!0,ot(E))},e.unstable_forceFrameRate=function(x){0>x||125<x?console.error("forceFrameRate takes a positive int between 0 and 125, forcing frame rates higher than 125 fps is not supported"):K=0<x?Math.floor(1e3/x):5},e.unstable_getCurrentPriorityLevel=function(){return m},e.unstable_getFirstCallbackNode=function(){return n(s)},e.unstable_next=function(x){switch(m){case 1:case 2:case 3:var z=3;break;default:z=m}var O=m;m=z;try{return x()}finally{m=O}},e.unstable_pauseExecution=function(){},e.unstable_requestPaint=function(){},e.unstable_runWithPriority=function(x,z){switch(x){case 1:case 2:case 3:case 4:case 5:break;default:x=3}var O=m;m=x;try{return z()}finally{m=O}},e.unstable_scheduleCallback=function(x,z,O){var B=e.unstable_now();switch(typeof O=="object"&&O!==null?(O=O.delay,O=typeof O=="number"&&0<O?B+O:B):O=B,x){case 1:var Z=-1;break;case 2:Z=250;break;case 5:Z=1073741823;break;case 4:Z=1e4;break;default:Z=5e3}return Z=O+Z,x={id:y++,callback:z,priorityLevel:x,startTime:O,expirationTime:Z,sortIndex:-1},O>B?(x.sortIndex=O,t(d,x),n(s)===null&&x===n(d)&&(v?(a(T),T=-1):v=!0,Kt(g,O-B))):(x.sortIndex=Z,t(s,x),k||w||(k=!0,ot(E))),x},e.unstable_shouldYield=ke,e.unstable_wrapCallback=function(x){var z=m;return function(){var O=m;m=z;try{return x.apply(this,arguments)}finally{m=O}}}})(ws);Ss.exports=ws;var bc=Ss.exports;/**
- * @license React
- * react-dom.production.min.js
- *
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */var ef=U,_e=bc;function S(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var xs=new Set,Jn={};function Vt(e,t){hn(e,t),hn(e+"Capture",t)}function hn(e,t){for(Jn[e]=t,e=0;e<t.length;e++)xs.add(t[e])}var tt=!(typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"),ui=Object.prototype.hasOwnProperty,tf=/^[:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD][:A-Z_a-z\u00C0-\u00D6\u00D8-\u00F6\u00F8-\u02FF\u0370-\u037D\u037F-\u1FFF\u200C-\u200D\u2070-\u218F\u2C00-\u2FEF\u3001-\uD7FF\uF900-\uFDCF\uFDF0-\uFFFD\-.0-9\u00B7\u0300-\u036F\u203F-\u2040]*$/,qo={},bo={};function nf(e){return ui.call(bo,e)?!0:ui.call(qo,e)?!1:tf.test(e)?bo[e]=!0:(qo[e]=!0,!1)}function rf(e,t,n,r){if(n!==null&&n.type===0)return!1;switch(typeof t){case"function":case"symbol":return!0;case"boolean":return r?!1:n!==null?!n.acceptsBooleans:(e=e.toLowerCase().slice(0,5),e!=="data-"&&e!=="aria-");default:return!1}}function lf(e,t,n,r){if(t===null||typeof t>"u"||rf(e,t,n,r))return!0;if(r)return!1;if(n!==null)switch(n.type){case 3:return!t;case 4:return t===!1;case 5:return isNaN(t);case 6:return isNaN(t)||1>t}return!1}function ve(e,t,n,r,l,i,o){this.acceptsBooleans=t===2||t===3||t===4,this.attributeName=r,this.attributeNamespace=l,this.mustUseProperty=n,this.propertyName=e,this.type=t,this.sanitizeURL=i,this.removeEmptyString=o}var se={};"children dangerouslySetInnerHTML defaultValue defaultChecked innerHTML suppressContentEditableWarning suppressHydrationWarning style".split(" ").forEach(function(e){se[e]=new ve(e,0,!1,e,null,!1,!1)});[["acceptCharset","accept-charset"],["className","class"],["htmlFor","for"],["httpEquiv","http-equiv"]].forEach(function(e){var t=e[0];se[t]=new ve(t,1,!1,e[1],null,!1,!1)});["contentEditable","draggable","spellCheck","value"].forEach(function(e){se[e]=new ve(e,2,!1,e.toLowerCase(),null,!1,!1)});["autoReverse","externalResourcesRequired","focusable","preserveAlpha"].forEach(function(e){se[e]=new ve(e,2,!1,e,null,!1,!1)});"allowFullScreen async autoFocus autoPlay controls default defer disabled disablePictureInPicture disableRemotePlayback formNoValidate hidden loop noModule noValidate open playsInline readOnly required reversed scoped seamless itemScope".split(" ").forEach(function(e){se[e]=new ve(e,3,!1,e.toLowerCase(),null,!1,!1)});["checked","multiple","muted","selected"].forEach(function(e){se[e]=new ve(e,3,!0,e,null,!1,!1)});["capture","download"].forEach(function(e){se[e]=new ve(e,4,!1,e,null,!1,!1)});["cols","rows","size","span"].forEach(function(e){se[e]=new ve(e,6,!1,e,null,!1,!1)});["rowSpan","start"].forEach(function(e){se[e]=new ve(e,5,!1,e.toLowerCase(),null,!1,!1)});var lo=/[\-:]([a-z])/g;function io(e){return e[1].toUpperCase()}"accent-height alignment-baseline arabic-form baseline-shift cap-height clip-path clip-rule color-interpolation color-interpolation-filters color-profile color-rendering dominant-baseline enable-background fill-opacity fill-rule flood-color flood-opacity font-family font-size font-size-adjust font-stretch font-style font-variant font-weight glyph-name glyph-orientation-horizontal glyph-orientation-vertical horiz-adv-x horiz-origin-x image-rendering letter-spacing lighting-color marker-end marker-mid marker-start overline-position overline-thickness paint-order panose-1 pointer-events rendering-intent shape-rendering stop-color stop-opacity strikethrough-position strikethrough-thickness stroke-dasharray stroke-dashoffset stroke-linecap stroke-linejoin stroke-miterlimit stroke-opacity stroke-width text-anchor text-decoration text-rendering underline-position underline-thickness unicode-bidi unicode-range units-per-em v-alphabetic v-hanging v-ideographic v-mathematical vector-effect vert-adv-y vert-origin-x vert-origin-y word-spacing writing-mode xmlns:xlink x-height".split(" ").forEach(function(e){var t=e.replace(lo,io);se[t]=new ve(t,1,!1,e,null,!1,!1)});"xlink:actuate xlink:arcrole xlink:role xlink:show xlink:title xlink:type".split(" ").forEach(function(e){var t=e.replace(lo,io);se[t]=new ve(t,1,!1,e,"http://www.w3.org/1999/xlink",!1,!1)});["xml:base","xml:lang","xml:space"].forEach(function(e){var t=e.replace(lo,io);se[t]=new ve(t,1,!1,e,"http://www.w3.org/XML/1998/namespace",!1,!1)});["tabIndex","crossOrigin"].forEach(function(e){se[e]=new ve(e,1,!1,e.toLowerCase(),null,!1,!1)});se.xlinkHref=new ve("xlinkHref",1,!1,"xlink:href","http://www.w3.org/1999/xlink",!0,!1);["src","href","action","formAction"].forEach(function(e){se[e]=new ve(e,1,!1,e.toLowerCase(),null,!0,!0)});function oo(e,t,n,r){var l=se.hasOwnProperty(t)?se[t]:null;(l!==null?l.type!==0:r||!(2<t.length)||t[0]!=="o"&&t[0]!=="O"||t[1]!=="n"&&t[1]!=="N")&&(lf(t,n,l,r)&&(n=null),r||l===null?nf(t)&&(n===null?e.removeAttribute(t):e.setAttribute(t,""+n)):l.mustUseProperty?e[l.propertyName]=n===null?l.type===3?!1:"":n:(t=l.attributeName,r=l.attributeNamespace,n===null?e.removeAttribute(t):(l=l.type,n=l===3||l===4&&n===!0?"":""+n,r?e.setAttributeNS(r,t,n):e.setAttribute(t,n))))}var it=ef.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,Sr=Symbol.for("react.element"),Xt=Symbol.for("react.portal"),Zt=Symbol.for("react.fragment"),uo=Symbol.for("react.strict_mode"),si=Symbol.for("react.profiler"),ks=Symbol.for("react.provider"),Cs=Symbol.for("react.context"),so=Symbol.for("react.forward_ref"),ai=Symbol.for("react.suspense"),ci=Symbol.for("react.suspense_list"),ao=Symbol.for("react.memo"),st=Symbol.for("react.lazy"),Es=Symbol.for("react.offscreen"),eu=Symbol.iterator;function Pn(e){return e===null||typeof e!="object"?null:(e=eu&&e[eu]||e["@@iterator"],typeof e=="function"?e:null)}var J=Object.assign,Rl;function Fn(e){if(Rl===void 0)try{throw Error()}catch(n){var t=n.stack.trim().match(/\n( *(at )?)/);Rl=t&&t[1]||""}return`
-`+Rl+e}var Dl=!1;function Fl(e,t){if(!e||Dl)return"";Dl=!0;var n=Error.prepareStackTrace;Error.prepareStackTrace=void 0;try{if(t)if(t=function(){throw Error()},Object.defineProperty(t.prototype,"props",{set:function(){throw Error()}}),typeof Reflect=="object"&&Reflect.construct){try{Reflect.construct(t,[])}catch(d){var r=d}Reflect.construct(e,[],t)}else{try{t.call()}catch(d){r=d}e.call(t.prototype)}else{try{throw Error()}catch(d){r=d}e()}}catch(d){if(d&&r&&typeof d.stack=="string"){for(var l=d.stack.split(`
-`),i=r.stack.split(`
-`),o=l.length-1,u=i.length-1;1<=o&&0<=u&&l[o]!==i[u];)u--;for(;1<=o&&0<=u;o--,u--)if(l[o]!==i[u]){if(o!==1||u!==1)do if(o--,u--,0>u||l[o]!==i[u]){var s=`
-`+l[o].replace(" at new "," at ");return e.displayName&&s.includes("<anonymous>")&&(s=s.replace("<anonymous>",e.displayName)),s}while(1<=o&&0<=u);break}}}finally{Dl=!1,Error.prepareStackTrace=n}return(e=e?e.displayName||e.name:"")?Fn(e):""}function of(e){switch(e.tag){case 5:return Fn(e.type);case 16:return Fn("Lazy");case 13:return Fn("Suspense");case 19:return Fn("SuspenseList");case 0:case 2:case 15:return e=Fl(e.type,!1),e;case 11:return e=Fl(e.type.render,!1),e;case 1:return e=Fl(e.type,!0),e;default:return""}}function fi(e){if(e==null)return null;if(typeof e=="function")return e.displayName||e.name||null;if(typeof e=="string")return e;switch(e){case Zt:return"Fragment";case Xt:return"Portal";case si:return"Profiler";case uo:return"StrictMode";case ai:return"Suspense";case ci:return"SuspenseList"}if(typeof e=="object")switch(e.$$typeof){case Cs:return(e.displayName||"Context")+".Consumer";case ks:return(e._context.displayName||"Context")+".Provider";case so:var t=e.render;return e=e.displayName,e||(e=t.displayName||t.name||"",e=e!==""?"ForwardRef("+e+")":"ForwardRef"),e;case ao:return t=e.displayName||null,t!==null?t:fi(e.type)||"Memo";case st:t=e._payload,e=e._init;try{return fi(e(t))}catch{}}return null}function uf(e){var t=e.type;switch(e.tag){case 24:return"Cache";case 9:return(t.displayName||"Context")+".Consumer";case 10:return(t._context.displayName||"Context")+".Provider";case 18:return"DehydratedFragment";case 11:return e=t.render,e=e.displayName||e.name||"",t.displayName||(e!==""?"ForwardRef("+e+")":"ForwardRef");case 7:return"Fragment";case 5:return t;case 4:return"Portal";case 3:return"Root";case 6:return"Text";case 16:return fi(t);case 8:return t===uo?"StrictMode":"Mode";case 22:return"Offscreen";case 12:return"Profiler";case 21:return"Scope";case 13:return"Suspense";case 19:return"SuspenseList";case 25:return"TracingMarker";case 1:case 0:case 17:case 2:case 14:case 15:if(typeof t=="function")return t.displayName||t.name||null;if(typeof t=="string")return t}return null}function kt(e){switch(typeof e){case"boolean":case"number":case"string":case"undefined":return e;case"object":return e;default:return""}}function Ns(e){var t=e.type;return(e=e.nodeName)&&e.toLowerCase()==="input"&&(t==="checkbox"||t==="radio")}function sf(e){var t=Ns(e)?"checked":"value",n=Object.getOwnPropertyDescriptor(e.constructor.prototype,t),r=""+e[t];if(!e.hasOwnProperty(t)&&typeof n<"u"&&typeof n.get=="function"&&typeof n.set=="function"){var l=n.get,i=n.set;return Object.defineProperty(e,t,{configurable:!0,get:function(){return l.call(this)},set:function(o){r=""+o,i.call(this,o)}}),Object.defineProperty(e,t,{enumerable:n.enumerable}),{getValue:function(){return r},setValue:function(o){r=""+o},stopTracking:function(){e._valueTracker=null,delete e[t]}}}}function wr(e){e._valueTracker||(e._valueTracker=sf(e))}function _s(e){if(!e)return!1;var t=e._valueTracker;if(!t)return!0;var n=t.getValue(),r="";return e&&(r=Ns(e)?e.checked?"true":"false":e.value),e=r,e!==n?(t.setValue(e),!0):!1}function Yr(e){if(e=e||(typeof document<"u"?document:void 0),typeof e>"u")return null;try{return e.activeElement||e.body}catch{return e.body}}function di(e,t){var n=t.checked;return J({},t,{defaultChecked:void 0,defaultValue:void 0,value:void 0,checked:n??e._wrapperState.initialChecked})}function tu(e,t){var n=t.defaultValue==null?"":t.defaultValue,r=t.checked!=null?t.checked:t.defaultChecked;n=kt(t.value!=null?t.value:n),e._wrapperState={initialChecked:r,initialValue:n,controlled:t.type==="checkbox"||t.type==="radio"?t.checked!=null:t.value!=null}}function js(e,t){t=t.checked,t!=null&&oo(e,"checked",t,!1)}function pi(e,t){js(e,t);var n=kt(t.value),r=t.type;if(n!=null)r==="number"?(n===0&&e.value===""||e.value!=n)&&(e.value=""+n):e.value!==""+n&&(e.value=""+n);else if(r==="submit"||r==="reset"){e.removeAttribute("value");return}t.hasOwnProperty("value")?mi(e,t.type,n):t.hasOwnProperty("defaultValue")&&mi(e,t.type,kt(t.defaultValue)),t.checked==null&&t.defaultChecked!=null&&(e.defaultChecked=!!t.defaultChecked)}function nu(e,t,n){if(t.hasOwnProperty("value")||t.hasOwnProperty("defaultValue")){var r=t.type;if(!(r!=="submit"&&r!=="reset"||t.value!==void 0&&t.value!==null))return;t=""+e._wrapperState.initialValue,n||t===e.value||(e.value=t),e.defaultValue=t}n=e.name,n!==""&&(e.name=""),e.defaultChecked=!!e._wrapperState.initialChecked,n!==""&&(e.name=n)}function mi(e,t,n){(t!=="number"||Yr(e.ownerDocument)!==e)&&(n==null?e.defaultValue=""+e._wrapperState.initialValue:e.defaultValue!==""+n&&(e.defaultValue=""+n))}var Mn=Array.isArray;function an(e,t,n,r){if(e=e.options,t){t={};for(var l=0;l<n.length;l++)t["$"+n[l]]=!0;for(n=0;n<e.length;n++)l=t.hasOwnProperty("$"+e[n].value),e[n].selected!==l&&(e[n].selected=l),l&&r&&(e[n].defaultSelected=!0)}else{for(n=""+kt(n),t=null,l=0;l<e.length;l++){if(e[l].value===n){e[l].selected=!0,r&&(e[l].defaultSelected=!0);return}t!==null||e[l].disabled||(t=e[l])}t!==null&&(t.selected=!0)}}function hi(e,t){if(t.dangerouslySetInnerHTML!=null)throw Error(S(91));return J({},t,{value:void 0,defaultValue:void 0,children:""+e._wrapperState.initialValue})}function ru(e,t){var n=t.value;if(n==null){if(n=t.children,t=t.defaultValue,n!=null){if(t!=null)throw Error(S(92));if(Mn(n)){if(1<n.length)throw Error(S(93));n=n[0]}t=n}t==null&&(t=""),n=t}e._wrapperState={initialValue:kt(n)}}function Ps(e,t){var n=kt(t.value),r=kt(t.defaultValue);n!=null&&(n=""+n,n!==e.value&&(e.value=n),t.defaultValue==null&&e.defaultValue!==n&&(e.defaultValue=n)),r!=null&&(e.defaultValue=""+r)}function lu(e){var t=e.textContent;t===e._wrapperState.initialValue&&t!==""&&t!==null&&(e.value=t)}function Ts(e){switch(e){case"svg":return"http://www.w3.org/2000/svg";case"math":return"http://www.w3.org/1998/Math/MathML";default:return"http://www.w3.org/1999/xhtml"}}function vi(e,t){return e==null||e==="http://www.w3.org/1999/xhtml"?Ts(t):e==="http://www.w3.org/2000/svg"&&t==="foreignObject"?"http://www.w3.org/1999/xhtml":e}var xr,zs=function(e){return typeof MSApp<"u"&&MSApp.execUnsafeLocalFunction?function(t,n,r,l){MSApp.execUnsafeLocalFunction(function(){return e(t,n,r,l)})}:e}(function(e,t){if(e.namespaceURI!=="http://www.w3.org/2000/svg"||"innerHTML"in e)e.innerHTML=t;else{for(xr=xr||document.createElement("div"),xr.innerHTML="<svg>"+t.valueOf().toString()+"</svg>",t=xr.firstChild;e.firstChild;)e.removeChild(e.firstChild);for(;t.firstChild;)e.appendChild(t.firstChild)}});function Xn(e,t){if(t){var n=e.firstChild;if(n&&n===e.lastChild&&n.nodeType===3){n.nodeValue=t;return}}e.textContent=t}var $n={animationIterationCount:!0,aspectRatio:!0,borderImageOutset:!0,borderImageSlice:!0,borderImageWidth:!0,boxFlex:!0,boxFlexGroup:!0,boxOrdinalGroup:!0,columnCount:!0,columns:!0,flex:!0,flexGrow:!0,flexPositive:!0,flexShrink:!0,flexNegative:!0,flexOrder:!0,gridArea:!0,gridRow:!0,gridRowEnd:!0,gridRowSpan:!0,gridRowStart:!0,gridColumn:!0,gridColumnEnd:!0,gridColumnSpan:!0,gridColumnStart:!0,fontWeight:!0,lineClamp:!0,lineHeight:!0,opacity:!0,order:!0,orphans:!0,tabSize:!0,widows:!0,zIndex:!0,zoom:!0,fillOpacity:!0,floodOpacity:!0,stopOpacity:!0,strokeDasharray:!0,strokeDashoffset:!0,strokeMiterlimit:!0,strokeOpacity:!0,strokeWidth:!0},af=["Webkit","ms","Moz","O"];Object.keys($n).forEach(function(e){af.forEach(function(t){t=t+e.charAt(0).toUpperCase()+e.substring(1),$n[t]=$n[e]})});function Ls(e,t,n){return t==null||typeof t=="boolean"||t===""?"":n||typeof t!="number"||t===0||$n.hasOwnProperty(e)&&$n[e]?(""+t).trim():t+"px"}function Os(e,t){e=e.style;for(var n in t)if(t.hasOwnProperty(n)){var r=n.indexOf("--")===0,l=Ls(n,t[n],r);n==="float"&&(n="cssFloat"),r?e.setProperty(n,l):e[n]=l}}var cf=J({menuitem:!0},{area:!0,base:!0,br:!0,col:!0,embed:!0,hr:!0,img:!0,input:!0,keygen:!0,link:!0,meta:!0,param:!0,source:!0,track:!0,wbr:!0});function yi(e,t){if(t){if(cf[e]&&(t.children!=null||t.dangerouslySetInnerHTML!=null))throw Error(S(137,e));if(t.dangerouslySetInnerHTML!=null){if(t.children!=null)throw Error(S(60));if(typeof t.dangerouslySetInnerHTML!="object"||!("__html"in t.dangerouslySetInnerHTML))throw Error(S(61))}if(t.style!=null&&typeof t.style!="object")throw Error(S(62))}}function gi(e,t){if(e.indexOf("-")===-1)return typeof t.is=="string";switch(e){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var Si=null;function co(e){return e=e.target||e.srcElement||window,e.correspondingUseElement&&(e=e.correspondingUseElement),e.nodeType===3?e.parentNode:e}var wi=null,cn=null,fn=null;function iu(e){if(e=hr(e)){if(typeof wi!="function")throw Error(S(280));var t=e.stateNode;t&&(t=kl(t),wi(e.stateNode,e.type,t))}}function Is(e){cn?fn?fn.push(e):fn=[e]:cn=e}function Rs(){if(cn){var e=cn,t=fn;if(fn=cn=null,iu(e),t)for(e=0;e<t.length;e++)iu(t[e])}}function Ds(e,t){return e(t)}function Fs(){}var Ml=!1;function Ms(e,t,n){if(Ml)return e(t,n);Ml=!0;try{return Ds(e,t,n)}finally{Ml=!1,(cn!==null||fn!==null)&&(Fs(),Rs())}}function Zn(e,t){var n=e.stateNode;if(n===null)return null;var r=kl(n);if(r===null)return null;n=r[t];e:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(r=!r.disabled)||(e=e.type,r=!(e==="button"||e==="input"||e==="select"||e==="textarea")),e=!r;break e;default:e=!1}if(e)return null;if(n&&typeof n!="function")throw Error(S(231,t,typeof n));return n}var xi=!1;if(tt)try{var Tn={};Object.defineProperty(Tn,"passive",{get:function(){xi=!0}}),window.addEventListener("test",Tn,Tn),window.removeEventListener("test",Tn,Tn)}catch{xi=!1}function ff(e,t,n,r,l,i,o,u,s){var d=Array.prototype.slice.call(arguments,3);try{t.apply(n,d)}catch(y){this.onError(y)}}var Bn=!1,Gr=null,Jr=!1,ki=null,df={onError:function(e){Bn=!0,Gr=e}};function pf(e,t,n,r,l,i,o,u,s){Bn=!1,Gr=null,ff.apply(df,arguments)}function mf(e,t,n,r,l,i,o,u,s){if(pf.apply(this,arguments),Bn){if(Bn){var d=Gr;Bn=!1,Gr=null}else throw Error(S(198));Jr||(Jr=!0,ki=d)}}function Wt(e){var t=e,n=e;if(e.alternate)for(;t.return;)t=t.return;else{e=t;do t=e,t.flags&4098&&(n=t.return),e=t.return;while(e)}return t.tag===3?n:null}function As(e){if(e.tag===13){var t=e.memoizedState;if(t===null&&(e=e.alternate,e!==null&&(t=e.memoizedState)),t!==null)return t.dehydrated}return null}function ou(e){if(Wt(e)!==e)throw Error(S(188))}function hf(e){var t=e.alternate;if(!t){if(t=Wt(e),t===null)throw Error(S(188));return t!==e?null:e}for(var n=e,r=t;;){var l=n.return;if(l===null)break;var i=l.alternate;if(i===null){if(r=l.return,r!==null){n=r;continue}break}if(l.child===i.child){for(i=l.child;i;){if(i===n)return ou(l),e;if(i===r)return ou(l),t;i=i.sibling}throw Error(S(188))}if(n.return!==r.return)n=l,r=i;else{for(var o=!1,u=l.child;u;){if(u===n){o=!0,n=l,r=i;break}if(u===r){o=!0,r=l,n=i;break}u=u.sibling}if(!o){for(u=i.child;u;){if(u===n){o=!0,n=i,r=l;break}if(u===r){o=!0,r=i,n=l;break}u=u.sibling}if(!o)throw Error(S(189))}}if(n.alternate!==r)throw Error(S(190))}if(n.tag!==3)throw Error(S(188));return n.stateNode.current===n?e:t}function Us(e){return e=hf(e),e!==null?$s(e):null}function $s(e){if(e.tag===5||e.tag===6)return e;for(e=e.child;e!==null;){var t=$s(e);if(t!==null)return t;e=e.sibling}return null}var Bs=_e.unstable_scheduleCallback,uu=_e.unstable_cancelCallback,vf=_e.unstable_shouldYield,yf=_e.unstable_requestPaint,q=_e.unstable_now,gf=_e.unstable_getCurrentPriorityLevel,fo=_e.unstable_ImmediatePriority,Hs=_e.unstable_UserBlockingPriority,Xr=_e.unstable_NormalPriority,Sf=_e.unstable_LowPriority,Vs=_e.unstable_IdlePriority,gl=null,Ke=null;function wf(e){if(Ke&&typeof Ke.onCommitFiberRoot=="function")try{Ke.onCommitFiberRoot(gl,e,void 0,(e.current.flags&128)===128)}catch{}}var Ue=Math.clz32?Math.clz32:Cf,xf=Math.log,kf=Math.LN2;function Cf(e){return e>>>=0,e===0?32:31-(xf(e)/kf|0)|0}var kr=64,Cr=4194304;function An(e){switch(e&-e){case 1:return 1;case 2:return 2;case 4:return 4;case 8:return 8;case 16:return 16;case 32:return 32;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return e&4194240;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return e&130023424;case 134217728:return 134217728;case 268435456:return 268435456;case 536870912:return 536870912;case 1073741824:return 1073741824;default:return e}}function Zr(e,t){var n=e.pendingLanes;if(n===0)return 0;var r=0,l=e.suspendedLanes,i=e.pingedLanes,o=n&268435455;if(o!==0){var u=o&~l;u!==0?r=An(u):(i&=o,i!==0&&(r=An(i)))}else o=n&~l,o!==0?r=An(o):i!==0&&(r=An(i));if(r===0)return 0;if(t!==0&&t!==r&&!(t&l)&&(l=r&-r,i=t&-t,l>=i||l===16&&(i&4194240)!==0))return t;if(r&4&&(r|=n&16),t=e.entangledLanes,t!==0)for(e=e.entanglements,t&=r;0<t;)n=31-Ue(t),l=1<<n,r|=e[n],t&=~l;return r}function Ef(e,t){switch(e){case 1:case 2:case 4:return t+250;case 8:case 16:case 32:case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:return t+5e3;case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:return-1;case 134217728:case 268435456:case 536870912:case 1073741824:return-1;default:return-1}}function Nf(e,t){for(var n=e.suspendedLanes,r=e.pingedLanes,l=e.expirationTimes,i=e.pendingLanes;0<i;){var o=31-Ue(i),u=1<<o,s=l[o];s===-1?(!(u&n)||u&r)&&(l[o]=Ef(u,t)):s<=t&&(e.expiredLanes|=u),i&=~u}}function Ci(e){return e=e.pendingLanes&-1073741825,e!==0?e:e&1073741824?1073741824:0}function Ws(){var e=kr;return kr<<=1,!(kr&4194240)&&(kr=64),e}function Al(e){for(var t=[],n=0;31>n;n++)t.push(e);return t}function pr(e,t,n){e.pendingLanes|=t,t!==536870912&&(e.suspendedLanes=0,e.pingedLanes=0),e=e.eventTimes,t=31-Ue(t),e[t]=n}function _f(e,t){var n=e.pendingLanes&~t;e.pendingLanes=t,e.suspendedLanes=0,e.pingedLanes=0,e.expiredLanes&=t,e.mutableReadLanes&=t,e.entangledLanes&=t,t=e.entanglements;var r=e.eventTimes;for(e=e.expirationTimes;0<n;){var l=31-Ue(n),i=1<<l;t[l]=0,r[l]=-1,e[l]=-1,n&=~i}}function po(e,t){var n=e.entangledLanes|=t;for(e=e.entanglements;n;){var r=31-Ue(n),l=1<<r;l&t|e[r]&t&&(e[r]|=t),n&=~l}}var A=0;function Qs(e){return e&=-e,1<e?4<e?e&268435455?16:536870912:4:1}var Ks,mo,Ys,Gs,Js,Ei=!1,Er=[],mt=null,ht=null,vt=null,qn=new Map,bn=new Map,ct=[],jf="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset submit".split(" ");function su(e,t){switch(e){case"focusin":case"focusout":mt=null;break;case"dragenter":case"dragleave":ht=null;break;case"mouseover":case"mouseout":vt=null;break;case"pointerover":case"pointerout":qn.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":bn.delete(t.pointerId)}}function zn(e,t,n,r,l,i){return e===null||e.nativeEvent!==i?(e={blockedOn:t,domEventName:n,eventSystemFlags:r,nativeEvent:i,targetContainers:[l]},t!==null&&(t=hr(t),t!==null&&mo(t)),e):(e.eventSystemFlags|=r,t=e.targetContainers,l!==null&&t.indexOf(l)===-1&&t.push(l),e)}function Pf(e,t,n,r,l){switch(t){case"focusin":return mt=zn(mt,e,t,n,r,l),!0;case"dragenter":return ht=zn(ht,e,t,n,r,l),!0;case"mouseover":return vt=zn(vt,e,t,n,r,l),!0;case"pointerover":var i=l.pointerId;return qn.set(i,zn(qn.get(i)||null,e,t,n,r,l)),!0;case"gotpointercapture":return i=l.pointerId,bn.set(i,zn(bn.get(i)||null,e,t,n,r,l)),!0}return!1}function Xs(e){var t=It(e.target);if(t!==null){var n=Wt(t);if(n!==null){if(t=n.tag,t===13){if(t=As(n),t!==null){e.blockedOn=t,Js(e.priority,function(){Ys(n)});return}}else if(t===3&&n.stateNode.current.memoizedState.isDehydrated){e.blockedOn=n.tag===3?n.stateNode.containerInfo:null;return}}}e.blockedOn=null}function Mr(e){if(e.blockedOn!==null)return!1;for(var t=e.targetContainers;0<t.length;){var n=Ni(e.domEventName,e.eventSystemFlags,t[0],e.nativeEvent);if(n===null){n=e.nativeEvent;var r=new n.constructor(n.type,n);Si=r,n.target.dispatchEvent(r),Si=null}else return t=hr(n),t!==null&&mo(t),e.blockedOn=n,!1;t.shift()}return!0}function au(e,t,n){Mr(e)&&n.delete(t)}function Tf(){Ei=!1,mt!==null&&Mr(mt)&&(mt=null),ht!==null&&Mr(ht)&&(ht=null),vt!==null&&Mr(vt)&&(vt=null),qn.forEach(au),bn.forEach(au)}function Ln(e,t){e.blockedOn===t&&(e.blockedOn=null,Ei||(Ei=!0,_e.unstable_scheduleCallback(_e.unstable_NormalPriority,Tf)))}function er(e){function t(l){return Ln(l,e)}if(0<Er.length){Ln(Er[0],e);for(var n=1;n<Er.length;n++){var r=Er[n];r.blockedOn===e&&(r.blockedOn=null)}}for(mt!==null&&Ln(mt,e),ht!==null&&Ln(ht,e),vt!==null&&Ln(vt,e),qn.forEach(t),bn.forEach(t),n=0;n<ct.length;n++)r=ct[n],r.blockedOn===e&&(r.blockedOn=null);for(;0<ct.length&&(n=ct[0],n.blockedOn===null);)Xs(n),n.blockedOn===null&&ct.shift()}var dn=it.ReactCurrentBatchConfig,qr=!0;function zf(e,t,n,r){var l=A,i=dn.transition;dn.transition=null;try{A=1,ho(e,t,n,r)}finally{A=l,dn.transition=i}}function Lf(e,t,n,r){var l=A,i=dn.transition;dn.transition=null;try{A=4,ho(e,t,n,r)}finally{A=l,dn.transition=i}}function ho(e,t,n,r){if(qr){var l=Ni(e,t,n,r);if(l===null)Gl(e,t,r,br,n),su(e,r);else if(Pf(l,e,t,n,r))r.stopPropagation();else if(su(e,r),t&4&&-1<jf.indexOf(e)){for(;l!==null;){var i=hr(l);if(i!==null&&Ks(i),i=Ni(e,t,n,r),i===null&&Gl(e,t,r,br,n),i===l)break;l=i}l!==null&&r.stopPropagation()}else Gl(e,t,r,null,n)}}var br=null;function Ni(e,t,n,r){if(br=null,e=co(r),e=It(e),e!==null)if(t=Wt(e),t===null)e=null;else if(n=t.tag,n===13){if(e=As(t),e!==null)return e;e=null}else if(n===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;e=null}else t!==e&&(e=null);return br=e,null}function Zs(e){switch(e){case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 1;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"toggle":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 4;case"message":switch(gf()){case fo:return 1;case Hs:return 4;case Xr:case Sf:return 16;case Vs:return 536870912;default:return 16}default:return 16}}var dt=null,vo=null,Ar=null;function qs(){if(Ar)return Ar;var e,t=vo,n=t.length,r,l="value"in dt?dt.value:dt.textContent,i=l.length;for(e=0;e<n&&t[e]===l[e];e++);var o=n-e;for(r=1;r<=o&&t[n-r]===l[i-r];r++);return Ar=l.slice(e,1<r?1-r:void 0)}function Ur(e){var t=e.keyCode;return"charCode"in e?(e=e.charCode,e===0&&t===13&&(e=13)):e=t,e===10&&(e=13),32<=e||e===13?e:0}function Nr(){return!0}function cu(){return!1}function Pe(e){function t(n,r,l,i,o){this._reactName=n,this._targetInst=l,this.type=r,this.nativeEvent=i,this.target=o,this.currentTarget=null;for(var u in e)e.hasOwnProperty(u)&&(n=e[u],this[u]=n?n(i):i[u]);return this.isDefaultPrevented=(i.defaultPrevented!=null?i.defaultPrevented:i.returnValue===!1)?Nr:cu,this.isPropagationStopped=cu,this}return J(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var n=this.nativeEvent;n&&(n.preventDefault?n.preventDefault():typeof n.returnValue!="unknown"&&(n.returnValue=!1),this.isDefaultPrevented=Nr)},stopPropagation:function(){var n=this.nativeEvent;n&&(n.stopPropagation?n.stopPropagation():typeof n.cancelBubble!="unknown"&&(n.cancelBubble=!0),this.isPropagationStopped=Nr)},persist:function(){},isPersistent:Nr}),t}var Cn={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(e){return e.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},yo=Pe(Cn),mr=J({},Cn,{view:0,detail:0}),Of=Pe(mr),Ul,$l,On,Sl=J({},mr,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:go,button:0,buttons:0,relatedTarget:function(e){return e.relatedTarget===void 0?e.fromElement===e.srcElement?e.toElement:e.fromElement:e.relatedTarget},movementX:function(e){return"movementX"in e?e.movementX:(e!==On&&(On&&e.type==="mousemove"?(Ul=e.screenX-On.screenX,$l=e.screenY-On.screenY):$l=Ul=0,On=e),Ul)},movementY:function(e){return"movementY"in e?e.movementY:$l}}),fu=Pe(Sl),If=J({},Sl,{dataTransfer:0}),Rf=Pe(If),Df=J({},mr,{relatedTarget:0}),Bl=Pe(Df),Ff=J({},Cn,{animationName:0,elapsedTime:0,pseudoElement:0}),Mf=Pe(Ff),Af=J({},Cn,{clipboardData:function(e){return"clipboardData"in e?e.clipboardData:window.clipboardData}}),Uf=Pe(Af),$f=J({},Cn,{data:0}),du=Pe($f),Bf={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},Hf={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},Vf={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function Wf(e){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(e):(e=Vf[e])?!!t[e]:!1}function go(){return Wf}var Qf=J({},mr,{key:function(e){if(e.key){var t=Bf[e.key]||e.key;if(t!=="Unidentified")return t}return e.type==="keypress"?(e=Ur(e),e===13?"Enter":String.fromCharCode(e)):e.type==="keydown"||e.type==="keyup"?Hf[e.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:go,charCode:function(e){return e.type==="keypress"?Ur(e):0},keyCode:function(e){return e.type==="keydown"||e.type==="keyup"?e.keyCode:0},which:function(e){return e.type==="keypress"?Ur(e):e.type==="keydown"||e.type==="keyup"?e.keyCode:0}}),Kf=Pe(Qf),Yf=J({},Sl,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0}),pu=Pe(Yf),Gf=J({},mr,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:go}),Jf=Pe(Gf),Xf=J({},Cn,{propertyName:0,elapsedTime:0,pseudoElement:0}),Zf=Pe(Xf),qf=J({},Sl,{deltaX:function(e){return"deltaX"in e?e.deltaX:"wheelDeltaX"in e?-e.wheelDeltaX:0},deltaY:function(e){return"deltaY"in e?e.deltaY:"wheelDeltaY"in e?-e.wheelDeltaY:"wheelDelta"in e?-e.wheelDelta:0},deltaZ:0,deltaMode:0}),bf=Pe(qf),ed=[9,13,27,32],So=tt&&"CompositionEvent"in window,Hn=null;tt&&"documentMode"in document&&(Hn=document.documentMode);var td=tt&&"TextEvent"in window&&!Hn,bs=tt&&(!So||Hn&&8<Hn&&11>=Hn),mu=" ",hu=!1;function ea(e,t){switch(e){case"keyup":return ed.indexOf(t.keyCode)!==-1;case"keydown":return t.keyCode!==229;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function ta(e){return e=e.detail,typeof e=="object"&&"data"in e?e.data:null}var qt=!1;function nd(e,t){switch(e){case"compositionend":return ta(t);case"keypress":return t.which!==32?null:(hu=!0,mu);case"textInput":return e=t.data,e===mu&&hu?null:e;default:return null}}function rd(e,t){if(qt)return e==="compositionend"||!So&&ea(e,t)?(e=qs(),Ar=vo=dt=null,qt=!1,e):null;switch(e){case"paste":return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return bs&&t.locale!=="ko"?null:t.data;default:return null}}var ld={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function vu(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t==="input"?!!ld[e.type]:t==="textarea"}function na(e,t,n,r){Is(r),t=el(t,"onChange"),0<t.length&&(n=new yo("onChange","change",null,n,r),e.push({event:n,listeners:t}))}var Vn=null,tr=null;function id(e){pa(e,0)}function wl(e){var t=tn(e);if(_s(t))return e}function od(e,t){if(e==="change")return t}var ra=!1;if(tt){var Hl;if(tt){var Vl="oninput"in document;if(!Vl){var yu=document.createElement("div");yu.setAttribute("oninput","return;"),Vl=typeof yu.oninput=="function"}Hl=Vl}else Hl=!1;ra=Hl&&(!document.documentMode||9<document.documentMode)}function gu(){Vn&&(Vn.detachEvent("onpropertychange",la),tr=Vn=null)}function la(e){if(e.propertyName==="value"&&wl(tr)){var t=[];na(t,tr,e,co(e)),Ms(id,t)}}function ud(e,t,n){e==="focusin"?(gu(),Vn=t,tr=n,Vn.attachEvent("onpropertychange",la)):e==="focusout"&&gu()}function sd(e){if(e==="selectionchange"||e==="keyup"||e==="keydown")return wl(tr)}function ad(e,t){if(e==="click")return wl(t)}function cd(e,t){if(e==="input"||e==="change")return wl(t)}function fd(e,t){return e===t&&(e!==0||1/e===1/t)||e!==e&&t!==t}var Be=typeof Object.is=="function"?Object.is:fd;function nr(e,t){if(Be(e,t))return!0;if(typeof e!="object"||e===null||typeof t!="object"||t===null)return!1;var n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(r=0;r<n.length;r++){var l=n[r];if(!ui.call(t,l)||!Be(e[l],t[l]))return!1}return!0}function Su(e){for(;e&&e.firstChild;)e=e.firstChild;return e}function wu(e,t){var n=Su(e);e=0;for(var r;n;){if(n.nodeType===3){if(r=e+n.textContent.length,e<=t&&r>=t)return{node:n,offset:t-e};e=r}e:{for(;n;){if(n.nextSibling){n=n.nextSibling;break e}n=n.parentNode}n=void 0}n=Su(n)}}function ia(e,t){return e&&t?e===t?!0:e&&e.nodeType===3?!1:t&&t.nodeType===3?ia(e,t.parentNode):"contains"in e?e.contains(t):e.compareDocumentPosition?!!(e.compareDocumentPosition(t)&16):!1:!1}function oa(){for(var e=window,t=Yr();t instanceof e.HTMLIFrameElement;){try{var n=typeof t.contentWindow.location.href=="string"}catch{n=!1}if(n)e=t.contentWindow;else break;t=Yr(e.document)}return t}function wo(e){var t=e&&e.nodeName&&e.nodeName.toLowerCase();return t&&(t==="input"&&(e.type==="text"||e.type==="search"||e.type==="tel"||e.type==="url"||e.type==="password")||t==="textarea"||e.contentEditable==="true")}function dd(e){var t=oa(),n=e.focusedElem,r=e.selectionRange;if(t!==n&&n&&n.ownerDocument&&ia(n.ownerDocument.documentElement,n)){if(r!==null&&wo(n)){if(t=r.start,e=r.end,e===void 0&&(e=t),"selectionStart"in n)n.selectionStart=t,n.selectionEnd=Math.min(e,n.value.length);else if(e=(t=n.ownerDocument||document)&&t.defaultView||window,e.getSelection){e=e.getSelection();var l=n.textContent.length,i=Math.min(r.start,l);r=r.end===void 0?i:Math.min(r.end,l),!e.extend&&i>r&&(l=r,r=i,i=l),l=wu(n,i);var o=wu(n,r);l&&o&&(e.rangeCount!==1||e.anchorNode!==l.node||e.anchorOffset!==l.offset||e.focusNode!==o.node||e.focusOffset!==o.offset)&&(t=t.createRange(),t.setStart(l.node,l.offset),e.removeAllRanges(),i>r?(e.addRange(t),e.extend(o.node,o.offset)):(t.setEnd(o.node,o.offset),e.addRange(t)))}}for(t=[],e=n;e=e.parentNode;)e.nodeType===1&&t.push({element:e,left:e.scrollLeft,top:e.scrollTop});for(typeof n.focus=="function"&&n.focus(),n=0;n<t.length;n++)e=t[n],e.element.scrollLeft=e.left,e.element.scrollTop=e.top}}var pd=tt&&"documentMode"in document&&11>=document.documentMode,bt=null,_i=null,Wn=null,ji=!1;function xu(e,t,n){var r=n.window===n?n.document:n.nodeType===9?n:n.ownerDocument;ji||bt==null||bt!==Yr(r)||(r=bt,"selectionStart"in r&&wo(r)?r={start:r.selectionStart,end:r.selectionEnd}:(r=(r.ownerDocument&&r.ownerDocument.defaultView||window).getSelection(),r={anchorNode:r.anchorNode,anchorOffset:r.anchorOffset,focusNode:r.focusNode,focusOffset:r.focusOffset}),Wn&&nr(Wn,r)||(Wn=r,r=el(_i,"onSelect"),0<r.length&&(t=new yo("onSelect","select",null,t,n),e.push({event:t,listeners:r}),t.target=bt)))}function _r(e,t){var n={};return n[e.toLowerCase()]=t.toLowerCase(),n["Webkit"+e]="webkit"+t,n["Moz"+e]="moz"+t,n}var en={animationend:_r("Animation","AnimationEnd"),animationiteration:_r("Animation","AnimationIteration"),animationstart:_r("Animation","AnimationStart"),transitionend:_r("Transition","TransitionEnd")},Wl={},ua={};tt&&(ua=document.createElement("div").style,"AnimationEvent"in window||(delete en.animationend.animation,delete en.animationiteration.animation,delete en.animationstart.animation),"TransitionEvent"in window||delete en.transitionend.transition);function xl(e){if(Wl[e])return Wl[e];if(!en[e])return e;var t=en[e],n;for(n in t)if(t.hasOwnProperty(n)&&n in ua)return Wl[e]=t[n];return e}var sa=xl("animationend"),aa=xl("animationiteration"),ca=xl("animationstart"),fa=xl("transitionend"),da=new Map,ku="abort auxClick cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");function Et(e,t){da.set(e,t),Vt(t,[e])}for(var Ql=0;Ql<ku.length;Ql++){var Kl=ku[Ql],md=Kl.toLowerCase(),hd=Kl[0].toUpperCase()+Kl.slice(1);Et(md,"on"+hd)}Et(sa,"onAnimationEnd");Et(aa,"onAnimationIteration");Et(ca,"onAnimationStart");Et("dblclick","onDoubleClick");Et("focusin","onFocus");Et("focusout","onBlur");Et(fa,"onTransitionEnd");hn("onMouseEnter",["mouseout","mouseover"]);hn("onMouseLeave",["mouseout","mouseover"]);hn("onPointerEnter",["pointerout","pointerover"]);hn("onPointerLeave",["pointerout","pointerover"]);Vt("onChange","change click focusin focusout input keydown keyup selectionchange".split(" "));Vt("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" "));Vt("onBeforeInput",["compositionend","keypress","textInput","paste"]);Vt("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" "));Vt("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" "));Vt("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var Un="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),vd=new Set("cancel close invalid load scroll toggle".split(" ").concat(Un));function Cu(e,t,n){var r=e.type||"unknown-event";e.currentTarget=n,mf(r,t,void 0,e),e.currentTarget=null}function pa(e,t){t=(t&4)!==0;for(var n=0;n<e.length;n++){var r=e[n],l=r.event;r=r.listeners;e:{var i=void 0;if(t)for(var o=r.length-1;0<=o;o--){var u=r[o],s=u.instance,d=u.currentTarget;if(u=u.listener,s!==i&&l.isPropagationStopped())break e;Cu(l,u,d),i=s}else for(o=0;o<r.length;o++){if(u=r[o],s=u.instance,d=u.currentTarget,u=u.listener,s!==i&&l.isPropagationStopped())break e;Cu(l,u,d),i=s}}}if(Jr)throw e=ki,Jr=!1,ki=null,e}function H(e,t){var n=t[Oi];n===void 0&&(n=t[Oi]=new Set);var r=e+"__bubble";n.has(r)||(ma(t,e,2,!1),n.add(r))}function Yl(e,t,n){var r=0;t&&(r|=4),ma(n,e,r,t)}var jr="_reactListening"+Math.random().toString(36).slice(2);function rr(e){if(!e[jr]){e[jr]=!0,xs.forEach(function(n){n!=="selectionchange"&&(vd.has(n)||Yl(n,!1,e),Yl(n,!0,e))});var t=e.nodeType===9?e:e.ownerDocument;t===null||t[jr]||(t[jr]=!0,Yl("selectionchange",!1,t))}}function ma(e,t,n,r){switch(Zs(t)){case 1:var l=zf;break;case 4:l=Lf;break;default:l=ho}n=l.bind(null,t,n,e),l=void 0,!xi||t!=="touchstart"&&t!=="touchmove"&&t!=="wheel"||(l=!0),r?l!==void 0?e.addEventListener(t,n,{capture:!0,passive:l}):e.addEventListener(t,n,!0):l!==void 0?e.addEventListener(t,n,{passive:l}):e.addEventListener(t,n,!1)}function Gl(e,t,n,r,l){var i=r;if(!(t&1)&&!(t&2)&&r!==null)e:for(;;){if(r===null)return;var o=r.tag;if(o===3||o===4){var u=r.stateNode.containerInfo;if(u===l||u.nodeType===8&&u.parentNode===l)break;if(o===4)for(o=r.return;o!==null;){var s=o.tag;if((s===3||s===4)&&(s=o.stateNode.containerInfo,s===l||s.nodeType===8&&s.parentNode===l))return;o=o.return}for(;u!==null;){if(o=It(u),o===null)return;if(s=o.tag,s===5||s===6){r=i=o;continue e}u=u.parentNode}}r=r.return}Ms(function(){var d=i,y=co(n),h=[];e:{var m=da.get(e);if(m!==void 0){var w=yo,k=e;switch(e){case"keypress":if(Ur(n)===0)break e;case"keydown":case"keyup":w=Kf;break;case"focusin":k="focus",w=Bl;break;case"focusout":k="blur",w=Bl;break;case"beforeblur":case"afterblur":w=Bl;break;case"click":if(n.button===2)break e;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":w=fu;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":w=Rf;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":w=Jf;break;case sa:case aa:case ca:w=Mf;break;case fa:w=Zf;break;case"scroll":w=Of;break;case"wheel":w=bf;break;case"copy":case"cut":case"paste":w=Uf;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":w=pu}var v=(t&4)!==0,D=!v&&e==="scroll",a=v?m!==null?m+"Capture":null:m;v=[];for(var c=d,p;c!==null;){p=c;var g=p.stateNode;if(p.tag===5&&g!==null&&(p=g,a!==null&&(g=Zn(c,a),g!=null&&v.push(lr(c,g,p)))),D)break;c=c.return}0<v.length&&(m=new w(m,k,null,n,y),h.push({event:m,listeners:v}))}}if(!(t&7)){e:{if(m=e==="mouseover"||e==="pointerover",w=e==="mouseout"||e==="pointerout",m&&n!==Si&&(k=n.relatedTarget||n.fromElement)&&(It(k)||k[nt]))break e;if((w||m)&&(m=y.window===y?y:(m=y.ownerDocument)?m.defaultView||m.parentWindow:window,w?(k=n.relatedTarget||n.toElement,w=d,k=k?It(k):null,k!==null&&(D=Wt(k),k!==D||k.tag!==5&&k.tag!==6)&&(k=null)):(w=null,k=d),w!==k)){if(v=fu,g="onMouseLeave",a="onMouseEnter",c="mouse",(e==="pointerout"||e==="pointerover")&&(v=pu,g="onPointerLeave",a="onPointerEnter",c="pointer"),D=w==null?m:tn(w),p=k==null?m:tn(k),m=new v(g,c+"leave",w,n,y),m.target=D,m.relatedTarget=p,g=null,It(y)===d&&(v=new v(a,c+"enter",k,n,y),v.target=p,v.relatedTarget=D,g=v),D=g,w&&k)t:{for(v=w,a=k,c=0,p=v;p;p=Jt(p))c++;for(p=0,g=a;g;g=Jt(g))p++;for(;0<c-p;)v=Jt(v),c--;for(;0<p-c;)a=Jt(a),p--;for(;c--;){if(v===a||a!==null&&v===a.alternate)break t;v=Jt(v),a=Jt(a)}v=null}else v=null;w!==null&&Eu(h,m,w,v,!1),k!==null&&D!==null&&Eu(h,D,k,v,!0)}}e:{if(m=d?tn(d):window,w=m.nodeName&&m.nodeName.toLowerCase(),w==="select"||w==="input"&&m.type==="file")var E=od;else if(vu(m))if(ra)E=cd;else{E=sd;var j=ud}else(w=m.nodeName)&&w.toLowerCase()==="input"&&(m.type==="checkbox"||m.type==="radio")&&(E=ad);if(E&&(E=E(e,d))){na(h,E,n,y);break e}j&&j(e,m,d),e==="focusout"&&(j=m._wrapperState)&&j.controlled&&m.type==="number"&&mi(m,"number",m.value)}switch(j=d?tn(d):window,e){case"focusin":(vu(j)||j.contentEditable==="true")&&(bt=j,_i=d,Wn=null);break;case"focusout":Wn=_i=bt=null;break;case"mousedown":ji=!0;break;case"contextmenu":case"mouseup":case"dragend":ji=!1,xu(h,n,y);break;case"selectionchange":if(pd)break;case"keydown":case"keyup":xu(h,n,y)}var _;if(So)e:{switch(e){case"compositionstart":var T="onCompositionStart";break e;case"compositionend":T="onCompositionEnd";break e;case"compositionupdate":T="onCompositionUpdate";break e}T=void 0}else qt?ea(e,n)&&(T="onCompositionEnd"):e==="keydown"&&n.keyCode===229&&(T="onCompositionStart");T&&(bs&&n.locale!=="ko"&&(qt||T!=="onCompositionStart"?T==="onCompositionEnd"&&qt&&(_=qs()):(dt=y,vo="value"in dt?dt.value:dt.textContent,qt=!0)),j=el(d,T),0<j.length&&(T=new du(T,e,null,n,y),h.push({event:T,listeners:j}),_?T.data=_:(_=ta(n),_!==null&&(T.data=_)))),(_=td?nd(e,n):rd(e,n))&&(d=el(d,"onBeforeInput"),0<d.length&&(y=new du("onBeforeInput","beforeinput",null,n,y),h.push({event:y,listeners:d}),y.data=_))}pa(h,t)})}function lr(e,t,n){return{instance:e,listener:t,currentTarget:n}}function el(e,t){for(var n=t+"Capture",r=[];e!==null;){var l=e,i=l.stateNode;l.tag===5&&i!==null&&(l=i,i=Zn(e,n),i!=null&&r.unshift(lr(e,i,l)),i=Zn(e,t),i!=null&&r.push(lr(e,i,l))),e=e.return}return r}function Jt(e){if(e===null)return null;do e=e.return;while(e&&e.tag!==5);return e||null}function Eu(e,t,n,r,l){for(var i=t._reactName,o=[];n!==null&&n!==r;){var u=n,s=u.alternate,d=u.stateNode;if(s!==null&&s===r)break;u.tag===5&&d!==null&&(u=d,l?(s=Zn(n,i),s!=null&&o.unshift(lr(n,s,u))):l||(s=Zn(n,i),s!=null&&o.push(lr(n,s,u)))),n=n.return}o.length!==0&&e.push({event:t,listeners:o})}var yd=/\r\n?/g,gd=/\u0000|\uFFFD/g;function Nu(e){return(typeof e=="string"?e:""+e).replace(yd,`
-`).replace(gd,"")}function Pr(e,t,n){if(t=Nu(t),Nu(e)!==t&&n)throw Error(S(425))}function tl(){}var Pi=null,Ti=null;function zi(e,t){return e==="textarea"||e==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var Li=typeof setTimeout=="function"?setTimeout:void 0,Sd=typeof clearTimeout=="function"?clearTimeout:void 0,_u=typeof Promise=="function"?Promise:void 0,wd=typeof queueMicrotask=="function"?queueMicrotask:typeof _u<"u"?function(e){return _u.resolve(null).then(e).catch(xd)}:Li;function xd(e){setTimeout(function(){throw e})}function Jl(e,t){var n=t,r=0;do{var l=n.nextSibling;if(e.removeChild(n),l&&l.nodeType===8)if(n=l.data,n==="/$"){if(r===0){e.removeChild(l),er(t);return}r--}else n!=="$"&&n!=="$?"&&n!=="$!"||r++;n=l}while(n);er(t)}function yt(e){for(;e!=null;e=e.nextSibling){var t=e.nodeType;if(t===1||t===3)break;if(t===8){if(t=e.data,t==="$"||t==="$!"||t==="$?")break;if(t==="/$")return null}}return e}function ju(e){e=e.previousSibling;for(var t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="$"||n==="$!"||n==="$?"){if(t===0)return e;t--}else n==="/$"&&t++}e=e.previousSibling}return null}var En=Math.random().toString(36).slice(2),Qe="__reactFiber$"+En,ir="__reactProps$"+En,nt="__reactContainer$"+En,Oi="__reactEvents$"+En,kd="__reactListeners$"+En,Cd="__reactHandles$"+En;function It(e){var t=e[Qe];if(t)return t;for(var n=e.parentNode;n;){if(t=n[nt]||n[Qe]){if(n=t.alternate,t.child!==null||n!==null&&n.child!==null)for(e=ju(e);e!==null;){if(n=e[Qe])return n;e=ju(e)}return t}e=n,n=e.parentNode}return null}function hr(e){return e=e[Qe]||e[nt],!e||e.tag!==5&&e.tag!==6&&e.tag!==13&&e.tag!==3?null:e}function tn(e){if(e.tag===5||e.tag===6)return e.stateNode;throw Error(S(33))}function kl(e){return e[ir]||null}var Ii=[],nn=-1;function Nt(e){return{current:e}}function V(e){0>nn||(e.current=Ii[nn],Ii[nn]=null,nn--)}function $(e,t){nn++,Ii[nn]=e.current,e.current=t}var Ct={},de=Nt(Ct),Se=Nt(!1),At=Ct;function vn(e,t){var n=e.type.contextTypes;if(!n)return Ct;var r=e.stateNode;if(r&&r.__reactInternalMemoizedUnmaskedChildContext===t)return r.__reactInternalMemoizedMaskedChildContext;var l={},i;for(i in n)l[i]=t[i];return r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=t,e.__reactInternalMemoizedMaskedChildContext=l),l}function we(e){return e=e.childContextTypes,e!=null}function nl(){V(Se),V(de)}function Pu(e,t,n){if(de.current!==Ct)throw Error(S(168));$(de,t),$(Se,n)}function ha(e,t,n){var r=e.stateNode;if(t=t.childContextTypes,typeof r.getChildContext!="function")return n;r=r.getChildContext();for(var l in r)if(!(l in t))throw Error(S(108,uf(e)||"Unknown",l));return J({},n,r)}function rl(e){return e=(e=e.stateNode)&&e.__reactInternalMemoizedMergedChildContext||Ct,At=de.current,$(de,e),$(Se,Se.current),!0}function Tu(e,t,n){var r=e.stateNode;if(!r)throw Error(S(169));n?(e=ha(e,t,At),r.__reactInternalMemoizedMergedChildContext=e,V(Se),V(de),$(de,e)):V(Se),$(Se,n)}var Ze=null,Cl=!1,Xl=!1;function va(e){Ze===null?Ze=[e]:Ze.push(e)}function Ed(e){Cl=!0,va(e)}function _t(){if(!Xl&&Ze!==null){Xl=!0;var e=0,t=A;try{var n=Ze;for(A=1;e<n.length;e++){var r=n[e];do r=r(!0);while(r!==null)}Ze=null,Cl=!1}catch(l){throw Ze!==null&&(Ze=Ze.slice(e+1)),Bs(fo,_t),l}finally{A=t,Xl=!1}}return null}var rn=[],ln=0,ll=null,il=0,Te=[],ze=0,Ut=null,qe=1,be="";function Lt(e,t){rn[ln++]=il,rn[ln++]=ll,ll=e,il=t}function ya(e,t,n){Te[ze++]=qe,Te[ze++]=be,Te[ze++]=Ut,Ut=e;var r=qe;e=be;var l=32-Ue(r)-1;r&=~(1<<l),n+=1;var i=32-Ue(t)+l;if(30<i){var o=l-l%5;i=(r&(1<<o)-1).toString(32),r>>=o,l-=o,qe=1<<32-Ue(t)+l|n<<l|r,be=i+e}else qe=1<<i|n<<l|r,be=e}function xo(e){e.return!==null&&(Lt(e,1),ya(e,1,0))}function ko(e){for(;e===ll;)ll=rn[--ln],rn[ln]=null,il=rn[--ln],rn[ln]=null;for(;e===Ut;)Ut=Te[--ze],Te[ze]=null,be=Te[--ze],Te[ze]=null,qe=Te[--ze],Te[ze]=null}var Ne=null,Ee=null,Q=!1,Ae=null;function ga(e,t){var n=Le(5,null,null,0);n.elementType="DELETED",n.stateNode=t,n.return=e,t=e.deletions,t===null?(e.deletions=[n],e.flags|=16):t.push(n)}function zu(e,t){switch(e.tag){case 5:var n=e.type;return t=t.nodeType!==1||n.toLowerCase()!==t.nodeName.toLowerCase()?null:t,t!==null?(e.stateNode=t,Ne=e,Ee=yt(t.firstChild),!0):!1;case 6:return t=e.pendingProps===""||t.nodeType!==3?null:t,t!==null?(e.stateNode=t,Ne=e,Ee=null,!0):!1;case 13:return t=t.nodeType!==8?null:t,t!==null?(n=Ut!==null?{id:qe,overflow:be}:null,e.memoizedState={dehydrated:t,treeContext:n,retryLane:1073741824},n=Le(18,null,null,0),n.stateNode=t,n.return=e,e.child=n,Ne=e,Ee=null,!0):!1;default:return!1}}function Ri(e){return(e.mode&1)!==0&&(e.flags&128)===0}function Di(e){if(Q){var t=Ee;if(t){var n=t;if(!zu(e,t)){if(Ri(e))throw Error(S(418));t=yt(n.nextSibling);var r=Ne;t&&zu(e,t)?ga(r,n):(e.flags=e.flags&-4097|2,Q=!1,Ne=e)}}else{if(Ri(e))throw Error(S(418));e.flags=e.flags&-4097|2,Q=!1,Ne=e}}}function Lu(e){for(e=e.return;e!==null&&e.tag!==5&&e.tag!==3&&e.tag!==13;)e=e.return;Ne=e}function Tr(e){if(e!==Ne)return!1;if(!Q)return Lu(e),Q=!0,!1;var t;if((t=e.tag!==3)&&!(t=e.tag!==5)&&(t=e.type,t=t!=="head"&&t!=="body"&&!zi(e.type,e.memoizedProps)),t&&(t=Ee)){if(Ri(e))throw Sa(),Error(S(418));for(;t;)ga(e,t),t=yt(t.nextSibling)}if(Lu(e),e.tag===13){if(e=e.memoizedState,e=e!==null?e.dehydrated:null,!e)throw Error(S(317));e:{for(e=e.nextSibling,t=0;e;){if(e.nodeType===8){var n=e.data;if(n==="/$"){if(t===0){Ee=yt(e.nextSibling);break e}t--}else n!=="$"&&n!=="$!"&&n!=="$?"||t++}e=e.nextSibling}Ee=null}}else Ee=Ne?yt(e.stateNode.nextSibling):null;return!0}function Sa(){for(var e=Ee;e;)e=yt(e.nextSibling)}function yn(){Ee=Ne=null,Q=!1}function Co(e){Ae===null?Ae=[e]:Ae.push(e)}var Nd=it.ReactCurrentBatchConfig;function In(e,t,n){if(e=n.ref,e!==null&&typeof e!="function"&&typeof e!="object"){if(n._owner){if(n=n._owner,n){if(n.tag!==1)throw Error(S(309));var r=n.stateNode}if(!r)throw Error(S(147,e));var l=r,i=""+e;return t!==null&&t.ref!==null&&typeof t.ref=="function"&&t.ref._stringRef===i?t.ref:(t=function(o){var u=l.refs;o===null?delete u[i]:u[i]=o},t._stringRef=i,t)}if(typeof e!="string")throw Error(S(284));if(!n._owner)throw Error(S(290,e))}return e}function zr(e,t){throw e=Object.prototype.toString.call(t),Error(S(31,e==="[object Object]"?"object with keys {"+Object.keys(t).join(", ")+"}":e))}function Ou(e){var t=e._init;return t(e._payload)}function wa(e){function t(a,c){if(e){var p=a.deletions;p===null?(a.deletions=[c],a.flags|=16):p.push(c)}}function n(a,c){if(!e)return null;for(;c!==null;)t(a,c),c=c.sibling;return null}function r(a,c){for(a=new Map;c!==null;)c.key!==null?a.set(c.key,c):a.set(c.index,c),c=c.sibling;return a}function l(a,c){return a=xt(a,c),a.index=0,a.sibling=null,a}function i(a,c,p){return a.index=p,e?(p=a.alternate,p!==null?(p=p.index,p<c?(a.flags|=2,c):p):(a.flags|=2,c)):(a.flags|=1048576,c)}function o(a){return e&&a.alternate===null&&(a.flags|=2),a}function u(a,c,p,g){return c===null||c.tag!==6?(c=ri(p,a.mode,g),c.return=a,c):(c=l(c,p),c.return=a,c)}function s(a,c,p,g){var E=p.type;return E===Zt?y(a,c,p.props.children,g,p.key):c!==null&&(c.elementType===E||typeof E=="object"&&E!==null&&E.$$typeof===st&&Ou(E)===c.type)?(g=l(c,p.props),g.ref=In(a,c,p),g.return=a,g):(g=Kr(p.type,p.key,p.props,null,a.mode,g),g.ref=In(a,c,p),g.return=a,g)}function d(a,c,p,g){return c===null||c.tag!==4||c.stateNode.containerInfo!==p.containerInfo||c.stateNode.implementation!==p.implementation?(c=li(p,a.mode,g),c.return=a,c):(c=l(c,p.children||[]),c.return=a,c)}function y(a,c,p,g,E){return c===null||c.tag!==7?(c=Mt(p,a.mode,g,E),c.return=a,c):(c=l(c,p),c.return=a,c)}function h(a,c,p){if(typeof c=="string"&&c!==""||typeof c=="number")return c=ri(""+c,a.mode,p),c.return=a,c;if(typeof c=="object"&&c!==null){switch(c.$$typeof){case Sr:return p=Kr(c.type,c.key,c.props,null,a.mode,p),p.ref=In(a,null,c),p.return=a,p;case Xt:return c=li(c,a.mode,p),c.return=a,c;case st:var g=c._init;return h(a,g(c._payload),p)}if(Mn(c)||Pn(c))return c=Mt(c,a.mode,p,null),c.return=a,c;zr(a,c)}return null}function m(a,c,p,g){var E=c!==null?c.key:null;if(typeof p=="string"&&p!==""||typeof p=="number")return E!==null?null:u(a,c,""+p,g);if(typeof p=="object"&&p!==null){switch(p.$$typeof){case Sr:return p.key===E?s(a,c,p,g):null;case Xt:return p.key===E?d(a,c,p,g):null;case st:return E=p._init,m(a,c,E(p._payload),g)}if(Mn(p)||Pn(p))return E!==null?null:y(a,c,p,g,null);zr(a,p)}return null}function w(a,c,p,g,E){if(typeof g=="string"&&g!==""||typeof g=="number")return a=a.get(p)||null,u(c,a,""+g,E);if(typeof g=="object"&&g!==null){switch(g.$$typeof){case Sr:return a=a.get(g.key===null?p:g.key)||null,s(c,a,g,E);case Xt:return a=a.get(g.key===null?p:g.key)||null,d(c,a,g,E);case st:var j=g._init;return w(a,c,p,j(g._payload),E)}if(Mn(g)||Pn(g))return a=a.get(p)||null,y(c,a,g,E,null);zr(c,g)}return null}function k(a,c,p,g){for(var E=null,j=null,_=c,T=c=0,K=null;_!==null&&T<p.length;T++){_.index>T?(K=_,_=null):K=_.sibling;var I=m(a,_,p[T],g);if(I===null){_===null&&(_=K);break}e&&_&&I.alternate===null&&t(a,_),c=i(I,c,T),j===null?E=I:j.sibling=I,j=I,_=K}if(T===p.length)return n(a,_),Q&&Lt(a,T),E;if(_===null){for(;T<p.length;T++)_=h(a,p[T],g),_!==null&&(c=i(_,c,T),j===null?E=_:j.sibling=_,j=_);return Q&&Lt(a,T),E}for(_=r(a,_);T<p.length;T++)K=w(_,a,T,p[T],g),K!==null&&(e&&K.alternate!==null&&_.delete(K.key===null?T:K.key),c=i(K,c,T),j===null?E=K:j.sibling=K,j=K);return e&&_.forEach(function(ke){return t(a,ke)}),Q&&Lt(a,T),E}function v(a,c,p,g){var E=Pn(p);if(typeof E!="function")throw Error(S(150));if(p=E.call(p),p==null)throw Error(S(151));for(var j=E=null,_=c,T=c=0,K=null,I=p.next();_!==null&&!I.done;T++,I=p.next()){_.index>T?(K=_,_=null):K=_.sibling;var ke=m(a,_,I.value,g);if(ke===null){_===null&&(_=K);break}e&&_&&ke.alternate===null&&t(a,_),c=i(ke,c,T),j===null?E=ke:j.sibling=ke,j=ke,_=K}if(I.done)return n(a,_),Q&&Lt(a,T),E;if(_===null){for(;!I.done;T++,I=p.next())I=h(a,I.value,g),I!==null&&(c=i(I,c,T),j===null?E=I:j.sibling=I,j=I);return Q&&Lt(a,T),E}for(_=r(a,_);!I.done;T++,I=p.next())I=w(_,a,T,I.value,g),I!==null&&(e&&I.alternate!==null&&_.delete(I.key===null?T:I.key),c=i(I,c,T),j===null?E=I:j.sibling=I,j=I);return e&&_.forEach(function(He){return t(a,He)}),Q&&Lt(a,T),E}function D(a,c,p,g){if(typeof p=="object"&&p!==null&&p.type===Zt&&p.key===null&&(p=p.props.children),typeof p=="object"&&p!==null){switch(p.$$typeof){case Sr:e:{for(var E=p.key,j=c;j!==null;){if(j.key===E){if(E=p.type,E===Zt){if(j.tag===7){n(a,j.sibling),c=l(j,p.props.children),c.return=a,a=c;break e}}else if(j.elementType===E||typeof E=="object"&&E!==null&&E.$$typeof===st&&Ou(E)===j.type){n(a,j.sibling),c=l(j,p.props),c.ref=In(a,j,p),c.return=a,a=c;break e}n(a,j);break}else t(a,j);j=j.sibling}p.type===Zt?(c=Mt(p.props.children,a.mode,g,p.key),c.return=a,a=c):(g=Kr(p.type,p.key,p.props,null,a.mode,g),g.ref=In(a,c,p),g.return=a,a=g)}return o(a);case Xt:e:{for(j=p.key;c!==null;){if(c.key===j)if(c.tag===4&&c.stateNode.containerInfo===p.containerInfo&&c.stateNode.implementation===p.implementation){n(a,c.sibling),c=l(c,p.children||[]),c.return=a,a=c;break e}else{n(a,c);break}else t(a,c);c=c.sibling}c=li(p,a.mode,g),c.return=a,a=c}return o(a);case st:return j=p._init,D(a,c,j(p._payload),g)}if(Mn(p))return k(a,c,p,g);if(Pn(p))return v(a,c,p,g);zr(a,p)}return typeof p=="string"&&p!==""||typeof p=="number"?(p=""+p,c!==null&&c.tag===6?(n(a,c.sibling),c=l(c,p),c.return=a,a=c):(n(a,c),c=ri(p,a.mode,g),c.return=a,a=c),o(a)):n(a,c)}return D}var gn=wa(!0),xa=wa(!1),ol=Nt(null),ul=null,on=null,Eo=null;function No(){Eo=on=ul=null}function _o(e){var t=ol.current;V(ol),e._currentValue=t}function Fi(e,t,n){for(;e!==null;){var r=e.alternate;if((e.childLanes&t)!==t?(e.childLanes|=t,r!==null&&(r.childLanes|=t)):r!==null&&(r.childLanes&t)!==t&&(r.childLanes|=t),e===n)break;e=e.return}}function pn(e,t){ul=e,Eo=on=null,e=e.dependencies,e!==null&&e.firstContext!==null&&(e.lanes&t&&(ge=!0),e.firstContext=null)}function Ie(e){var t=e._currentValue;if(Eo!==e)if(e={context:e,memoizedValue:t,next:null},on===null){if(ul===null)throw Error(S(308));on=e,ul.dependencies={lanes:0,firstContext:e}}else on=on.next=e;return t}var Rt=null;function jo(e){Rt===null?Rt=[e]:Rt.push(e)}function ka(e,t,n,r){var l=t.interleaved;return l===null?(n.next=n,jo(t)):(n.next=l.next,l.next=n),t.interleaved=n,rt(e,r)}function rt(e,t){e.lanes|=t;var n=e.alternate;for(n!==null&&(n.lanes|=t),n=e,e=e.return;e!==null;)e.childLanes|=t,n=e.alternate,n!==null&&(n.childLanes|=t),n=e,e=e.return;return n.tag===3?n.stateNode:null}var at=!1;function Po(e){e.updateQueue={baseState:e.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,interleaved:null,lanes:0},effects:null}}function Ca(e,t){e=e.updateQueue,t.updateQueue===e&&(t.updateQueue={baseState:e.baseState,firstBaseUpdate:e.firstBaseUpdate,lastBaseUpdate:e.lastBaseUpdate,shared:e.shared,effects:e.effects})}function et(e,t){return{eventTime:e,lane:t,tag:0,payload:null,callback:null,next:null}}function gt(e,t,n){var r=e.updateQueue;if(r===null)return null;if(r=r.shared,M&2){var l=r.pending;return l===null?t.next=t:(t.next=l.next,l.next=t),r.pending=t,rt(e,n)}return l=r.interleaved,l===null?(t.next=t,jo(r)):(t.next=l.next,l.next=t),r.interleaved=t,rt(e,n)}function $r(e,t,n){if(t=t.updateQueue,t!==null&&(t=t.shared,(n&4194240)!==0)){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,po(e,n)}}function Iu(e,t){var n=e.updateQueue,r=e.alternate;if(r!==null&&(r=r.updateQueue,n===r)){var l=null,i=null;if(n=n.firstBaseUpdate,n!==null){do{var o={eventTime:n.eventTime,lane:n.lane,tag:n.tag,payload:n.payload,callback:n.callback,next:null};i===null?l=i=o:i=i.next=o,n=n.next}while(n!==null);i===null?l=i=t:i=i.next=t}else l=i=t;n={baseState:r.baseState,firstBaseUpdate:l,lastBaseUpdate:i,shared:r.shared,effects:r.effects},e.updateQueue=n;return}e=n.lastBaseUpdate,e===null?n.firstBaseUpdate=t:e.next=t,n.lastBaseUpdate=t}function sl(e,t,n,r){var l=e.updateQueue;at=!1;var i=l.firstBaseUpdate,o=l.lastBaseUpdate,u=l.shared.pending;if(u!==null){l.shared.pending=null;var s=u,d=s.next;s.next=null,o===null?i=d:o.next=d,o=s;var y=e.alternate;y!==null&&(y=y.updateQueue,u=y.lastBaseUpdate,u!==o&&(u===null?y.firstBaseUpdate=d:u.next=d,y.lastBaseUpdate=s))}if(i!==null){var h=l.baseState;o=0,y=d=s=null,u=i;do{var m=u.lane,w=u.eventTime;if((r&m)===m){y!==null&&(y=y.next={eventTime:w,lane:0,tag:u.tag,payload:u.payload,callback:u.callback,next:null});e:{var k=e,v=u;switch(m=t,w=n,v.tag){case 1:if(k=v.payload,typeof k=="function"){h=k.call(w,h,m);break e}h=k;break e;case 3:k.flags=k.flags&-65537|128;case 0:if(k=v.payload,m=typeof k=="function"?k.call(w,h,m):k,m==null)break e;h=J({},h,m);break e;case 2:at=!0}}u.callback!==null&&u.lane!==0&&(e.flags|=64,m=l.effects,m===null?l.effects=[u]:m.push(u))}else w={eventTime:w,lane:m,tag:u.tag,payload:u.payload,callback:u.callback,next:null},y===null?(d=y=w,s=h):y=y.next=w,o|=m;if(u=u.next,u===null){if(u=l.shared.pending,u===null)break;m=u,u=m.next,m.next=null,l.lastBaseUpdate=m,l.shared.pending=null}}while(!0);if(y===null&&(s=h),l.baseState=s,l.firstBaseUpdate=d,l.lastBaseUpdate=y,t=l.shared.interleaved,t!==null){l=t;do o|=l.lane,l=l.next;while(l!==t)}else i===null&&(l.shared.lanes=0);Bt|=o,e.lanes=o,e.memoizedState=h}}function Ru(e,t,n){if(e=t.effects,t.effects=null,e!==null)for(t=0;t<e.length;t++){var r=e[t],l=r.callback;if(l!==null){if(r.callback=null,r=n,typeof l!="function")throw Error(S(191,l));l.call(r)}}}var vr={},Ye=Nt(vr),or=Nt(vr),ur=Nt(vr);function Dt(e){if(e===vr)throw Error(S(174));return e}function To(e,t){switch($(ur,t),$(or,e),$(Ye,vr),e=t.nodeType,e){case 9:case 11:t=(t=t.documentElement)?t.namespaceURI:vi(null,"");break;default:e=e===8?t.parentNode:t,t=e.namespaceURI||null,e=e.tagName,t=vi(t,e)}V(Ye),$(Ye,t)}function Sn(){V(Ye),V(or),V(ur)}function Ea(e){Dt(ur.current);var t=Dt(Ye.current),n=vi(t,e.type);t!==n&&($(or,e),$(Ye,n))}function zo(e){or.current===e&&(V(Ye),V(or))}var Y=Nt(0);function al(e){for(var t=e;t!==null;){if(t.tag===13){var n=t.memoizedState;if(n!==null&&(n=n.dehydrated,n===null||n.data==="$?"||n.data==="$!"))return t}else if(t.tag===19&&t.memoizedProps.revealOrder!==void 0){if(t.flags&128)return t}else if(t.child!==null){t.child.return=t,t=t.child;continue}if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}var Zl=[];function Lo(){for(var e=0;e<Zl.length;e++)Zl[e]._workInProgressVersionPrimary=null;Zl.length=0}var Br=it.ReactCurrentDispatcher,ql=it.ReactCurrentBatchConfig,$t=0,G=null,ne=null,le=null,cl=!1,Qn=!1,sr=0,_d=0;function ae(){throw Error(S(321))}function Oo(e,t){if(t===null)return!1;for(var n=0;n<t.length&&n<e.length;n++)if(!Be(e[n],t[n]))return!1;return!0}function Io(e,t,n,r,l,i){if($t=i,G=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,Br.current=e===null||e.memoizedState===null?zd:Ld,e=n(r,l),Qn){i=0;do{if(Qn=!1,sr=0,25<=i)throw Error(S(301));i+=1,le=ne=null,t.updateQueue=null,Br.current=Od,e=n(r,l)}while(Qn)}if(Br.current=fl,t=ne!==null&&ne.next!==null,$t=0,le=ne=G=null,cl=!1,t)throw Error(S(300));return e}function Ro(){var e=sr!==0;return sr=0,e}function We(){var e={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return le===null?G.memoizedState=le=e:le=le.next=e,le}function Re(){if(ne===null){var e=G.alternate;e=e!==null?e.memoizedState:null}else e=ne.next;var t=le===null?G.memoizedState:le.next;if(t!==null)le=t,ne=e;else{if(e===null)throw Error(S(310));ne=e,e={memoizedState:ne.memoizedState,baseState:ne.baseState,baseQueue:ne.baseQueue,queue:ne.queue,next:null},le===null?G.memoizedState=le=e:le=le.next=e}return le}function ar(e,t){return typeof t=="function"?t(e):t}function bl(e){var t=Re(),n=t.queue;if(n===null)throw Error(S(311));n.lastRenderedReducer=e;var r=ne,l=r.baseQueue,i=n.pending;if(i!==null){if(l!==null){var o=l.next;l.next=i.next,i.next=o}r.baseQueue=l=i,n.pending=null}if(l!==null){i=l.next,r=r.baseState;var u=o=null,s=null,d=i;do{var y=d.lane;if(($t&y)===y)s!==null&&(s=s.next={lane:0,action:d.action,hasEagerState:d.hasEagerState,eagerState:d.eagerState,next:null}),r=d.hasEagerState?d.eagerState:e(r,d.action);else{var h={lane:y,action:d.action,hasEagerState:d.hasEagerState,eagerState:d.eagerState,next:null};s===null?(u=s=h,o=r):s=s.next=h,G.lanes|=y,Bt|=y}d=d.next}while(d!==null&&d!==i);s===null?o=r:s.next=u,Be(r,t.memoizedState)||(ge=!0),t.memoizedState=r,t.baseState=o,t.baseQueue=s,n.lastRenderedState=r}if(e=n.interleaved,e!==null){l=e;do i=l.lane,G.lanes|=i,Bt|=i,l=l.next;while(l!==e)}else l===null&&(n.lanes=0);return[t.memoizedState,n.dispatch]}function ei(e){var t=Re(),n=t.queue;if(n===null)throw Error(S(311));n.lastRenderedReducer=e;var r=n.dispatch,l=n.pending,i=t.memoizedState;if(l!==null){n.pending=null;var o=l=l.next;do i=e(i,o.action),o=o.next;while(o!==l);Be(i,t.memoizedState)||(ge=!0),t.memoizedState=i,t.baseQueue===null&&(t.baseState=i),n.lastRenderedState=i}return[i,r]}function Na(){}function _a(e,t){var n=G,r=Re(),l=t(),i=!Be(r.memoizedState,l);if(i&&(r.memoizedState=l,ge=!0),r=r.queue,Do(Ta.bind(null,n,r,e),[e]),r.getSnapshot!==t||i||le!==null&&le.memoizedState.tag&1){if(n.flags|=2048,cr(9,Pa.bind(null,n,r,l,t),void 0,null),ie===null)throw Error(S(349));$t&30||ja(n,t,l)}return l}function ja(e,t,n){e.flags|=16384,e={getSnapshot:t,value:n},t=G.updateQueue,t===null?(t={lastEffect:null,stores:null},G.updateQueue=t,t.stores=[e]):(n=t.stores,n===null?t.stores=[e]:n.push(e))}function Pa(e,t,n,r){t.value=n,t.getSnapshot=r,za(t)&&La(e)}function Ta(e,t,n){return n(function(){za(t)&&La(e)})}function za(e){var t=e.getSnapshot;e=e.value;try{var n=t();return!Be(e,n)}catch{return!0}}function La(e){var t=rt(e,1);t!==null&&$e(t,e,1,-1)}function Du(e){var t=We();return typeof e=="function"&&(e=e()),t.memoizedState=t.baseState=e,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:ar,lastRenderedState:e},t.queue=e,e=e.dispatch=Td.bind(null,G,e),[t.memoizedState,e]}function cr(e,t,n,r){return e={tag:e,create:t,destroy:n,deps:r,next:null},t=G.updateQueue,t===null?(t={lastEffect:null,stores:null},G.updateQueue=t,t.lastEffect=e.next=e):(n=t.lastEffect,n===null?t.lastEffect=e.next=e:(r=n.next,n.next=e,e.next=r,t.lastEffect=e)),e}function Oa(){return Re().memoizedState}function Hr(e,t,n,r){var l=We();G.flags|=e,l.memoizedState=cr(1|t,n,void 0,r===void 0?null:r)}function El(e,t,n,r){var l=Re();r=r===void 0?null:r;var i=void 0;if(ne!==null){var o=ne.memoizedState;if(i=o.destroy,r!==null&&Oo(r,o.deps)){l.memoizedState=cr(t,n,i,r);return}}G.flags|=e,l.memoizedState=cr(1|t,n,i,r)}function Fu(e,t){return Hr(8390656,8,e,t)}function Do(e,t){return El(2048,8,e,t)}function Ia(e,t){return El(4,2,e,t)}function Ra(e,t){return El(4,4,e,t)}function Da(e,t){if(typeof t=="function")return e=e(),t(e),function(){t(null)};if(t!=null)return e=e(),t.current=e,function(){t.current=null}}function Fa(e,t,n){return n=n!=null?n.concat([e]):null,El(4,4,Da.bind(null,t,e),n)}function Fo(){}function Ma(e,t){var n=Re();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&Oo(t,r[1])?r[0]:(n.memoizedState=[e,t],e)}function Aa(e,t){var n=Re();t=t===void 0?null:t;var r=n.memoizedState;return r!==null&&t!==null&&Oo(t,r[1])?r[0]:(e=e(),n.memoizedState=[e,t],e)}function Ua(e,t,n){return $t&21?(Be(n,t)||(n=Ws(),G.lanes|=n,Bt|=n,e.baseState=!0),t):(e.baseState&&(e.baseState=!1,ge=!0),e.memoizedState=n)}function jd(e,t){var n=A;A=n!==0&&4>n?n:4,e(!0);var r=ql.transition;ql.transition={};try{e(!1),t()}finally{A=n,ql.transition=r}}function $a(){return Re().memoizedState}function Pd(e,t,n){var r=wt(e);if(n={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null},Ba(e))Ha(t,n);else if(n=ka(e,t,n,r),n!==null){var l=me();$e(n,e,r,l),Va(n,t,r)}}function Td(e,t,n){var r=wt(e),l={lane:r,action:n,hasEagerState:!1,eagerState:null,next:null};if(Ba(e))Ha(t,l);else{var i=e.alternate;if(e.lanes===0&&(i===null||i.lanes===0)&&(i=t.lastRenderedReducer,i!==null))try{var o=t.lastRenderedState,u=i(o,n);if(l.hasEagerState=!0,l.eagerState=u,Be(u,o)){var s=t.interleaved;s===null?(l.next=l,jo(t)):(l.next=s.next,s.next=l),t.interleaved=l;return}}catch{}finally{}n=ka(e,t,l,r),n!==null&&(l=me(),$e(n,e,r,l),Va(n,t,r))}}function Ba(e){var t=e.alternate;return e===G||t!==null&&t===G}function Ha(e,t){Qn=cl=!0;var n=e.pending;n===null?t.next=t:(t.next=n.next,n.next=t),e.pending=t}function Va(e,t,n){if(n&4194240){var r=t.lanes;r&=e.pendingLanes,n|=r,t.lanes=n,po(e,n)}}var fl={readContext:Ie,useCallback:ae,useContext:ae,useEffect:ae,useImperativeHandle:ae,useInsertionEffect:ae,useLayoutEffect:ae,useMemo:ae,useReducer:ae,useRef:ae,useState:ae,useDebugValue:ae,useDeferredValue:ae,useTransition:ae,useMutableSource:ae,useSyncExternalStore:ae,useId:ae,unstable_isNewReconciler:!1},zd={readContext:Ie,useCallback:function(e,t){return We().memoizedState=[e,t===void 0?null:t],e},useContext:Ie,useEffect:Fu,useImperativeHandle:function(e,t,n){return n=n!=null?n.concat([e]):null,Hr(4194308,4,Da.bind(null,t,e),n)},useLayoutEffect:function(e,t){return Hr(4194308,4,e,t)},useInsertionEffect:function(e,t){return Hr(4,2,e,t)},useMemo:function(e,t){var n=We();return t=t===void 0?null:t,e=e(),n.memoizedState=[e,t],e},useReducer:function(e,t,n){var r=We();return t=n!==void 0?n(t):t,r.memoizedState=r.baseState=t,e={pending:null,interleaved:null,lanes:0,dispatch:null,lastRenderedReducer:e,lastRenderedState:t},r.queue=e,e=e.dispatch=Pd.bind(null,G,e),[r.memoizedState,e]},useRef:function(e){var t=We();return e={current:e},t.memoizedState=e},useState:Du,useDebugValue:Fo,useDeferredValue:function(e){return We().memoizedState=e},useTransition:function(){var e=Du(!1),t=e[0];return e=jd.bind(null,e[1]),We().memoizedState=e,[t,e]},useMutableSource:function(){},useSyncExternalStore:function(e,t,n){var r=G,l=We();if(Q){if(n===void 0)throw Error(S(407));n=n()}else{if(n=t(),ie===null)throw Error(S(349));$t&30||ja(r,t,n)}l.memoizedState=n;var i={value:n,getSnapshot:t};return l.queue=i,Fu(Ta.bind(null,r,i,e),[e]),r.flags|=2048,cr(9,Pa.bind(null,r,i,n,t),void 0,null),n},useId:function(){var e=We(),t=ie.identifierPrefix;if(Q){var n=be,r=qe;n=(r&~(1<<32-Ue(r)-1)).toString(32)+n,t=":"+t+"R"+n,n=sr++,0<n&&(t+="H"+n.toString(32)),t+=":"}else n=_d++,t=":"+t+"r"+n.toString(32)+":";return e.memoizedState=t},unstable_isNewReconciler:!1},Ld={readContext:Ie,useCallback:Ma,useContext:Ie,useEffect:Do,useImperativeHandle:Fa,useInsertionEffect:Ia,useLayoutEffect:Ra,useMemo:Aa,useReducer:bl,useRef:Oa,useState:function(){return bl(ar)},useDebugValue:Fo,useDeferredValue:function(e){var t=Re();return Ua(t,ne.memoizedState,e)},useTransition:function(){var e=bl(ar)[0],t=Re().memoizedState;return[e,t]},useMutableSource:Na,useSyncExternalStore:_a,useId:$a,unstable_isNewReconciler:!1},Od={readContext:Ie,useCallback:Ma,useContext:Ie,useEffect:Do,useImperativeHandle:Fa,useInsertionEffect:Ia,useLayoutEffect:Ra,useMemo:Aa,useReducer:ei,useRef:Oa,useState:function(){return ei(ar)},useDebugValue:Fo,useDeferredValue:function(e){var t=Re();return ne===null?t.memoizedState=e:Ua(t,ne.memoizedState,e)},useTransition:function(){var e=ei(ar)[0],t=Re().memoizedState;return[e,t]},useMutableSource:Na,useSyncExternalStore:_a,useId:$a,unstable_isNewReconciler:!1};function Fe(e,t){if(e&&e.defaultProps){t=J({},t),e=e.defaultProps;for(var n in e)t[n]===void 0&&(t[n]=e[n]);return t}return t}function Mi(e,t,n,r){t=e.memoizedState,n=n(r,t),n=n==null?t:J({},t,n),e.memoizedState=n,e.lanes===0&&(e.updateQueue.baseState=n)}var Nl={isMounted:function(e){return(e=e._reactInternals)?Wt(e)===e:!1},enqueueSetState:function(e,t,n){e=e._reactInternals;var r=me(),l=wt(e),i=et(r,l);i.payload=t,n!=null&&(i.callback=n),t=gt(e,i,l),t!==null&&($e(t,e,l,r),$r(t,e,l))},enqueueReplaceState:function(e,t,n){e=e._reactInternals;var r=me(),l=wt(e),i=et(r,l);i.tag=1,i.payload=t,n!=null&&(i.callback=n),t=gt(e,i,l),t!==null&&($e(t,e,l,r),$r(t,e,l))},enqueueForceUpdate:function(e,t){e=e._reactInternals;var n=me(),r=wt(e),l=et(n,r);l.tag=2,t!=null&&(l.callback=t),t=gt(e,l,r),t!==null&&($e(t,e,r,n),$r(t,e,r))}};function Mu(e,t,n,r,l,i,o){return e=e.stateNode,typeof e.shouldComponentUpdate=="function"?e.shouldComponentUpdate(r,i,o):t.prototype&&t.prototype.isPureReactComponent?!nr(n,r)||!nr(l,i):!0}function Wa(e,t,n){var r=!1,l=Ct,i=t.contextType;return typeof i=="object"&&i!==null?i=Ie(i):(l=we(t)?At:de.current,r=t.contextTypes,i=(r=r!=null)?vn(e,l):Ct),t=new t(n,i),e.memoizedState=t.state!==null&&t.state!==void 0?t.state:null,t.updater=Nl,e.stateNode=t,t._reactInternals=e,r&&(e=e.stateNode,e.__reactInternalMemoizedUnmaskedChildContext=l,e.__reactInternalMemoizedMaskedChildContext=i),t}function Au(e,t,n,r){e=t.state,typeof t.componentWillReceiveProps=="function"&&t.componentWillReceiveProps(n,r),typeof t.UNSAFE_componentWillReceiveProps=="function"&&t.UNSAFE_componentWillReceiveProps(n,r),t.state!==e&&Nl.enqueueReplaceState(t,t.state,null)}function Ai(e,t,n,r){var l=e.stateNode;l.props=n,l.state=e.memoizedState,l.refs={},Po(e);var i=t.contextType;typeof i=="object"&&i!==null?l.context=Ie(i):(i=we(t)?At:de.current,l.context=vn(e,i)),l.state=e.memoizedState,i=t.getDerivedStateFromProps,typeof i=="function"&&(Mi(e,t,i,n),l.state=e.memoizedState),typeof t.getDerivedStateFromProps=="function"||typeof l.getSnapshotBeforeUpdate=="function"||typeof l.UNSAFE_componentWillMount!="function"&&typeof l.componentWillMount!="function"||(t=l.state,typeof l.componentWillMount=="function"&&l.componentWillMount(),typeof l.UNSAFE_componentWillMount=="function"&&l.UNSAFE_componentWillMount(),t!==l.state&&Nl.enqueueReplaceState(l,l.state,null),sl(e,n,l,r),l.state=e.memoizedState),typeof l.componentDidMount=="function"&&(e.flags|=4194308)}function wn(e,t){try{var n="",r=t;do n+=of(r),r=r.return;while(r);var l=n}catch(i){l=`
-Error generating stack: `+i.message+`
-`+i.stack}return{value:e,source:t,stack:l,digest:null}}function ti(e,t,n){return{value:e,source:null,stack:n??null,digest:t??null}}function Ui(e,t){try{console.error(t.value)}catch(n){setTimeout(function(){throw n})}}var Id=typeof WeakMap=="function"?WeakMap:Map;function Qa(e,t,n){n=et(-1,n),n.tag=3,n.payload={element:null};var r=t.value;return n.callback=function(){pl||(pl=!0,Ji=r),Ui(e,t)},n}function Ka(e,t,n){n=et(-1,n),n.tag=3;var r=e.type.getDerivedStateFromError;if(typeof r=="function"){var l=t.value;n.payload=function(){return r(l)},n.callback=function(){Ui(e,t)}}var i=e.stateNode;return i!==null&&typeof i.componentDidCatch=="function"&&(n.callback=function(){Ui(e,t),typeof r!="function"&&(St===null?St=new Set([this]):St.add(this));var o=t.stack;this.componentDidCatch(t.value,{componentStack:o!==null?o:""})}),n}function Uu(e,t,n){var r=e.pingCache;if(r===null){r=e.pingCache=new Id;var l=new Set;r.set(t,l)}else l=r.get(t),l===void 0&&(l=new Set,r.set(t,l));l.has(n)||(l.add(n),e=Yd.bind(null,e,t,n),t.then(e,e))}function $u(e){do{var t;if((t=e.tag===13)&&(t=e.memoizedState,t=t!==null?t.dehydrated!==null:!0),t)return e;e=e.return}while(e!==null);return null}function Bu(e,t,n,r,l){return e.mode&1?(e.flags|=65536,e.lanes=l,e):(e===t?e.flags|=65536:(e.flags|=128,n.flags|=131072,n.flags&=-52805,n.tag===1&&(n.alternate===null?n.tag=17:(t=et(-1,1),t.tag=2,gt(n,t,1))),n.lanes|=1),e)}var Rd=it.ReactCurrentOwner,ge=!1;function pe(e,t,n,r){t.child=e===null?xa(t,null,n,r):gn(t,e.child,n,r)}function Hu(e,t,n,r,l){n=n.render;var i=t.ref;return pn(t,l),r=Io(e,t,n,r,i,l),n=Ro(),e!==null&&!ge?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,lt(e,t,l)):(Q&&n&&xo(t),t.flags|=1,pe(e,t,r,l),t.child)}function Vu(e,t,n,r,l){if(e===null){var i=n.type;return typeof i=="function"&&!Wo(i)&&i.defaultProps===void 0&&n.compare===null&&n.defaultProps===void 0?(t.tag=15,t.type=i,Ya(e,t,i,r,l)):(e=Kr(n.type,null,r,t,t.mode,l),e.ref=t.ref,e.return=t,t.child=e)}if(i=e.child,!(e.lanes&l)){var o=i.memoizedProps;if(n=n.compare,n=n!==null?n:nr,n(o,r)&&e.ref===t.ref)return lt(e,t,l)}return t.flags|=1,e=xt(i,r),e.ref=t.ref,e.return=t,t.child=e}function Ya(e,t,n,r,l){if(e!==null){var i=e.memoizedProps;if(nr(i,r)&&e.ref===t.ref)if(ge=!1,t.pendingProps=r=i,(e.lanes&l)!==0)e.flags&131072&&(ge=!0);else return t.lanes=e.lanes,lt(e,t,l)}return $i(e,t,n,r,l)}function Ga(e,t,n){var r=t.pendingProps,l=r.children,i=e!==null?e.memoizedState:null;if(r.mode==="hidden")if(!(t.mode&1))t.memoizedState={baseLanes:0,cachePool:null,transitions:null},$(sn,Ce),Ce|=n;else{if(!(n&1073741824))return e=i!==null?i.baseLanes|n:n,t.lanes=t.childLanes=1073741824,t.memoizedState={baseLanes:e,cachePool:null,transitions:null},t.updateQueue=null,$(sn,Ce),Ce|=e,null;t.memoizedState={baseLanes:0,cachePool:null,transitions:null},r=i!==null?i.baseLanes:n,$(sn,Ce),Ce|=r}else i!==null?(r=i.baseLanes|n,t.memoizedState=null):r=n,$(sn,Ce),Ce|=r;return pe(e,t,l,n),t.child}function Ja(e,t){var n=t.ref;(e===null&&n!==null||e!==null&&e.ref!==n)&&(t.flags|=512,t.flags|=2097152)}function $i(e,t,n,r,l){var i=we(n)?At:de.current;return i=vn(t,i),pn(t,l),n=Io(e,t,n,r,i,l),r=Ro(),e!==null&&!ge?(t.updateQueue=e.updateQueue,t.flags&=-2053,e.lanes&=~l,lt(e,t,l)):(Q&&r&&xo(t),t.flags|=1,pe(e,t,n,l),t.child)}function Wu(e,t,n,r,l){if(we(n)){var i=!0;rl(t)}else i=!1;if(pn(t,l),t.stateNode===null)Vr(e,t),Wa(t,n,r),Ai(t,n,r,l),r=!0;else if(e===null){var o=t.stateNode,u=t.memoizedProps;o.props=u;var s=o.context,d=n.contextType;typeof d=="object"&&d!==null?d=Ie(d):(d=we(n)?At:de.current,d=vn(t,d));var y=n.getDerivedStateFromProps,h=typeof y=="function"||typeof o.getSnapshotBeforeUpdate=="function";h||typeof o.UNSAFE_componentWillReceiveProps!="function"&&typeof o.componentWillReceiveProps!="function"||(u!==r||s!==d)&&Au(t,o,r,d),at=!1;var m=t.memoizedState;o.state=m,sl(t,r,o,l),s=t.memoizedState,u!==r||m!==s||Se.current||at?(typeof y=="function"&&(Mi(t,n,y,r),s=t.memoizedState),(u=at||Mu(t,n,u,r,m,s,d))?(h||typeof o.UNSAFE_componentWillMount!="function"&&typeof o.componentWillMount!="function"||(typeof o.componentWillMount=="function"&&o.componentWillMount(),typeof o.UNSAFE_componentWillMount=="function"&&o.UNSAFE_componentWillMount()),typeof o.componentDidMount=="function"&&(t.flags|=4194308)):(typeof o.componentDidMount=="function"&&(t.flags|=4194308),t.memoizedProps=r,t.memoizedState=s),o.props=r,o.state=s,o.context=d,r=u):(typeof o.componentDidMount=="function"&&(t.flags|=4194308),r=!1)}else{o=t.stateNode,Ca(e,t),u=t.memoizedProps,d=t.type===t.elementType?u:Fe(t.type,u),o.props=d,h=t.pendingProps,m=o.context,s=n.contextType,typeof s=="object"&&s!==null?s=Ie(s):(s=we(n)?At:de.current,s=vn(t,s));var w=n.getDerivedStateFromProps;(y=typeof w=="function"||typeof o.getSnapshotBeforeUpdate=="function")||typeof o.UNSAFE_componentWillReceiveProps!="function"&&typeof o.componentWillReceiveProps!="function"||(u!==h||m!==s)&&Au(t,o,r,s),at=!1,m=t.memoizedState,o.state=m,sl(t,r,o,l);var k=t.memoizedState;u!==h||m!==k||Se.current||at?(typeof w=="function"&&(Mi(t,n,w,r),k=t.memoizedState),(d=at||Mu(t,n,d,r,m,k,s)||!1)?(y||typeof o.UNSAFE_componentWillUpdate!="function"&&typeof o.componentWillUpdate!="function"||(typeof o.componentWillUpdate=="function"&&o.componentWillUpdate(r,k,s),typeof o.UNSAFE_componentWillUpdate=="function"&&o.UNSAFE_componentWillUpdate(r,k,s)),typeof o.componentDidUpdate=="function"&&(t.flags|=4),typeof o.getSnapshotBeforeUpdate=="function"&&(t.flags|=1024)):(typeof o.componentDidUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=4),typeof o.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=1024),t.memoizedProps=r,t.memoizedState=k),o.props=r,o.state=k,o.context=s,r=d):(typeof o.componentDidUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=4),typeof o.getSnapshotBeforeUpdate!="function"||u===e.memoizedProps&&m===e.memoizedState||(t.flags|=1024),r=!1)}return Bi(e,t,n,r,i,l)}function Bi(e,t,n,r,l,i){Ja(e,t);var o=(t.flags&128)!==0;if(!r&&!o)return l&&Tu(t,n,!1),lt(e,t,i);r=t.stateNode,Rd.current=t;var u=o&&typeof n.getDerivedStateFromError!="function"?null:r.render();return t.flags|=1,e!==null&&o?(t.child=gn(t,e.child,null,i),t.child=gn(t,null,u,i)):pe(e,t,u,i),t.memoizedState=r.state,l&&Tu(t,n,!0),t.child}function Xa(e){var t=e.stateNode;t.pendingContext?Pu(e,t.pendingContext,t.pendingContext!==t.context):t.context&&Pu(e,t.context,!1),To(e,t.containerInfo)}function Qu(e,t,n,r,l){return yn(),Co(l),t.flags|=256,pe(e,t,n,r),t.child}var Hi={dehydrated:null,treeContext:null,retryLane:0};function Vi(e){return{baseLanes:e,cachePool:null,transitions:null}}function Za(e,t,n){var r=t.pendingProps,l=Y.current,i=!1,o=(t.flags&128)!==0,u;if((u=o)||(u=e!==null&&e.memoizedState===null?!1:(l&2)!==0),u?(i=!0,t.flags&=-129):(e===null||e.memoizedState!==null)&&(l|=1),$(Y,l&1),e===null)return Di(t),e=t.memoizedState,e!==null&&(e=e.dehydrated,e!==null)?(t.mode&1?e.data==="$!"?t.lanes=8:t.lanes=1073741824:t.lanes=1,null):(o=r.children,e=r.fallback,i?(r=t.mode,i=t.child,o={mode:"hidden",children:o},!(r&1)&&i!==null?(i.childLanes=0,i.pendingProps=o):i=Pl(o,r,0,null),e=Mt(e,r,n,null),i.return=t,e.return=t,i.sibling=e,t.child=i,t.child.memoizedState=Vi(n),t.memoizedState=Hi,e):Mo(t,o));if(l=e.memoizedState,l!==null&&(u=l.dehydrated,u!==null))return Dd(e,t,o,r,u,l,n);if(i){i=r.fallback,o=t.mode,l=e.child,u=l.sibling;var s={mode:"hidden",children:r.children};return!(o&1)&&t.child!==l?(r=t.child,r.childLanes=0,r.pendingProps=s,t.deletions=null):(r=xt(l,s),r.subtreeFlags=l.subtreeFlags&14680064),u!==null?i=xt(u,i):(i=Mt(i,o,n,null),i.flags|=2),i.return=t,r.return=t,r.sibling=i,t.child=r,r=i,i=t.child,o=e.child.memoizedState,o=o===null?Vi(n):{baseLanes:o.baseLanes|n,cachePool:null,transitions:o.transitions},i.memoizedState=o,i.childLanes=e.childLanes&~n,t.memoizedState=Hi,r}return i=e.child,e=i.sibling,r=xt(i,{mode:"visible",children:r.children}),!(t.mode&1)&&(r.lanes=n),r.return=t,r.sibling=null,e!==null&&(n=t.deletions,n===null?(t.deletions=[e],t.flags|=16):n.push(e)),t.child=r,t.memoizedState=null,r}function Mo(e,t){return t=Pl({mode:"visible",children:t},e.mode,0,null),t.return=e,e.child=t}function Lr(e,t,n,r){return r!==null&&Co(r),gn(t,e.child,null,n),e=Mo(t,t.pendingProps.children),e.flags|=2,t.memoizedState=null,e}function Dd(e,t,n,r,l,i,o){if(n)return t.flags&256?(t.flags&=-257,r=ti(Error(S(422))),Lr(e,t,o,r)):t.memoizedState!==null?(t.child=e.child,t.flags|=128,null):(i=r.fallback,l=t.mode,r=Pl({mode:"visible",children:r.children},l,0,null),i=Mt(i,l,o,null),i.flags|=2,r.return=t,i.return=t,r.sibling=i,t.child=r,t.mode&1&&gn(t,e.child,null,o),t.child.memoizedState=Vi(o),t.memoizedState=Hi,i);if(!(t.mode&1))return Lr(e,t,o,null);if(l.data==="$!"){if(r=l.nextSibling&&l.nextSibling.dataset,r)var u=r.dgst;return r=u,i=Error(S(419)),r=ti(i,r,void 0),Lr(e,t,o,r)}if(u=(o&e.childLanes)!==0,ge||u){if(r=ie,r!==null){switch(o&-o){case 4:l=2;break;case 16:l=8;break;case 64:case 128:case 256:case 512:case 1024:case 2048:case 4096:case 8192:case 16384:case 32768:case 65536:case 131072:case 262144:case 524288:case 1048576:case 2097152:case 4194304:case 8388608:case 16777216:case 33554432:case 67108864:l=32;break;case 536870912:l=268435456;break;default:l=0}l=l&(r.suspendedLanes|o)?0:l,l!==0&&l!==i.retryLane&&(i.retryLane=l,rt(e,l),$e(r,e,l,-1))}return Vo(),r=ti(Error(S(421))),Lr(e,t,o,r)}return l.data==="$?"?(t.flags|=128,t.child=e.child,t=Gd.bind(null,e),l._reactRetry=t,null):(e=i.treeContext,Ee=yt(l.nextSibling),Ne=t,Q=!0,Ae=null,e!==null&&(Te[ze++]=qe,Te[ze++]=be,Te[ze++]=Ut,qe=e.id,be=e.overflow,Ut=t),t=Mo(t,r.children),t.flags|=4096,t)}function Ku(e,t,n){e.lanes|=t;var r=e.alternate;r!==null&&(r.lanes|=t),Fi(e.return,t,n)}function ni(e,t,n,r,l){var i=e.memoizedState;i===null?e.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:r,tail:n,tailMode:l}:(i.isBackwards=t,i.rendering=null,i.renderingStartTime=0,i.last=r,i.tail=n,i.tailMode=l)}function qa(e,t,n){var r=t.pendingProps,l=r.revealOrder,i=r.tail;if(pe(e,t,r.children,n),r=Y.current,r&2)r=r&1|2,t.flags|=128;else{if(e!==null&&e.flags&128)e:for(e=t.child;e!==null;){if(e.tag===13)e.memoizedState!==null&&Ku(e,n,t);else if(e.tag===19)Ku(e,n,t);else if(e.child!==null){e.child.return=e,e=e.child;continue}if(e===t)break e;for(;e.sibling===null;){if(e.return===null||e.return===t)break e;e=e.return}e.sibling.return=e.return,e=e.sibling}r&=1}if($(Y,r),!(t.mode&1))t.memoizedState=null;else switch(l){case"forwards":for(n=t.child,l=null;n!==null;)e=n.alternate,e!==null&&al(e)===null&&(l=n),n=n.sibling;n=l,n===null?(l=t.child,t.child=null):(l=n.sibling,n.sibling=null),ni(t,!1,l,n,i);break;case"backwards":for(n=null,l=t.child,t.child=null;l!==null;){if(e=l.alternate,e!==null&&al(e)===null){t.child=l;break}e=l.sibling,l.sibling=n,n=l,l=e}ni(t,!0,n,null,i);break;case"together":ni(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function Vr(e,t){!(t.mode&1)&&e!==null&&(e.alternate=null,t.alternate=null,t.flags|=2)}function lt(e,t,n){if(e!==null&&(t.dependencies=e.dependencies),Bt|=t.lanes,!(n&t.childLanes))return null;if(e!==null&&t.child!==e.child)throw Error(S(153));if(t.child!==null){for(e=t.child,n=xt(e,e.pendingProps),t.child=n,n.return=t;e.sibling!==null;)e=e.sibling,n=n.sibling=xt(e,e.pendingProps),n.return=t;n.sibling=null}return t.child}function Fd(e,t,n){switch(t.tag){case 3:Xa(t),yn();break;case 5:Ea(t);break;case 1:we(t.type)&&rl(t);break;case 4:To(t,t.stateNode.containerInfo);break;case 10:var r=t.type._context,l=t.memoizedProps.value;$(ol,r._currentValue),r._currentValue=l;break;case 13:if(r=t.memoizedState,r!==null)return r.dehydrated!==null?($(Y,Y.current&1),t.flags|=128,null):n&t.child.childLanes?Za(e,t,n):($(Y,Y.current&1),e=lt(e,t,n),e!==null?e.sibling:null);$(Y,Y.current&1);break;case 19:if(r=(n&t.childLanes)!==0,e.flags&128){if(r)return qa(e,t,n);t.flags|=128}if(l=t.memoizedState,l!==null&&(l.rendering=null,l.tail=null,l.lastEffect=null),$(Y,Y.current),r)break;return null;case 22:case 23:return t.lanes=0,Ga(e,t,n)}return lt(e,t,n)}var ba,Wi,ec,tc;ba=function(e,t){for(var n=t.child;n!==null;){if(n.tag===5||n.tag===6)e.appendChild(n.stateNode);else if(n.tag!==4&&n.child!==null){n.child.return=n,n=n.child;continue}if(n===t)break;for(;n.sibling===null;){if(n.return===null||n.return===t)return;n=n.return}n.sibling.return=n.return,n=n.sibling}};Wi=function(){};ec=function(e,t,n,r){var l=e.memoizedProps;if(l!==r){e=t.stateNode,Dt(Ye.current);var i=null;switch(n){case"input":l=di(e,l),r=di(e,r),i=[];break;case"select":l=J({},l,{value:void 0}),r=J({},r,{value:void 0}),i=[];break;case"textarea":l=hi(e,l),r=hi(e,r),i=[];break;default:typeof l.onClick!="function"&&typeof r.onClick=="function"&&(e.onclick=tl)}yi(n,r);var o;n=null;for(d in l)if(!r.hasOwnProperty(d)&&l.hasOwnProperty(d)&&l[d]!=null)if(d==="style"){var u=l[d];for(o in u)u.hasOwnProperty(o)&&(n||(n={}),n[o]="")}else d!=="dangerouslySetInnerHTML"&&d!=="children"&&d!=="suppressContentEditableWarning"&&d!=="suppressHydrationWarning"&&d!=="autoFocus"&&(Jn.hasOwnProperty(d)?i||(i=[]):(i=i||[]).push(d,null));for(d in r){var s=r[d];if(u=l!=null?l[d]:void 0,r.hasOwnProperty(d)&&s!==u&&(s!=null||u!=null))if(d==="style")if(u){for(o in u)!u.hasOwnProperty(o)||s&&s.hasOwnProperty(o)||(n||(n={}),n[o]="");for(o in s)s.hasOwnProperty(o)&&u[o]!==s[o]&&(n||(n={}),n[o]=s[o])}else n||(i||(i=[]),i.push(d,n)),n=s;else d==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,u=u?u.__html:void 0,s!=null&&u!==s&&(i=i||[]).push(d,s)):d==="children"?typeof s!="string"&&typeof s!="number"||(i=i||[]).push(d,""+s):d!=="suppressContentEditableWarning"&&d!=="suppressHydrationWarning"&&(Jn.hasOwnProperty(d)?(s!=null&&d==="onScroll"&&H("scroll",e),i||u===s||(i=[])):(i=i||[]).push(d,s))}n&&(i=i||[]).push("style",n);var d=i;(t.updateQueue=d)&&(t.flags|=4)}};tc=function(e,t,n,r){n!==r&&(t.flags|=4)};function Rn(e,t){if(!Q)switch(e.tailMode){case"hidden":t=e.tail;for(var n=null;t!==null;)t.alternate!==null&&(n=t),t=t.sibling;n===null?e.tail=null:n.sibling=null;break;case"collapsed":n=e.tail;for(var r=null;n!==null;)n.alternate!==null&&(r=n),n=n.sibling;r===null?t||e.tail===null?e.tail=null:e.tail.sibling=null:r.sibling=null}}function ce(e){var t=e.alternate!==null&&e.alternate.child===e.child,n=0,r=0;if(t)for(var l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags&14680064,r|=l.flags&14680064,l.return=e,l=l.sibling;else for(l=e.child;l!==null;)n|=l.lanes|l.childLanes,r|=l.subtreeFlags,r|=l.flags,l.return=e,l=l.sibling;return e.subtreeFlags|=r,e.childLanes=n,t}function Md(e,t,n){var r=t.pendingProps;switch(ko(t),t.tag){case 2:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return ce(t),null;case 1:return we(t.type)&&nl(),ce(t),null;case 3:return r=t.stateNode,Sn(),V(Se),V(de),Lo(),r.pendingContext&&(r.context=r.pendingContext,r.pendingContext=null),(e===null||e.child===null)&&(Tr(t)?t.flags|=4:e===null||e.memoizedState.isDehydrated&&!(t.flags&256)||(t.flags|=1024,Ae!==null&&(qi(Ae),Ae=null))),Wi(e,t),ce(t),null;case 5:zo(t);var l=Dt(ur.current);if(n=t.type,e!==null&&t.stateNode!=null)ec(e,t,n,r,l),e.ref!==t.ref&&(t.flags|=512,t.flags|=2097152);else{if(!r){if(t.stateNode===null)throw Error(S(166));return ce(t),null}if(e=Dt(Ye.current),Tr(t)){r=t.stateNode,n=t.type;var i=t.memoizedProps;switch(r[Qe]=t,r[ir]=i,e=(t.mode&1)!==0,n){case"dialog":H("cancel",r),H("close",r);break;case"iframe":case"object":case"embed":H("load",r);break;case"video":case"audio":for(l=0;l<Un.length;l++)H(Un[l],r);break;case"source":H("error",r);break;case"img":case"image":case"link":H("error",r),H("load",r);break;case"details":H("toggle",r);break;case"input":tu(r,i),H("invalid",r);break;case"select":r._wrapperState={wasMultiple:!!i.multiple},H("invalid",r);break;case"textarea":ru(r,i),H("invalid",r)}yi(n,i),l=null;for(var o in i)if(i.hasOwnProperty(o)){var u=i[o];o==="children"?typeof u=="string"?r.textContent!==u&&(i.suppressHydrationWarning!==!0&&Pr(r.textContent,u,e),l=["children",u]):typeof u=="number"&&r.textContent!==""+u&&(i.suppressHydrationWarning!==!0&&Pr(r.textContent,u,e),l=["children",""+u]):Jn.hasOwnProperty(o)&&u!=null&&o==="onScroll"&&H("scroll",r)}switch(n){case"input":wr(r),nu(r,i,!0);break;case"textarea":wr(r),lu(r);break;case"select":case"option":break;default:typeof i.onClick=="function"&&(r.onclick=tl)}r=l,t.updateQueue=r,r!==null&&(t.flags|=4)}else{o=l.nodeType===9?l:l.ownerDocument,e==="http://www.w3.org/1999/xhtml"&&(e=Ts(n)),e==="http://www.w3.org/1999/xhtml"?n==="script"?(e=o.createElement("div"),e.innerHTML="<script><\/script>",e=e.removeChild(e.firstChild)):typeof r.is=="string"?e=o.createElement(n,{is:r.is}):(e=o.createElement(n),n==="select"&&(o=e,r.multiple?o.multiple=!0:r.size&&(o.size=r.size))):e=o.createElementNS(e,n),e[Qe]=t,e[ir]=r,ba(e,t,!1,!1),t.stateNode=e;e:{switch(o=gi(n,r),n){case"dialog":H("cancel",e),H("close",e),l=r;break;case"iframe":case"object":case"embed":H("load",e),l=r;break;case"video":case"audio":for(l=0;l<Un.length;l++)H(Un[l],e);l=r;break;case"source":H("error",e),l=r;break;case"img":case"image":case"link":H("error",e),H("load",e),l=r;break;case"details":H("toggle",e),l=r;break;case"input":tu(e,r),l=di(e,r),H("invalid",e);break;case"option":l=r;break;case"select":e._wrapperState={wasMultiple:!!r.multiple},l=J({},r,{value:void 0}),H("invalid",e);break;case"textarea":ru(e,r),l=hi(e,r),H("invalid",e);break;default:l=r}yi(n,l),u=l;for(i in u)if(u.hasOwnProperty(i)){var s=u[i];i==="style"?Os(e,s):i==="dangerouslySetInnerHTML"?(s=s?s.__html:void 0,s!=null&&zs(e,s)):i==="children"?typeof s=="string"?(n!=="textarea"||s!=="")&&Xn(e,s):typeof s=="number"&&Xn(e,""+s):i!=="suppressContentEditableWarning"&&i!=="suppressHydrationWarning"&&i!=="autoFocus"&&(Jn.hasOwnProperty(i)?s!=null&&i==="onScroll"&&H("scroll",e):s!=null&&oo(e,i,s,o))}switch(n){case"input":wr(e),nu(e,r,!1);break;case"textarea":wr(e),lu(e);break;case"option":r.value!=null&&e.setAttribute("value",""+kt(r.value));break;case"select":e.multiple=!!r.multiple,i=r.value,i!=null?an(e,!!r.multiple,i,!1):r.defaultValue!=null&&an(e,!!r.multiple,r.defaultValue,!0);break;default:typeof l.onClick=="function"&&(e.onclick=tl)}switch(n){case"button":case"input":case"select":case"textarea":r=!!r.autoFocus;break e;case"img":r=!0;break e;default:r=!1}}r&&(t.flags|=4)}t.ref!==null&&(t.flags|=512,t.flags|=2097152)}return ce(t),null;case 6:if(e&&t.stateNode!=null)tc(e,t,e.memoizedProps,r);else{if(typeof r!="string"&&t.stateNode===null)throw Error(S(166));if(n=Dt(ur.current),Dt(Ye.current),Tr(t)){if(r=t.stateNode,n=t.memoizedProps,r[Qe]=t,(i=r.nodeValue!==n)&&(e=Ne,e!==null))switch(e.tag){case 3:Pr(r.nodeValue,n,(e.mode&1)!==0);break;case 5:e.memoizedProps.suppressHydrationWarning!==!0&&Pr(r.nodeValue,n,(e.mode&1)!==0)}i&&(t.flags|=4)}else r=(n.nodeType===9?n:n.ownerDocument).createTextNode(r),r[Qe]=t,t.stateNode=r}return ce(t),null;case 13:if(V(Y),r=t.memoizedState,e===null||e.memoizedState!==null&&e.memoizedState.dehydrated!==null){if(Q&&Ee!==null&&t.mode&1&&!(t.flags&128))Sa(),yn(),t.flags|=98560,i=!1;else if(i=Tr(t),r!==null&&r.dehydrated!==null){if(e===null){if(!i)throw Error(S(318));if(i=t.memoizedState,i=i!==null?i.dehydrated:null,!i)throw Error(S(317));i[Qe]=t}else yn(),!(t.flags&128)&&(t.memoizedState=null),t.flags|=4;ce(t),i=!1}else Ae!==null&&(qi(Ae),Ae=null),i=!0;if(!i)return t.flags&65536?t:null}return t.flags&128?(t.lanes=n,t):(r=r!==null,r!==(e!==null&&e.memoizedState!==null)&&r&&(t.child.flags|=8192,t.mode&1&&(e===null||Y.current&1?re===0&&(re=3):Vo())),t.updateQueue!==null&&(t.flags|=4),ce(t),null);case 4:return Sn(),Wi(e,t),e===null&&rr(t.stateNode.containerInfo),ce(t),null;case 10:return _o(t.type._context),ce(t),null;case 17:return we(t.type)&&nl(),ce(t),null;case 19:if(V(Y),i=t.memoizedState,i===null)return ce(t),null;if(r=(t.flags&128)!==0,o=i.rendering,o===null)if(r)Rn(i,!1);else{if(re!==0||e!==null&&e.flags&128)for(e=t.child;e!==null;){if(o=al(e),o!==null){for(t.flags|=128,Rn(i,!1),r=o.updateQueue,r!==null&&(t.updateQueue=r,t.flags|=4),t.subtreeFlags=0,r=n,n=t.child;n!==null;)i=n,e=r,i.flags&=14680066,o=i.alternate,o===null?(i.childLanes=0,i.lanes=e,i.child=null,i.subtreeFlags=0,i.memoizedProps=null,i.memoizedState=null,i.updateQueue=null,i.dependencies=null,i.stateNode=null):(i.childLanes=o.childLanes,i.lanes=o.lanes,i.child=o.child,i.subtreeFlags=0,i.deletions=null,i.memoizedProps=o.memoizedProps,i.memoizedState=o.memoizedState,i.updateQueue=o.updateQueue,i.type=o.type,e=o.dependencies,i.dependencies=e===null?null:{lanes:e.lanes,firstContext:e.firstContext}),n=n.sibling;return $(Y,Y.current&1|2),t.child}e=e.sibling}i.tail!==null&&q()>xn&&(t.flags|=128,r=!0,Rn(i,!1),t.lanes=4194304)}else{if(!r)if(e=al(o),e!==null){if(t.flags|=128,r=!0,n=e.updateQueue,n!==null&&(t.updateQueue=n,t.flags|=4),Rn(i,!0),i.tail===null&&i.tailMode==="hidden"&&!o.alternate&&!Q)return ce(t),null}else 2*q()-i.renderingStartTime>xn&&n!==1073741824&&(t.flags|=128,r=!0,Rn(i,!1),t.lanes=4194304);i.isBackwards?(o.sibling=t.child,t.child=o):(n=i.last,n!==null?n.sibling=o:t.child=o,i.last=o)}return i.tail!==null?(t=i.tail,i.rendering=t,i.tail=t.sibling,i.renderingStartTime=q(),t.sibling=null,n=Y.current,$(Y,r?n&1|2:n&1),t):(ce(t),null);case 22:case 23:return Ho(),r=t.memoizedState!==null,e!==null&&e.memoizedState!==null!==r&&(t.flags|=8192),r&&t.mode&1?Ce&1073741824&&(ce(t),t.subtreeFlags&6&&(t.flags|=8192)):ce(t),null;case 24:return null;case 25:return null}throw Error(S(156,t.tag))}function Ad(e,t){switch(ko(t),t.tag){case 1:return we(t.type)&&nl(),e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 3:return Sn(),V(Se),V(de),Lo(),e=t.flags,e&65536&&!(e&128)?(t.flags=e&-65537|128,t):null;case 5:return zo(t),null;case 13:if(V(Y),e=t.memoizedState,e!==null&&e.dehydrated!==null){if(t.alternate===null)throw Error(S(340));yn()}return e=t.flags,e&65536?(t.flags=e&-65537|128,t):null;case 19:return V(Y),null;case 4:return Sn(),null;case 10:return _o(t.type._context),null;case 22:case 23:return Ho(),null;case 24:return null;default:return null}}var Or=!1,fe=!1,Ud=typeof WeakSet=="function"?WeakSet:Set,N=null;function un(e,t){var n=e.ref;if(n!==null)if(typeof n=="function")try{n(null)}catch(r){X(e,t,r)}else n.current=null}function Qi(e,t,n){try{n()}catch(r){X(e,t,r)}}var Yu=!1;function $d(e,t){if(Pi=qr,e=oa(),wo(e)){if("selectionStart"in e)var n={start:e.selectionStart,end:e.selectionEnd};else e:{n=(n=e.ownerDocument)&&n.defaultView||window;var r=n.getSelection&&n.getSelection();if(r&&r.rangeCount!==0){n=r.anchorNode;var l=r.anchorOffset,i=r.focusNode;r=r.focusOffset;try{n.nodeType,i.nodeType}catch{n=null;break e}var o=0,u=-1,s=-1,d=0,y=0,h=e,m=null;t:for(;;){for(var w;h!==n||l!==0&&h.nodeType!==3||(u=o+l),h!==i||r!==0&&h.nodeType!==3||(s=o+r),h.nodeType===3&&(o+=h.nodeValue.length),(w=h.firstChild)!==null;)m=h,h=w;for(;;){if(h===e)break t;if(m===n&&++d===l&&(u=o),m===i&&++y===r&&(s=o),(w=h.nextSibling)!==null)break;h=m,m=h.parentNode}h=w}n=u===-1||s===-1?null:{start:u,end:s}}else n=null}n=n||{start:0,end:0}}else n=null;for(Ti={focusedElem:e,selectionRange:n},qr=!1,N=t;N!==null;)if(t=N,e=t.child,(t.subtreeFlags&1028)!==0&&e!==null)e.return=t,N=e;else for(;N!==null;){t=N;try{var k=t.alternate;if(t.flags&1024)switch(t.tag){case 0:case 11:case 15:break;case 1:if(k!==null){var v=k.memoizedProps,D=k.memoizedState,a=t.stateNode,c=a.getSnapshotBeforeUpdate(t.elementType===t.type?v:Fe(t.type,v),D);a.__reactInternalSnapshotBeforeUpdate=c}break;case 3:var p=t.stateNode.containerInfo;p.nodeType===1?p.textContent="":p.nodeType===9&&p.documentElement&&p.removeChild(p.documentElement);break;case 5:case 6:case 4:case 17:break;default:throw Error(S(163))}}catch(g){X(t,t.return,g)}if(e=t.sibling,e!==null){e.return=t.return,N=e;break}N=t.return}return k=Yu,Yu=!1,k}function Kn(e,t,n){var r=t.updateQueue;if(r=r!==null?r.lastEffect:null,r!==null){var l=r=r.next;do{if((l.tag&e)===e){var i=l.destroy;l.destroy=void 0,i!==void 0&&Qi(t,n,i)}l=l.next}while(l!==r)}}function _l(e,t){if(t=t.updateQueue,t=t!==null?t.lastEffect:null,t!==null){var n=t=t.next;do{if((n.tag&e)===e){var r=n.create;n.destroy=r()}n=n.next}while(n!==t)}}function Ki(e){var t=e.ref;if(t!==null){var n=e.stateNode;switch(e.tag){case 5:e=n;break;default:e=n}typeof t=="function"?t(e):t.current=e}}function nc(e){var t=e.alternate;t!==null&&(e.alternate=null,nc(t)),e.child=null,e.deletions=null,e.sibling=null,e.tag===5&&(t=e.stateNode,t!==null&&(delete t[Qe],delete t[ir],delete t[Oi],delete t[kd],delete t[Cd])),e.stateNode=null,e.return=null,e.dependencies=null,e.memoizedProps=null,e.memoizedState=null,e.pendingProps=null,e.stateNode=null,e.updateQueue=null}function rc(e){return e.tag===5||e.tag===3||e.tag===4}function Gu(e){e:for(;;){for(;e.sibling===null;){if(e.return===null||rc(e.return))return null;e=e.return}for(e.sibling.return=e.return,e=e.sibling;e.tag!==5&&e.tag!==6&&e.tag!==18;){if(e.flags&2||e.child===null||e.tag===4)continue e;e.child.return=e,e=e.child}if(!(e.flags&2))return e.stateNode}}function Yi(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.nodeType===8?n.parentNode.insertBefore(e,t):n.insertBefore(e,t):(n.nodeType===8?(t=n.parentNode,t.insertBefore(e,n)):(t=n,t.appendChild(e)),n=n._reactRootContainer,n!=null||t.onclick!==null||(t.onclick=tl));else if(r!==4&&(e=e.child,e!==null))for(Yi(e,t,n),e=e.sibling;e!==null;)Yi(e,t,n),e=e.sibling}function Gi(e,t,n){var r=e.tag;if(r===5||r===6)e=e.stateNode,t?n.insertBefore(e,t):n.appendChild(e);else if(r!==4&&(e=e.child,e!==null))for(Gi(e,t,n),e=e.sibling;e!==null;)Gi(e,t,n),e=e.sibling}var oe=null,Me=!1;function ut(e,t,n){for(n=n.child;n!==null;)lc(e,t,n),n=n.sibling}function lc(e,t,n){if(Ke&&typeof Ke.onCommitFiberUnmount=="function")try{Ke.onCommitFiberUnmount(gl,n)}catch{}switch(n.tag){case 5:fe||un(n,t);case 6:var r=oe,l=Me;oe=null,ut(e,t,n),oe=r,Me=l,oe!==null&&(Me?(e=oe,n=n.stateNode,e.nodeType===8?e.parentNode.removeChild(n):e.removeChild(n)):oe.removeChild(n.stateNode));break;case 18:oe!==null&&(Me?(e=oe,n=n.stateNode,e.nodeType===8?Jl(e.parentNode,n):e.nodeType===1&&Jl(e,n),er(e)):Jl(oe,n.stateNode));break;case 4:r=oe,l=Me,oe=n.stateNode.containerInfo,Me=!0,ut(e,t,n),oe=r,Me=l;break;case 0:case 11:case 14:case 15:if(!fe&&(r=n.updateQueue,r!==null&&(r=r.lastEffect,r!==null))){l=r=r.next;do{var i=l,o=i.destroy;i=i.tag,o!==void 0&&(i&2||i&4)&&Qi(n,t,o),l=l.next}while(l!==r)}ut(e,t,n);break;case 1:if(!fe&&(un(n,t),r=n.stateNode,typeof r.componentWillUnmount=="function"))try{r.props=n.memoizedProps,r.state=n.memoizedState,r.componentWillUnmount()}catch(u){X(n,t,u)}ut(e,t,n);break;case 21:ut(e,t,n);break;case 22:n.mode&1?(fe=(r=fe)||n.memoizedState!==null,ut(e,t,n),fe=r):ut(e,t,n);break;default:ut(e,t,n)}}function Ju(e){var t=e.updateQueue;if(t!==null){e.updateQueue=null;var n=e.stateNode;n===null&&(n=e.stateNode=new Ud),t.forEach(function(r){var l=Jd.bind(null,e,r);n.has(r)||(n.add(r),r.then(l,l))})}}function De(e,t){var n=t.deletions;if(n!==null)for(var r=0;r<n.length;r++){var l=n[r];try{var i=e,o=t,u=o;e:for(;u!==null;){switch(u.tag){case 5:oe=u.stateNode,Me=!1;break e;case 3:oe=u.stateNode.containerInfo,Me=!0;break e;case 4:oe=u.stateNode.containerInfo,Me=!0;break e}u=u.return}if(oe===null)throw Error(S(160));lc(i,o,l),oe=null,Me=!1;var s=l.alternate;s!==null&&(s.return=null),l.return=null}catch(d){X(l,t,d)}}if(t.subtreeFlags&12854)for(t=t.child;t!==null;)ic(t,e),t=t.sibling}function ic(e,t){var n=e.alternate,r=e.flags;switch(e.tag){case 0:case 11:case 14:case 15:if(De(t,e),Ve(e),r&4){try{Kn(3,e,e.return),_l(3,e)}catch(v){X(e,e.return,v)}try{Kn(5,e,e.return)}catch(v){X(e,e.return,v)}}break;case 1:De(t,e),Ve(e),r&512&&n!==null&&un(n,n.return);break;case 5:if(De(t,e),Ve(e),r&512&&n!==null&&un(n,n.return),e.flags&32){var l=e.stateNode;try{Xn(l,"")}catch(v){X(e,e.return,v)}}if(r&4&&(l=e.stateNode,l!=null)){var i=e.memoizedProps,o=n!==null?n.memoizedProps:i,u=e.type,s=e.updateQueue;if(e.updateQueue=null,s!==null)try{u==="input"&&i.type==="radio"&&i.name!=null&&js(l,i),gi(u,o);var d=gi(u,i);for(o=0;o<s.length;o+=2){var y=s[o],h=s[o+1];y==="style"?Os(l,h):y==="dangerouslySetInnerHTML"?zs(l,h):y==="children"?Xn(l,h):oo(l,y,h,d)}switch(u){case"input":pi(l,i);break;case"textarea":Ps(l,i);break;case"select":var m=l._wrapperState.wasMultiple;l._wrapperState.wasMultiple=!!i.multiple;var w=i.value;w!=null?an(l,!!i.multiple,w,!1):m!==!!i.multiple&&(i.defaultValue!=null?an(l,!!i.multiple,i.defaultValue,!0):an(l,!!i.multiple,i.multiple?[]:"",!1))}l[ir]=i}catch(v){X(e,e.return,v)}}break;case 6:if(De(t,e),Ve(e),r&4){if(e.stateNode===null)throw Error(S(162));l=e.stateNode,i=e.memoizedProps;try{l.nodeValue=i}catch(v){X(e,e.return,v)}}break;case 3:if(De(t,e),Ve(e),r&4&&n!==null&&n.memoizedState.isDehydrated)try{er(t.containerInfo)}catch(v){X(e,e.return,v)}break;case 4:De(t,e),Ve(e);break;case 13:De(t,e),Ve(e),l=e.child,l.flags&8192&&(i=l.memoizedState!==null,l.stateNode.isHidden=i,!i||l.alternate!==null&&l.alternate.memoizedState!==null||($o=q())),r&4&&Ju(e);break;case 22:if(y=n!==null&&n.memoizedState!==null,e.mode&1?(fe=(d=fe)||y,De(t,e),fe=d):De(t,e),Ve(e),r&8192){if(d=e.memoizedState!==null,(e.stateNode.isHidden=d)&&!y&&e.mode&1)for(N=e,y=e.child;y!==null;){for(h=N=y;N!==null;){switch(m=N,w=m.child,m.tag){case 0:case 11:case 14:case 15:Kn(4,m,m.return);break;case 1:un(m,m.return);var k=m.stateNode;if(typeof k.componentWillUnmount=="function"){r=m,n=m.return;try{t=r,k.props=t.memoizedProps,k.state=t.memoizedState,k.componentWillUnmount()}catch(v){X(r,n,v)}}break;case 5:un(m,m.return);break;case 22:if(m.memoizedState!==null){Zu(h);continue}}w!==null?(w.return=m,N=w):Zu(h)}y=y.sibling}e:for(y=null,h=e;;){if(h.tag===5){if(y===null){y=h;try{l=h.stateNode,d?(i=l.style,typeof i.setProperty=="function"?i.setProperty("display","none","important"):i.display="none"):(u=h.stateNode,s=h.memoizedProps.style,o=s!=null&&s.hasOwnProperty("display")?s.display:null,u.style.display=Ls("display",o))}catch(v){X(e,e.return,v)}}}else if(h.tag===6){if(y===null)try{h.stateNode.nodeValue=d?"":h.memoizedProps}catch(v){X(e,e.return,v)}}else if((h.tag!==22&&h.tag!==23||h.memoizedState===null||h===e)&&h.child!==null){h.child.return=h,h=h.child;continue}if(h===e)break e;for(;h.sibling===null;){if(h.return===null||h.return===e)break e;y===h&&(y=null),h=h.return}y===h&&(y=null),h.sibling.return=h.return,h=h.sibling}}break;case 19:De(t,e),Ve(e),r&4&&Ju(e);break;case 21:break;default:De(t,e),Ve(e)}}function Ve(e){var t=e.flags;if(t&2){try{e:{for(var n=e.return;n!==null;){if(rc(n)){var r=n;break e}n=n.return}throw Error(S(160))}switch(r.tag){case 5:var l=r.stateNode;r.flags&32&&(Xn(l,""),r.flags&=-33);var i=Gu(e);Gi(e,i,l);break;case 3:case 4:var o=r.stateNode.containerInfo,u=Gu(e);Yi(e,u,o);break;default:throw Error(S(161))}}catch(s){X(e,e.return,s)}e.flags&=-3}t&4096&&(e.flags&=-4097)}function Bd(e,t,n){N=e,oc(e)}function oc(e,t,n){for(var r=(e.mode&1)!==0;N!==null;){var l=N,i=l.child;if(l.tag===22&&r){var o=l.memoizedState!==null||Or;if(!o){var u=l.alternate,s=u!==null&&u.memoizedState!==null||fe;u=Or;var d=fe;if(Or=o,(fe=s)&&!d)for(N=l;N!==null;)o=N,s=o.child,o.tag===22&&o.memoizedState!==null?qu(l):s!==null?(s.return=o,N=s):qu(l);for(;i!==null;)N=i,oc(i),i=i.sibling;N=l,Or=u,fe=d}Xu(e)}else l.subtreeFlags&8772&&i!==null?(i.return=l,N=i):Xu(e)}}function Xu(e){for(;N!==null;){var t=N;if(t.flags&8772){var n=t.alternate;try{if(t.flags&8772)switch(t.tag){case 0:case 11:case 15:fe||_l(5,t);break;case 1:var r=t.stateNode;if(t.flags&4&&!fe)if(n===null)r.componentDidMount();else{var l=t.elementType===t.type?n.memoizedProps:Fe(t.type,n.memoizedProps);r.componentDidUpdate(l,n.memoizedState,r.__reactInternalSnapshotBeforeUpdate)}var i=t.updateQueue;i!==null&&Ru(t,i,r);break;case 3:var o=t.updateQueue;if(o!==null){if(n=null,t.child!==null)switch(t.child.tag){case 5:n=t.child.stateNode;break;case 1:n=t.child.stateNode}Ru(t,o,n)}break;case 5:var u=t.stateNode;if(n===null&&t.flags&4){n=u;var s=t.memoizedProps;switch(t.type){case"button":case"input":case"select":case"textarea":s.autoFocus&&n.focus();break;case"img":s.src&&(n.src=s.src)}}break;case 6:break;case 4:break;case 12:break;case 13:if(t.memoizedState===null){var d=t.alternate;if(d!==null){var y=d.memoizedState;if(y!==null){var h=y.dehydrated;h!==null&&er(h)}}}break;case 19:case 17:case 21:case 22:case 23:case 25:break;default:throw Error(S(163))}fe||t.flags&512&&Ki(t)}catch(m){X(t,t.return,m)}}if(t===e){N=null;break}if(n=t.sibling,n!==null){n.return=t.return,N=n;break}N=t.return}}function Zu(e){for(;N!==null;){var t=N;if(t===e){N=null;break}var n=t.sibling;if(n!==null){n.return=t.return,N=n;break}N=t.return}}function qu(e){for(;N!==null;){var t=N;try{switch(t.tag){case 0:case 11:case 15:var n=t.return;try{_l(4,t)}catch(s){X(t,n,s)}break;case 1:var r=t.stateNode;if(typeof r.componentDidMount=="function"){var l=t.return;try{r.componentDidMount()}catch(s){X(t,l,s)}}var i=t.return;try{Ki(t)}catch(s){X(t,i,s)}break;case 5:var o=t.return;try{Ki(t)}catch(s){X(t,o,s)}}}catch(s){X(t,t.return,s)}if(t===e){N=null;break}var u=t.sibling;if(u!==null){u.return=t.return,N=u;break}N=t.return}}var Hd=Math.ceil,dl=it.ReactCurrentDispatcher,Ao=it.ReactCurrentOwner,Oe=it.ReactCurrentBatchConfig,M=0,ie=null,b=null,ue=0,Ce=0,sn=Nt(0),re=0,fr=null,Bt=0,jl=0,Uo=0,Yn=null,ye=null,$o=0,xn=1/0,Xe=null,pl=!1,Ji=null,St=null,Ir=!1,pt=null,ml=0,Gn=0,Xi=null,Wr=-1,Qr=0;function me(){return M&6?q():Wr!==-1?Wr:Wr=q()}function wt(e){return e.mode&1?M&2&&ue!==0?ue&-ue:Nd.transition!==null?(Qr===0&&(Qr=Ws()),Qr):(e=A,e!==0||(e=window.event,e=e===void 0?16:Zs(e.type)),e):1}function $e(e,t,n,r){if(50<Gn)throw Gn=0,Xi=null,Error(S(185));pr(e,n,r),(!(M&2)||e!==ie)&&(e===ie&&(!(M&2)&&(jl|=n),re===4&&ft(e,ue)),xe(e,r),n===1&&M===0&&!(t.mode&1)&&(xn=q()+500,Cl&&_t()))}function xe(e,t){var n=e.callbackNode;Nf(e,t);var r=Zr(e,e===ie?ue:0);if(r===0)n!==null&&uu(n),e.callbackNode=null,e.callbackPriority=0;else if(t=r&-r,e.callbackPriority!==t){if(n!=null&&uu(n),t===1)e.tag===0?Ed(bu.bind(null,e)):va(bu.bind(null,e)),wd(function(){!(M&6)&&_t()}),n=null;else{switch(Qs(r)){case 1:n=fo;break;case 4:n=Hs;break;case 16:n=Xr;break;case 536870912:n=Vs;break;default:n=Xr}n=mc(n,uc.bind(null,e))}e.callbackPriority=t,e.callbackNode=n}}function uc(e,t){if(Wr=-1,Qr=0,M&6)throw Error(S(327));var n=e.callbackNode;if(mn()&&e.callbackNode!==n)return null;var r=Zr(e,e===ie?ue:0);if(r===0)return null;if(r&30||r&e.expiredLanes||t)t=hl(e,r);else{t=r;var l=M;M|=2;var i=ac();(ie!==e||ue!==t)&&(Xe=null,xn=q()+500,Ft(e,t));do try{Qd();break}catch(u){sc(e,u)}while(!0);No(),dl.current=i,M=l,b!==null?t=0:(ie=null,ue=0,t=re)}if(t!==0){if(t===2&&(l=Ci(e),l!==0&&(r=l,t=Zi(e,l))),t===1)throw n=fr,Ft(e,0),ft(e,r),xe(e,q()),n;if(t===6)ft(e,r);else{if(l=e.current.alternate,!(r&30)&&!Vd(l)&&(t=hl(e,r),t===2&&(i=Ci(e),i!==0&&(r=i,t=Zi(e,i))),t===1))throw n=fr,Ft(e,0),ft(e,r),xe(e,q()),n;switch(e.finishedWork=l,e.finishedLanes=r,t){case 0:case 1:throw Error(S(345));case 2:Ot(e,ye,Xe);break;case 3:if(ft(e,r),(r&130023424)===r&&(t=$o+500-q(),10<t)){if(Zr(e,0)!==0)break;if(l=e.suspendedLanes,(l&r)!==r){me(),e.pingedLanes|=e.suspendedLanes&l;break}e.timeoutHandle=Li(Ot.bind(null,e,ye,Xe),t);break}Ot(e,ye,Xe);break;case 4:if(ft(e,r),(r&4194240)===r)break;for(t=e.eventTimes,l=-1;0<r;){var o=31-Ue(r);i=1<<o,o=t[o],o>l&&(l=o),r&=~i}if(r=l,r=q()-r,r=(120>r?120:480>r?480:1080>r?1080:1920>r?1920:3e3>r?3e3:4320>r?4320:1960*Hd(r/1960))-r,10<r){e.timeoutHandle=Li(Ot.bind(null,e,ye,Xe),r);break}Ot(e,ye,Xe);break;case 5:Ot(e,ye,Xe);break;default:throw Error(S(329))}}}return xe(e,q()),e.callbackNode===n?uc.bind(null,e):null}function Zi(e,t){var n=Yn;return e.current.memoizedState.isDehydrated&&(Ft(e,t).flags|=256),e=hl(e,t),e!==2&&(t=ye,ye=n,t!==null&&qi(t)),e}function qi(e){ye===null?ye=e:ye.push.apply(ye,e)}function Vd(e){for(var t=e;;){if(t.flags&16384){var n=t.updateQueue;if(n!==null&&(n=n.stores,n!==null))for(var r=0;r<n.length;r++){var l=n[r],i=l.getSnapshot;l=l.value;try{if(!Be(i(),l))return!1}catch{return!1}}}if(n=t.child,t.subtreeFlags&16384&&n!==null)n.return=t,t=n;else{if(t===e)break;for(;t.sibling===null;){if(t.return===null||t.return===e)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}function ft(e,t){for(t&=~Uo,t&=~jl,e.suspendedLanes|=t,e.pingedLanes&=~t,e=e.expirationTimes;0<t;){var n=31-Ue(t),r=1<<n;e[n]=-1,t&=~r}}function bu(e){if(M&6)throw Error(S(327));mn();var t=Zr(e,0);if(!(t&1))return xe(e,q()),null;var n=hl(e,t);if(e.tag!==0&&n===2){var r=Ci(e);r!==0&&(t=r,n=Zi(e,r))}if(n===1)throw n=fr,Ft(e,0),ft(e,t),xe(e,q()),n;if(n===6)throw Error(S(345));return e.finishedWork=e.current.alternate,e.finishedLanes=t,Ot(e,ye,Xe),xe(e,q()),null}function Bo(e,t){var n=M;M|=1;try{return e(t)}finally{M=n,M===0&&(xn=q()+500,Cl&&_t())}}function Ht(e){pt!==null&&pt.tag===0&&!(M&6)&&mn();var t=M;M|=1;var n=Oe.transition,r=A;try{if(Oe.transition=null,A=1,e)return e()}finally{A=r,Oe.transition=n,M=t,!(M&6)&&_t()}}function Ho(){Ce=sn.current,V(sn)}function Ft(e,t){e.finishedWork=null,e.finishedLanes=0;var n=e.timeoutHandle;if(n!==-1&&(e.timeoutHandle=-1,Sd(n)),b!==null)for(n=b.return;n!==null;){var r=n;switch(ko(r),r.tag){case 1:r=r.type.childContextTypes,r!=null&&nl();break;case 3:Sn(),V(Se),V(de),Lo();break;case 5:zo(r);break;case 4:Sn();break;case 13:V(Y);break;case 19:V(Y);break;case 10:_o(r.type._context);break;case 22:case 23:Ho()}n=n.return}if(ie=e,b=e=xt(e.current,null),ue=Ce=t,re=0,fr=null,Uo=jl=Bt=0,ye=Yn=null,Rt!==null){for(t=0;t<Rt.length;t++)if(n=Rt[t],r=n.interleaved,r!==null){n.interleaved=null;var l=r.next,i=n.pending;if(i!==null){var o=i.next;i.next=l,r.next=o}n.pending=r}Rt=null}return e}function sc(e,t){do{var n=b;try{if(No(),Br.current=fl,cl){for(var r=G.memoizedState;r!==null;){var l=r.queue;l!==null&&(l.pending=null),r=r.next}cl=!1}if($t=0,le=ne=G=null,Qn=!1,sr=0,Ao.current=null,n===null||n.return===null){re=1,fr=t,b=null;break}e:{var i=e,o=n.return,u=n,s=t;if(t=ue,u.flags|=32768,s!==null&&typeof s=="object"&&typeof s.then=="function"){var d=s,y=u,h=y.tag;if(!(y.mode&1)&&(h===0||h===11||h===15)){var m=y.alternate;m?(y.updateQueue=m.updateQueue,y.memoizedState=m.memoizedState,y.lanes=m.lanes):(y.updateQueue=null,y.memoizedState=null)}var w=$u(o);if(w!==null){w.flags&=-257,Bu(w,o,u,i,t),w.mode&1&&Uu(i,d,t),t=w,s=d;var k=t.updateQueue;if(k===null){var v=new Set;v.add(s),t.updateQueue=v}else k.add(s);break e}else{if(!(t&1)){Uu(i,d,t),Vo();break e}s=Error(S(426))}}else if(Q&&u.mode&1){var D=$u(o);if(D!==null){!(D.flags&65536)&&(D.flags|=256),Bu(D,o,u,i,t),Co(wn(s,u));break e}}i=s=wn(s,u),re!==4&&(re=2),Yn===null?Yn=[i]:Yn.push(i),i=o;do{switch(i.tag){case 3:i.flags|=65536,t&=-t,i.lanes|=t;var a=Qa(i,s,t);Iu(i,a);break e;case 1:u=s;var c=i.type,p=i.stateNode;if(!(i.flags&128)&&(typeof c.getDerivedStateFromError=="function"||p!==null&&typeof p.componentDidCatch=="function"&&(St===null||!St.has(p)))){i.flags|=65536,t&=-t,i.lanes|=t;var g=Ka(i,u,t);Iu(i,g);break e}}i=i.return}while(i!==null)}fc(n)}catch(E){t=E,b===n&&n!==null&&(b=n=n.return);continue}break}while(!0)}function ac(){var e=dl.current;return dl.current=fl,e===null?fl:e}function Vo(){(re===0||re===3||re===2)&&(re=4),ie===null||!(Bt&268435455)&&!(jl&268435455)||ft(ie,ue)}function hl(e,t){var n=M;M|=2;var r=ac();(ie!==e||ue!==t)&&(Xe=null,Ft(e,t));do try{Wd();break}catch(l){sc(e,l)}while(!0);if(No(),M=n,dl.current=r,b!==null)throw Error(S(261));return ie=null,ue=0,re}function Wd(){for(;b!==null;)cc(b)}function Qd(){for(;b!==null&&!vf();)cc(b)}function cc(e){var t=pc(e.alternate,e,Ce);e.memoizedProps=e.pendingProps,t===null?fc(e):b=t,Ao.current=null}function fc(e){var t=e;do{var n=t.alternate;if(e=t.return,t.flags&32768){if(n=Ad(n,t),n!==null){n.flags&=32767,b=n;return}if(e!==null)e.flags|=32768,e.subtreeFlags=0,e.deletions=null;else{re=6,b=null;return}}else if(n=Md(n,t,Ce),n!==null){b=n;return}if(t=t.sibling,t!==null){b=t;return}b=t=e}while(t!==null);re===0&&(re=5)}function Ot(e,t,n){var r=A,l=Oe.transition;try{Oe.transition=null,A=1,Kd(e,t,n,r)}finally{Oe.transition=l,A=r}return null}function Kd(e,t,n,r){do mn();while(pt!==null);if(M&6)throw Error(S(327));n=e.finishedWork;var l=e.finishedLanes;if(n===null)return null;if(e.finishedWork=null,e.finishedLanes=0,n===e.current)throw Error(S(177));e.callbackNode=null,e.callbackPriority=0;var i=n.lanes|n.childLanes;if(_f(e,i),e===ie&&(b=ie=null,ue=0),!(n.subtreeFlags&2064)&&!(n.flags&2064)||Ir||(Ir=!0,mc(Xr,function(){return mn(),null})),i=(n.flags&15990)!==0,n.subtreeFlags&15990||i){i=Oe.transition,Oe.transition=null;var o=A;A=1;var u=M;M|=4,Ao.current=null,$d(e,n),ic(n,e),dd(Ti),qr=!!Pi,Ti=Pi=null,e.current=n,Bd(n),yf(),M=u,A=o,Oe.transition=i}else e.current=n;if(Ir&&(Ir=!1,pt=e,ml=l),i=e.pendingLanes,i===0&&(St=null),wf(n.stateNode),xe(e,q()),t!==null)for(r=e.onRecoverableError,n=0;n<t.length;n++)l=t[n],r(l.value,{componentStack:l.stack,digest:l.digest});if(pl)throw pl=!1,e=Ji,Ji=null,e;return ml&1&&e.tag!==0&&mn(),i=e.pendingLanes,i&1?e===Xi?Gn++:(Gn=0,Xi=e):Gn=0,_t(),null}function mn(){if(pt!==null){var e=Qs(ml),t=Oe.transition,n=A;try{if(Oe.transition=null,A=16>e?16:e,pt===null)var r=!1;else{if(e=pt,pt=null,ml=0,M&6)throw Error(S(331));var l=M;for(M|=4,N=e.current;N!==null;){var i=N,o=i.child;if(N.flags&16){var u=i.deletions;if(u!==null){for(var s=0;s<u.length;s++){var d=u[s];for(N=d;N!==null;){var y=N;switch(y.tag){case 0:case 11:case 15:Kn(8,y,i)}var h=y.child;if(h!==null)h.return=y,N=h;else for(;N!==null;){y=N;var m=y.sibling,w=y.return;if(nc(y),y===d){N=null;break}if(m!==null){m.return=w,N=m;break}N=w}}}var k=i.alternate;if(k!==null){var v=k.child;if(v!==null){k.child=null;do{var D=v.sibling;v.sibling=null,v=D}while(v!==null)}}N=i}}if(i.subtreeFlags&2064&&o!==null)o.return=i,N=o;else e:for(;N!==null;){if(i=N,i.flags&2048)switch(i.tag){case 0:case 11:case 15:Kn(9,i,i.return)}var a=i.sibling;if(a!==null){a.return=i.return,N=a;break e}N=i.return}}var c=e.current;for(N=c;N!==null;){o=N;var p=o.child;if(o.subtreeFlags&2064&&p!==null)p.return=o,N=p;else e:for(o=c;N!==null;){if(u=N,u.flags&2048)try{switch(u.tag){case 0:case 11:case 15:_l(9,u)}}catch(E){X(u,u.return,E)}if(u===o){N=null;break e}var g=u.sibling;if(g!==null){g.return=u.return,N=g;break e}N=u.return}}if(M=l,_t(),Ke&&typeof Ke.onPostCommitFiberRoot=="function")try{Ke.onPostCommitFiberRoot(gl,e)}catch{}r=!0}return r}finally{A=n,Oe.transition=t}}return!1}function es(e,t,n){t=wn(n,t),t=Qa(e,t,1),e=gt(e,t,1),t=me(),e!==null&&(pr(e,1,t),xe(e,t))}function X(e,t,n){if(e.tag===3)es(e,e,n);else for(;t!==null;){if(t.tag===3){es(t,e,n);break}else if(t.tag===1){var r=t.stateNode;if(typeof t.type.getDerivedStateFromError=="function"||typeof r.componentDidCatch=="function"&&(St===null||!St.has(r))){e=wn(n,e),e=Ka(t,e,1),t=gt(t,e,1),e=me(),t!==null&&(pr(t,1,e),xe(t,e));break}}t=t.return}}function Yd(e,t,n){var r=e.pingCache;r!==null&&r.delete(t),t=me(),e.pingedLanes|=e.suspendedLanes&n,ie===e&&(ue&n)===n&&(re===4||re===3&&(ue&130023424)===ue&&500>q()-$o?Ft(e,0):Uo|=n),xe(e,t)}function dc(e,t){t===0&&(e.mode&1?(t=Cr,Cr<<=1,!(Cr&130023424)&&(Cr=4194304)):t=1);var n=me();e=rt(e,t),e!==null&&(pr(e,t,n),xe(e,n))}function Gd(e){var t=e.memoizedState,n=0;t!==null&&(n=t.retryLane),dc(e,n)}function Jd(e,t){var n=0;switch(e.tag){case 13:var r=e.stateNode,l=e.memoizedState;l!==null&&(n=l.retryLane);break;case 19:r=e.stateNode;break;default:throw Error(S(314))}r!==null&&r.delete(t),dc(e,n)}var pc;pc=function(e,t,n){if(e!==null)if(e.memoizedProps!==t.pendingProps||Se.current)ge=!0;else{if(!(e.lanes&n)&&!(t.flags&128))return ge=!1,Fd(e,t,n);ge=!!(e.flags&131072)}else ge=!1,Q&&t.flags&1048576&&ya(t,il,t.index);switch(t.lanes=0,t.tag){case 2:var r=t.type;Vr(e,t),e=t.pendingProps;var l=vn(t,de.current);pn(t,n),l=Io(null,t,r,e,l,n);var i=Ro();return t.flags|=1,typeof l=="object"&&l!==null&&typeof l.render=="function"&&l.$$typeof===void 0?(t.tag=1,t.memoizedState=null,t.updateQueue=null,we(r)?(i=!0,rl(t)):i=!1,t.memoizedState=l.state!==null&&l.state!==void 0?l.state:null,Po(t),l.updater=Nl,t.stateNode=l,l._reactInternals=t,Ai(t,r,e,n),t=Bi(null,t,r,!0,i,n)):(t.tag=0,Q&&i&&xo(t),pe(null,t,l,n),t=t.child),t;case 16:r=t.elementType;e:{switch(Vr(e,t),e=t.pendingProps,l=r._init,r=l(r._payload),t.type=r,l=t.tag=Zd(r),e=Fe(r,e),l){case 0:t=$i(null,t,r,e,n);break e;case 1:t=Wu(null,t,r,e,n);break e;case 11:t=Hu(null,t,r,e,n);break e;case 14:t=Vu(null,t,r,Fe(r.type,e),n);break e}throw Error(S(306,r,""))}return t;case 0:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Fe(r,l),$i(e,t,r,l,n);case 1:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Fe(r,l),Wu(e,t,r,l,n);case 3:e:{if(Xa(t),e===null)throw Error(S(387));r=t.pendingProps,i=t.memoizedState,l=i.element,Ca(e,t),sl(t,r,null,n);var o=t.memoizedState;if(r=o.element,i.isDehydrated)if(i={element:r,isDehydrated:!1,cache:o.cache,pendingSuspenseBoundaries:o.pendingSuspenseBoundaries,transitions:o.transitions},t.updateQueue.baseState=i,t.memoizedState=i,t.flags&256){l=wn(Error(S(423)),t),t=Qu(e,t,r,n,l);break e}else if(r!==l){l=wn(Error(S(424)),t),t=Qu(e,t,r,n,l);break e}else for(Ee=yt(t.stateNode.containerInfo.firstChild),Ne=t,Q=!0,Ae=null,n=xa(t,null,r,n),t.child=n;n;)n.flags=n.flags&-3|4096,n=n.sibling;else{if(yn(),r===l){t=lt(e,t,n);break e}pe(e,t,r,n)}t=t.child}return t;case 5:return Ea(t),e===null&&Di(t),r=t.type,l=t.pendingProps,i=e!==null?e.memoizedProps:null,o=l.children,zi(r,l)?o=null:i!==null&&zi(r,i)&&(t.flags|=32),Ja(e,t),pe(e,t,o,n),t.child;case 6:return e===null&&Di(t),null;case 13:return Za(e,t,n);case 4:return To(t,t.stateNode.containerInfo),r=t.pendingProps,e===null?t.child=gn(t,null,r,n):pe(e,t,r,n),t.child;case 11:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Fe(r,l),Hu(e,t,r,l,n);case 7:return pe(e,t,t.pendingProps,n),t.child;case 8:return pe(e,t,t.pendingProps.children,n),t.child;case 12:return pe(e,t,t.pendingProps.children,n),t.child;case 10:e:{if(r=t.type._context,l=t.pendingProps,i=t.memoizedProps,o=l.value,$(ol,r._currentValue),r._currentValue=o,i!==null)if(Be(i.value,o)){if(i.children===l.children&&!Se.current){t=lt(e,t,n);break e}}else for(i=t.child,i!==null&&(i.return=t);i!==null;){var u=i.dependencies;if(u!==null){o=i.child;for(var s=u.firstContext;s!==null;){if(s.context===r){if(i.tag===1){s=et(-1,n&-n),s.tag=2;var d=i.updateQueue;if(d!==null){d=d.shared;var y=d.pending;y===null?s.next=s:(s.next=y.next,y.next=s),d.pending=s}}i.lanes|=n,s=i.alternate,s!==null&&(s.lanes|=n),Fi(i.return,n,t),u.lanes|=n;break}s=s.next}}else if(i.tag===10)o=i.type===t.type?null:i.child;else if(i.tag===18){if(o=i.return,o===null)throw Error(S(341));o.lanes|=n,u=o.alternate,u!==null&&(u.lanes|=n),Fi(o,n,t),o=i.sibling}else o=i.child;if(o!==null)o.return=i;else for(o=i;o!==null;){if(o===t){o=null;break}if(i=o.sibling,i!==null){i.return=o.return,o=i;break}o=o.return}i=o}pe(e,t,l.children,n),t=t.child}return t;case 9:return l=t.type,r=t.pendingProps.children,pn(t,n),l=Ie(l),r=r(l),t.flags|=1,pe(e,t,r,n),t.child;case 14:return r=t.type,l=Fe(r,t.pendingProps),l=Fe(r.type,l),Vu(e,t,r,l,n);case 15:return Ya(e,t,t.type,t.pendingProps,n);case 17:return r=t.type,l=t.pendingProps,l=t.elementType===r?l:Fe(r,l),Vr(e,t),t.tag=1,we(r)?(e=!0,rl(t)):e=!1,pn(t,n),Wa(t,r,l),Ai(t,r,l,n),Bi(null,t,r,!0,e,n);case 19:return qa(e,t,n);case 22:return Ga(e,t,n)}throw Error(S(156,t.tag))};function mc(e,t){return Bs(e,t)}function Xd(e,t,n,r){this.tag=e,this.key=n,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=r,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function Le(e,t,n,r){return new Xd(e,t,n,r)}function Wo(e){return e=e.prototype,!(!e||!e.isReactComponent)}function Zd(e){if(typeof e=="function")return Wo(e)?1:0;if(e!=null){if(e=e.$$typeof,e===so)return 11;if(e===ao)return 14}return 2}function xt(e,t){var n=e.alternate;return n===null?(n=Le(e.tag,t,e.key,e.mode),n.elementType=e.elementType,n.type=e.type,n.stateNode=e.stateNode,n.alternate=e,e.alternate=n):(n.pendingProps=t,n.type=e.type,n.flags=0,n.subtreeFlags=0,n.deletions=null),n.flags=e.flags&14680064,n.childLanes=e.childLanes,n.lanes=e.lanes,n.child=e.child,n.memoizedProps=e.memoizedProps,n.memoizedState=e.memoizedState,n.updateQueue=e.updateQueue,t=e.dependencies,n.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext},n.sibling=e.sibling,n.index=e.index,n.ref=e.ref,n}function Kr(e,t,n,r,l,i){var o=2;if(r=e,typeof e=="function")Wo(e)&&(o=1);else if(typeof e=="string")o=5;else e:switch(e){case Zt:return Mt(n.children,l,i,t);case uo:o=8,l|=8;break;case si:return e=Le(12,n,t,l|2),e.elementType=si,e.lanes=i,e;case ai:return e=Le(13,n,t,l),e.elementType=ai,e.lanes=i,e;case ci:return e=Le(19,n,t,l),e.elementType=ci,e.lanes=i,e;case Es:return Pl(n,l,i,t);default:if(typeof e=="object"&&e!==null)switch(e.$$typeof){case ks:o=10;break e;case Cs:o=9;break e;case so:o=11;break e;case ao:o=14;break e;case st:o=16,r=null;break e}throw Error(S(130,e==null?e:typeof e,""))}return t=Le(o,n,t,l),t.elementType=e,t.type=r,t.lanes=i,t}function Mt(e,t,n,r){return e=Le(7,e,r,t),e.lanes=n,e}function Pl(e,t,n,r){return e=Le(22,e,r,t),e.elementType=Es,e.lanes=n,e.stateNode={isHidden:!1},e}function ri(e,t,n){return e=Le(6,e,null,t),e.lanes=n,e}function li(e,t,n){return t=Le(4,e.children!==null?e.children:[],e.key,t),t.lanes=n,t.stateNode={containerInfo:e.containerInfo,pendingChildren:null,implementation:e.implementation},t}function qd(e,t,n,r,l){this.tag=t,this.containerInfo=e,this.finishedWork=this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.pendingContext=this.context=null,this.callbackPriority=0,this.eventTimes=Al(0),this.expirationTimes=Al(-1),this.entangledLanes=this.finishedLanes=this.mutableReadLanes=this.expiredLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=Al(0),this.identifierPrefix=r,this.onRecoverableError=l,this.mutableSourceEagerHydrationData=null}function Qo(e,t,n,r,l,i,o,u,s){return e=new qd(e,t,n,u,s),t===1?(t=1,i===!0&&(t|=8)):t=0,i=Le(3,null,null,t),e.current=i,i.stateNode=e,i.memoizedState={element:r,isDehydrated:n,cache:null,transitions:null,pendingSuspenseBoundaries:null},Po(i),e}function bd(e,t,n){var r=3<arguments.length&&arguments[3]!==void 0?arguments[3]:null;return{$$typeof:Xt,key:r==null?null:""+r,children:e,containerInfo:t,implementation:n}}function hc(e){if(!e)return Ct;e=e._reactInternals;e:{if(Wt(e)!==e||e.tag!==1)throw Error(S(170));var t=e;do{switch(t.tag){case 3:t=t.stateNode.context;break e;case 1:if(we(t.type)){t=t.stateNode.__reactInternalMemoizedMergedChildContext;break e}}t=t.return}while(t!==null);throw Error(S(171))}if(e.tag===1){var n=e.type;if(we(n))return ha(e,n,t)}return t}function vc(e,t,n,r,l,i,o,u,s){return e=Qo(n,r,!0,e,l,i,o,u,s),e.context=hc(null),n=e.current,r=me(),l=wt(n),i=et(r,l),i.callback=t??null,gt(n,i,l),e.current.lanes=l,pr(e,l,r),xe(e,r),e}function Tl(e,t,n,r){var l=t.current,i=me(),o=wt(l);return n=hc(n),t.context===null?t.context=n:t.pendingContext=n,t=et(i,o),t.payload={element:e},r=r===void 0?null:r,r!==null&&(t.callback=r),e=gt(l,t,o),e!==null&&($e(e,l,o,i),$r(e,l,o)),o}function vl(e){if(e=e.current,!e.child)return null;switch(e.child.tag){case 5:return e.child.stateNode;default:return e.child.stateNode}}function ts(e,t){if(e=e.memoizedState,e!==null&&e.dehydrated!==null){var n=e.retryLane;e.retryLane=n!==0&&n<t?n:t}}function Ko(e,t){ts(e,t),(e=e.alternate)&&ts(e,t)}function ep(){return null}var yc=typeof reportError=="function"?reportError:function(e){console.error(e)};function Yo(e){this._internalRoot=e}zl.prototype.render=Yo.prototype.render=function(e){var t=this._internalRoot;if(t===null)throw Error(S(409));Tl(e,t,null,null)};zl.prototype.unmount=Yo.prototype.unmount=function(){var e=this._internalRoot;if(e!==null){this._internalRoot=null;var t=e.containerInfo;Ht(function(){Tl(null,e,null,null)}),t[nt]=null}};function zl(e){this._internalRoot=e}zl.prototype.unstable_scheduleHydration=function(e){if(e){var t=Gs();e={blockedOn:null,target:e,priority:t};for(var n=0;n<ct.length&&t!==0&&t<ct[n].priority;n++);ct.splice(n,0,e),n===0&&Xs(e)}};function Go(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11)}function Ll(e){return!(!e||e.nodeType!==1&&e.nodeType!==9&&e.nodeType!==11&&(e.nodeType!==8||e.nodeValue!==" react-mount-point-unstable "))}function ns(){}function tp(e,t,n,r,l){if(l){if(typeof r=="function"){var i=r;r=function(){var d=vl(o);i.call(d)}}var o=vc(t,r,e,0,null,!1,!1,"",ns);return e._reactRootContainer=o,e[nt]=o.current,rr(e.nodeType===8?e.parentNode:e),Ht(),o}for(;l=e.lastChild;)e.removeChild(l);if(typeof r=="function"){var u=r;r=function(){var d=vl(s);u.call(d)}}var s=Qo(e,0,!1,null,null,!1,!1,"",ns);return e._reactRootContainer=s,e[nt]=s.current,rr(e.nodeType===8?e.parentNode:e),Ht(function(){Tl(t,s,n,r)}),s}function Ol(e,t,n,r,l){var i=n._reactRootContainer;if(i){var o=i;if(typeof l=="function"){var u=l;l=function(){var s=vl(o);u.call(s)}}Tl(t,o,e,l)}else o=tp(n,t,e,l,r);return vl(o)}Ks=function(e){switch(e.tag){case 3:var t=e.stateNode;if(t.current.memoizedState.isDehydrated){var n=An(t.pendingLanes);n!==0&&(po(t,n|1),xe(t,q()),!(M&6)&&(xn=q()+500,_t()))}break;case 13:Ht(function(){var r=rt(e,1);if(r!==null){var l=me();$e(r,e,1,l)}}),Ko(e,1)}};mo=function(e){if(e.tag===13){var t=rt(e,134217728);if(t!==null){var n=me();$e(t,e,134217728,n)}Ko(e,134217728)}};Ys=function(e){if(e.tag===13){var t=wt(e),n=rt(e,t);if(n!==null){var r=me();$e(n,e,t,r)}Ko(e,t)}};Gs=function(){return A};Js=function(e,t){var n=A;try{return A=e,t()}finally{A=n}};wi=function(e,t,n){switch(t){case"input":if(pi(e,n),t=n.name,n.type==="radio"&&t!=null){for(n=e;n.parentNode;)n=n.parentNode;for(n=n.querySelectorAll("input[name="+JSON.stringify(""+t)+'][type="radio"]'),t=0;t<n.length;t++){var r=n[t];if(r!==e&&r.form===e.form){var l=kl(r);if(!l)throw Error(S(90));_s(r),pi(r,l)}}}break;case"textarea":Ps(e,n);break;case"select":t=n.value,t!=null&&an(e,!!n.multiple,t,!1)}};Ds=Bo;Fs=Ht;var np={usingClientEntryPoint:!1,Events:[hr,tn,kl,Is,Rs,Bo]},Dn={findFiberByHostInstance:It,bundleType:0,version:"18.3.1",rendererPackageName:"react-dom"},rp={bundleType:Dn.bundleType,version:Dn.version,rendererPackageName:Dn.rendererPackageName,rendererConfig:Dn.rendererConfig,overrideHookState:null,overrideHookStateDeletePath:null,overrideHookStateRenamePath:null,overrideProps:null,overridePropsDeletePath:null,overridePropsRenamePath:null,setErrorHandler:null,setSuspenseHandler:null,scheduleUpdate:null,currentDispatcherRef:it.ReactCurrentDispatcher,findHostInstanceByFiber:function(e){return e=Us(e),e===null?null:e.stateNode},findFiberByHostInstance:Dn.findFiberByHostInstance||ep,findHostInstancesForRefresh:null,scheduleRefresh:null,scheduleRoot:null,setRefreshHandler:null,getCurrentFiber:null,reconcilerVersion:"18.3.1-next-f1338f8080-20240426"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var Rr=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!Rr.isDisabled&&Rr.supportsFiber)try{gl=Rr.inject(rp),Ke=Rr}catch{}}je.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=np;je.createPortal=function(e,t){var n=2<arguments.length&&arguments[2]!==void 0?arguments[2]:null;if(!Go(t))throw Error(S(200));return bd(e,t,null,n)};je.createRoot=function(e,t){if(!Go(e))throw Error(S(299));var n=!1,r="",l=yc;return t!=null&&(t.unstable_strictMode===!0&&(n=!0),t.identifierPrefix!==void 0&&(r=t.identifierPrefix),t.onRecoverableError!==void 0&&(l=t.onRecoverableError)),t=Qo(e,1,!1,null,null,n,!1,r,l),e[nt]=t.current,rr(e.nodeType===8?e.parentNode:e),new Yo(t)};je.findDOMNode=function(e){if(e==null)return null;if(e.nodeType===1)return e;var t=e._reactInternals;if(t===void 0)throw typeof e.render=="function"?Error(S(188)):(e=Object.keys(e).join(","),Error(S(268,e)));return e=Us(t),e=e===null?null:e.stateNode,e};je.flushSync=function(e){return Ht(e)};je.hydrate=function(e,t,n){if(!Ll(t))throw Error(S(200));return Ol(null,e,t,!0,n)};je.hydrateRoot=function(e,t,n){if(!Go(e))throw Error(S(405));var r=n!=null&&n.hydratedSources||null,l=!1,i="",o=yc;if(n!=null&&(n.unstable_strictMode===!0&&(l=!0),n.identifierPrefix!==void 0&&(i=n.identifierPrefix),n.onRecoverableError!==void 0&&(o=n.onRecoverableError)),t=vc(t,null,e,1,n??null,l,!1,i,o),e[nt]=t.current,rr(e),r)for(e=0;e<r.length;e++)n=r[e],l=n._getVersion,l=l(n._source),t.mutableSourceEagerHydrationData==null?t.mutableSourceEagerHydrationData=[n,l]:t.mutableSourceEagerHydrationData.push(n,l);return new zl(t)};je.render=function(e,t,n){if(!Ll(t))throw Error(S(200));return Ol(null,e,t,!1,n)};je.unmountComponentAtNode=function(e){if(!Ll(e))throw Error(S(40));return e._reactRootContainer?(Ht(function(){Ol(null,null,e,!1,function(){e._reactRootContainer=null,e[nt]=null})}),!0):!1};je.unstable_batchedUpdates=Bo;je.unstable_renderSubtreeIntoContainer=function(e,t,n,r){if(!Ll(n))throw Error(S(200));if(e==null||e._reactInternals===void 0)throw Error(S(38));return Ol(e,t,n,!1,r)};je.version="18.3.1-next-f1338f8080-20240426";function gc(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(gc)}catch(e){console.error(e)}}gc(),gs.exports=je;var lp=gs.exports,Sc,rs=lp;Sc=rs.createRoot,rs.hydrateRoot;const wc="contactscout_state",xc="contactscout_jurisdiction",bi="cs_api_key",kc=[{id:"us-congress",label:"US Congress",desc:"All current US reps + senators for your state"},{id:"state-exec",label:"State Executive",desc:"Governor, Lt. Gov, AG, Treasurer, Auditor, commissioners"},{id:"state-senate",label:"State Senate",desc:"Full current state senate roster"},{id:"state-house",label:"State House",desc:"House members for your tracked counties"},{id:"city-1",label:"City Council — City 1",desc:"Current mayor and all council members"},{id:"city-2",label:"City Council — City 2",desc:"Current mayor and all council members"},{id:"city-3",label:"City Council — City 3",desc:"Current mayor and all council members"}],ip=`Verify this elected official for 2025-2026 using web search. Respond ONLY with valid JSON no markdown:
-{"stillInOffice":true,"currentTitle":"","directEmail":"","officeEmail":"","officePhone":"","changes":"No changes detected","flags":"","replacedBy":"","confidence":"high"}`,op=`Find all current elected officials for 2025-2026 using web search. Respond ONLY with valid JSON no markdown:
-{"officials":[{"name":"","title":"","district":"","county":"","category":"US Senate|US House|Executive|State Senate|House|City Council","directEmail":"","officeEmail":"","officePhone":""}],"confidence":"high","notes":""}`,up=["All","US Senate","US House","Executive","State Senate","House","City Council"],sp={pending:"PENDING",checking:"CHECKING",done:"VERIFIED",changed:"CHANGED",left_office:"LEFT OFFICE",error:"ERROR"},ls={pending:"#8b949e",checking:"#f59e0b",done:"#3fb950",changed:"#58a6ff",left_office:"#f85149",error:"#e3b341"},is={officials:[],newOfficials:[],scanStatus:{},scanMeta:{}},os={state:"",counties:"",city1:"",city2:"",city3:""};function ap(){try{const e=localStorage.getItem(wc);return e?JSON.parse(e):{...is}}catch{return{...is}}}function cp(e){try{localStorage.setItem(wc,JSON.stringify(e))}catch{}}function fp(){try{const e=localStorage.getItem(xc);return e?JSON.parse(e):{...os}}catch{return{...os}}}function dp(e){try{localStorage.setItem(xc,JSON.stringify(e))}catch{}}const pp=e=>new Promise(t=>setTimeout(t,e));function mp(e){const t=e.name.trim().split(/\s+/);return{firstName:t[0]??"",lastName:t.slice(1).join(" "),title:e.title,category:e.category,email:e.officeEmail||e.directEmail||"",rsvpLink:"",inviteStatus:"pending",sentAt:"",rsvpStatus:"No Response",rsvpDate:"",notes:[e.district,e.county].filter(Boolean).join(" · ")}}function hp(e){const t=e.state.trim()||"[YOUR STATE]",n=e.counties.trim()||"[YOUR COUNTIES]",r=e.city1.trim()||"[CITY 1]",l=e.city2.trim()||"[CITY 2]",i=e.city3.trim()||"[CITY 3]";return{"us-congress":`Search for every current US Congress member for ${t} (119th Congress 2025-2026): all House reps + 2 senators. For each: full name, title, district, official scheduler/contact email.`,"state-exec":`Search for all current ${t} statewide executive elected officials as of 2025: Governor, Lt. Governor, AG, Treasurer, Auditor, Superintendent of Public Instruction, and other commissioners. Full name, title, email.`,"state-senate":`Search for all current ${t} State Senate members as of 2025. Full name, district, official email.`,"state-house":`Search for all current ${t} House members for ${n} as of 2025. Full name, district, county, official email.`,"city-1":`Search for current ${r} City Council as of 2025. Mayor + every council member: full name, title, official email.`,"city-2":`Search for current ${l} City Council as of 2025. Mayor + every council member: full name, title, official email.`,"city-3":`Search for current ${i} City Council as of 2025. Mayor + every council member: full name, title, official email.`}}function ii(e,t){const n=document.createElement("a");n.href=URL.createObjectURL(e),n.download=t,n.click(),URL.revokeObjectURL(n.href)}function oi(){return new Date().toISOString().slice(0,10)}async function vp(e,t,n){var o,u;const r=await fetch("https://api.anthropic.com/v1/messages",{method:"POST",headers:{"Content-Type":"application/json","x-api-key":e,"anthropic-version":"2023-06-01","anthropic-dangerous-direct-browser-access":"true"},body:JSON.stringify({model:"claude-sonnet-4-20250514",max_tokens:1500,system:t,tools:[{type:"web_search_20250305",name:"web_search"}],messages:[{role:"user",content:n}]})});if(r.status===401)throw sessionStorage.removeItem(bi),new Error("Invalid API key");const l=await r.json();if(l.error)throw new Error(l.error.message??"API error");const i=((u=(o=l.content)==null?void 0:o.find(s=>s.type==="text"))==null?void 0:u.text)??"";if(!i)throw new Error("No text response");return JSON.parse(i.replace(/```json|```/g,"").trim())}function yp({apiKey:e,onSave:t,onClose:n}){const[r,l]=U.useState(e),[i,o]=U.useState(!1);function u(){if(!r.startsWith("sk-ant-")){o(!0);return}t(r),n()}return f.jsx("div",{className:"if-modal-backdrop",onClick:n,children:f.jsxs("div",{className:"if-modal",onClick:s=>s.stopPropagation(),children:[f.jsx("div",{className:"if-modal-title",children:"Claude API Key"}),f.jsxs("div",{className:"if-modal-sub",children:["Enter your Anthropic API key. Get a free key at"," ",f.jsx("a",{href:"https://console.anthropic.com/",target:"_blank",rel:"noreferrer",style:{color:"#58a6ff"},children:"console.anthropic.com"})," ","→ API Keys → Create Key. Starts with ",f.jsx("code",{children:"sk-ant-"}),". Stored in session only."]}),f.jsx("input",{className:`if-input${i?" err":""}`,type:"password",placeholder:"sk-ant-...",value:r,onChange:s=>{l(s.target.value),o(!1)},onKeyDown:s=>s.key==="Enter"&&u(),autoFocus:!0}),i&&f.jsx("div",{style:{fontSize:10,color:"#f85149",marginTop:6},children:"Must start with sk-ant-"}),f.jsxs("div",{style:{display:"flex",gap:8,justifyContent:"flex-end",marginTop:12},children:[f.jsx("button",{className:"if-btn",onClick:n,children:"Cancel"}),f.jsx("button",{className:"if-btn pri",onClick:u,children:"Save"})]})]})})}const gp=[{key:"state",label:"State name",placeholder:"e.g. California"},{key:"counties",label:"Counties (comma-separated)",placeholder:"e.g. San Francisco, Alameda"},{key:"city1",label:"City 1",placeholder:"e.g. San Francisco"},{key:"city2",label:"City 2 (optional)",placeholder:"e.g. Oakland"},{key:"city3",label:"City 3 (optional)",placeholder:"e.g. Berkeley"}];function Sp({jx:e,onSave:t,onClose:n}){const[r,l]=U.useState(e);function i(){dp(r),t(r),n()}return f.jsx("div",{className:"if-modal-backdrop",onClick:n,children:f.jsxs("div",{className:"if-modal",onClick:o=>o.stopPropagation(),children:[f.jsx("div",{className:"if-modal-title",children:"Jurisdiction Settings"}),f.jsx("div",{className:"if-modal-sub",children:"These values replace [YOUR STATE] and [CITY N] placeholders in scan prompts. Saved locally in your browser."}),gp.map(o=>f.jsxs("div",{className:"if-field",children:[f.jsx("div",{className:"if-field-label",children:o.label}),f.jsx("input",{className:"if-input",placeholder:o.placeholder,value:r[o.key],onChange:u=>l(s=>({...s,[o.key]:u.target.value}))})]},o.key)),f.jsxs("div",{style:{display:"flex",gap:8,justifyContent:"flex-end",marginTop:8},children:[f.jsx("button",{className:"if-btn",onClick:n,children:"Cancel"}),f.jsx("button",{className:"if-btn pri",onClick:i,children:"Save"})]})]})})}function wp({log:e,show:t,onClose:n}){return f.jsxs("div",{className:`if-log-sidebar${t?" show":""}`,children:[f.jsxs("div",{style:{display:"flex",justifyContent:"space-between",alignItems:"center",flexShrink:0},children:[f.jsx("div",{className:"if-section-label",children:"ACTIVITY LOG"}),f.jsx("button",{className:"if-btn sm if-log-btn",onClick:n,style:{minHeight:24,padding:"0 8px"},"aria-label":"Close log",children:"✕"})]}),e.length===0&&f.jsx("div",{style:{fontSize:10,color:"#7d8590",fontStyle:"italic",marginTop:6},children:"No activity yet."}),f.jsx("div",{style:{flex:1,overflowY:"auto",minHeight:0,marginTop:6},children:e.map((r,l)=>f.jsx("div",{style:{fontSize:10,color:l===0?"#8b949e":"#7d8590",marginBottom:3,lineHeight:1.5,borderLeft:`2px solid ${l===0?"#1f6feb":"transparent"}`,paddingLeft:5},children:r},l))})]})}function xp({jx:e,scanStatus:t,scanMeta:n,newOfficials:r,running:l,runScan:i,addNew:o,dismissNew:u,onConfigureJx:s}){const d=!e.state.trim(),y=Object.values(t).some(h=>h==="done");return f.jsxs("div",{style:{padding:"16px 20px"},children:[d&&f.jsxs("div",{style:{background:"#1a0e00",border:"1px solid #bb8009",borderRadius:7,padding:"12px 16px",marginBottom:16},children:[f.jsx("div",{style:{fontSize:11,color:"#e3b341",fontWeight:600,marginBottom:4},children:"Scan targets need jurisdiction"}),f.jsx("div",{style:{fontSize:10,color:"#8b949e",lineHeight:1.7,marginBottom:8},children:"Configure your state, counties, and cities so scans target the right officials."}),f.jsx("button",{className:"if-btn sm",style:{borderColor:"#bb8009",color:"#e3b341"},onClick:s,children:"Configure Jurisdiction"})]}),!d&&r.length===0&&!y&&f.jsxs("div",{style:{fontSize:10,color:"#8b949e",lineHeight:1.7,marginBottom:16},children:["Scan each category to discover current officials. New officials not already in your list will surface as add candidates. Run after each election cycle, then head to ",f.jsx("strong",{style:{color:"#f0f6fc"},children:"Officials"})," to verify, and ",f.jsx("strong",{style:{color:"#f0f6fc"},children:"Export"})," when ready."]}),kc.map(h=>{const m=t[h.id],w=n[h.id],k=r.filter(v=>v._scanId===h.id);return f.jsxs("div",{style:{marginBottom:10},children:[f.jsx("div",{className:"if-card",children:f.jsxs("div",{style:{display:"flex",justifyContent:"space-between",alignItems:"flex-start",gap:8},children:[f.jsxs("div",{style:{flex:1,minWidth:0},children:[f.jsx("div",{style:{fontSize:12,color:"#f0f6fc",fontWeight:500,marginBottom:2},children:h.label}),f.jsx("div",{style:{fontSize:10,color:"#8b949e"},children:h.desc}),w&&f.jsxs("div",{style:{fontSize:10,color:"#8b949e",marginTop:4},children:["Found ",w.total," ·"," ",f.jsx("span",{style:{color:w.confidence==="high"?"#3fb950":w.confidence==="medium"?"#e3b341":"#f85149"},children:w.confidence}),w.notes&&f.jsxs("span",{style:{color:"#e3b341"},children:[" · ",w.notes]})]})]}),f.jsxs("div",{style:{display:"flex",gap:6,alignItems:"center",flexShrink:0},children:[m==="scanning"&&f.jsx("span",{className:"if-tag",style:{borderColor:"#f59e0b",color:"#f59e0b",animation:"cs-pulse 1.4s ease-in-out infinite"},children:"SCANNING"}),m==="done"&&f.jsx("span",{className:"if-tag",style:{borderColor:"#238636",color:"#3fb950"},children:"DONE"}),m==="error"&&f.jsx("span",{className:"if-tag",style:{borderColor:"#da3633",color:"#f85149"},children:"ERROR"}),f.jsx("button",{className:"if-btn sm pri",onClick:()=>i(h.id),disabled:l||d,children:m==="done"?"↻ Re-scan":"▶ Scan"})]})]})}),k.length>0&&f.jsxs("div",{style:{margin:"3px 0 0 12px",paddingLeft:12,borderLeft:"2px solid #1f6feb"},children:[f.jsxs("div",{style:{fontSize:9,color:"#58a6ff",letterSpacing:"0.1em",margin:"6px 0 5px"},children:["NEW CANDIDATES (",k.length,")"]}),k.map(v=>f.jsxs("div",{className:"if-new-card",style:{marginBottom:4},children:[f.jsxs("div",{style:{minWidth:0},children:[f.jsx("div",{style:{fontSize:11,color:"#f0f6fc",fontWeight:500},children:v.name}),f.jsxs("div",{style:{fontSize:9,color:"#8b949e"},children:[v.title,v.district?` · D${v.district}`:""]}),(v.directEmail||v.officeEmail)&&f.jsx("div",{style:{fontSize:9,color:"#8b949e"},children:v.directEmail||v.officeEmail})]}),f.jsxs("div",{style:{display:"flex",gap:5,flexShrink:0},children:[f.jsx("button",{className:"if-btn sm grn",onClick:()=>o(v),children:"+ Add"}),f.jsx("button",{className:"if-btn sm",onClick:()=>u(v.name),children:"✕"})]})]},v.name))]}),m==="done"&&k.length===0&&w&&f.jsxs("div",{style:{margin:"3px 0 0 12px",padding:"4px 0 4px 12px",borderLeft:"2px solid #238636",fontSize:10,color:"#3fb950"},children:["✓ All ",w.total," already in list."]})]},h.id)})]})}function kp(e){return{display:"inline-flex",alignItems:"center",padding:"2px 7px",borderRadius:4,border:`1px solid ${ls[e]}`,color:ls[e],fontSize:9,letterSpacing:"0.07em",animation:e==="checking"?"cs-pulse 1.4s ease-in-out infinite":void 0}}function Cp({officials:e,allOfficials:t,activeCat:n,setActiveCat:r,selected:l,toggleSel:i,setSelected:o,running:u,runVerify:s,addManually:d,onGoDiscover:y}){const[h,m]=U.useState(new Set),w=e.map(v=>v.name);function k(v){m(D=>{const a=new Set(D);return a.has(v)?a.delete(v):a.add(v),a})}return t.length===0?f.jsxs("div",{className:"if-empty",children:[f.jsx("div",{className:"if-empty-title",children:"No officials in your list yet."}),f.jsx("div",{className:"if-empty-sub",children:"Use Discover to scan for officials, then add them here."}),f.jsxs("div",{style:{display:"flex",gap:8,justifyContent:"center",flexWrap:"wrap"},children:[f.jsx("button",{className:"if-btn pri",onClick:y,children:"Go to Discover"}),f.jsx("button",{className:"if-btn",onClick:d,children:"+ Add Manually"})]})]}):f.jsxs("div",{children:[f.jsxs("div",{style:{padding:"8px 16px",borderBottom:"1px solid #161b22",display:"flex",gap:6,alignItems:"center",flexWrap:"wrap",position:"sticky",top:0,background:"#080c10",zIndex:2},children:[f.jsx("div",{style:{display:"flex",gap:4,flexWrap:"wrap",flex:1,minWidth:0},children:up.map(v=>f.jsx("button",{className:`if-pill${n===v?" active":""}`,onClick:()=>r(v),children:v.toUpperCase()},v))}),f.jsxs("div",{style:{display:"flex",gap:4,flexShrink:0},children:[f.jsx("button",{className:"if-btn sm",onClick:()=>o(new Set(w)),disabled:u,children:"Sel All"}),f.jsx("button",{className:"if-btn sm",onClick:()=>o(new Set),disabled:u,children:"Clear"}),f.jsxs("button",{className:"if-btn sm pri",onClick:()=>s([...l]),disabled:u||l.size===0,children:["▶ (",l.size,")"]}),f.jsx("button",{className:"if-btn sm pri",onClick:()=>s(w),disabled:u,children:"▶ All"})]})]}),e.length===0&&f.jsx("div",{style:{padding:"24px 20px",textAlign:"center",color:"#8b949e",fontSize:11},children:"No officials in this category."}),e.map(v=>{const D=h.has(v.name),a=v.result;return f.jsxs("div",{children:[f.jsxs("div",{className:"if-official-row",onClick:()=>k(v.name),children:[f.jsx("input",{type:"checkbox",checked:l.has(v.name),onClick:c=>{c.stopPropagation(),i(v.name)},style:{accentColor:"#1f6feb",marginTop:2},readOnly:!0}),f.jsxs("div",{style:{minWidth:0},children:[f.jsx("div",{style:{fontSize:11,fontWeight:500,color:"#f0f6fc",whiteSpace:"nowrap",overflow:"hidden",textOverflow:"ellipsis"},children:v.name}),f.jsxs("div",{style:{fontSize:9,color:"#8b949e"},children:[v.title,v.district?` · D${v.district}`:"",v.county?` · ${v.county}`:""]}),f.jsx("div",{style:{fontSize:9,color:"#8b949e",fontStyle:"italic"},children:v.directEmail||v.officeEmail||"no email"})]}),f.jsx("div",{className:"if-official-cat",children:f.jsx("span",{style:{fontSize:9,color:"#8b949e"},children:v.category})}),f.jsx("div",{children:f.jsx("span",{style:kp(v.status),children:sp[v.status]})}),f.jsx("div",{style:{fontSize:10,color:"#30363d"},children:D?"▲":"▼"})]}),D&&a&&f.jsx("div",{className:"if-official-expand",children:a.error?f.jsxs("span",{style:{color:"#e3b341"},children:["Error: ",a.error]}):f.jsxs(f.Fragment,{children:[a.stillInOffice!==void 0&&f.jsxs("div",{children:[f.jsx("span",{style:{color:"#8b949e"},children:"IN OFFICE "}),a.stillInOffice?f.jsx("span",{style:{color:"#3fb950"},children:"Yes"}):f.jsxs("span",{style:{color:"#f85149"},children:["No",a.replacedBy?` → ${a.replacedBy}`:""]})]}),a.currentTitle&&f.jsxs("div",{children:[f.jsx("span",{style:{color:"#8b949e"},children:"TITLE "}),a.currentTitle]}),a.directEmail&&f.jsxs("div",{children:[f.jsx("span",{style:{color:"#8b949e"},children:"DIRECT "}),a.directEmail]}),a.officeEmail&&f.jsxs("div",{children:[f.jsx("span",{style:{color:"#8b949e"},children:"OFFICE "}),a.officeEmail]}),a.changes&&a.changes!=="No changes detected"&&f.jsxs("div",{children:[f.jsx("span",{style:{color:"#8b949e"},children:"CHANGES "}),f.jsx("span",{style:{color:"#58a6ff"},children:a.changes})]}),a.flags&&f.jsxs("div",{children:[f.jsx("span",{style:{color:"#8b949e"},children:"FLAGS "}),f.jsx("span",{style:{color:"#e3b341"},children:a.flags})]}),a.confidence&&f.jsxs("div",{children:[f.jsx("span",{style:{color:"#8b949e"},children:"CONF "}),f.jsx("span",{style:{color:a.confidence==="high"?"#3fb950":a.confidence==="medium"?"#e3b341":"#f85149"},children:a.confidence.toUpperCase()})]})]})})]},v.name)}),f.jsx("div",{style:{padding:"10px 18px",borderTop:"1px solid #161b22"},children:f.jsx("button",{className:"if-btn sm",onClick:d,children:"+ Add Manually"})})]})}function Ep({officials:e,exportForInviteFlow:t,exportBackup:n,exportCSV:r,importBackup:l,clearAll:i}){const o=e.filter(u=>u.status!=="left_office"&&(u.officeEmail||u.directEmail)).length;return f.jsxs("div",{style:{padding:"16px 20px",maxWidth:600},children:[f.jsxs("div",{className:"if-card",style:{marginBottom:10},children:[f.jsx("div",{style:{fontSize:12,color:"#f0f6fc",fontWeight:600,marginBottom:4},children:"Export to InviteFlow"}),f.jsxs("div",{style:{fontSize:10,color:"#8b949e",lineHeight:1.7,marginBottom:12},children:["Downloads a JSON file that InviteFlow imports directly. Includes only officials with an email address who are still in office.",e.length>0&&f.jsxs(f.Fragment,{children:[" ",f.jsxs("strong",{style:{color:o>0?"#a371f7":"#f85149"},children:[o," of ",e.length]})," officials are eligible."]})]}),f.jsxs("button",{className:"if-btn grn",onClick:t,disabled:o===0,children:["⬡ Export ",o>0?`${o} invitees`:"(no eligible officials)"," → InviteFlow JSON"]})]}),f.jsxs("div",{className:"if-card",style:{marginBottom:10},children:[f.jsx("div",{style:{fontSize:12,color:"#f0f6fc",fontWeight:600,marginBottom:4},children:"Download CSV"}),f.jsxs("div",{style:{fontSize:10,color:"#8b949e",lineHeight:1.7,marginBottom:12},children:["All ",e.length," officials with full detail columns, suitable for spreadsheet analysis."]}),f.jsx("button",{className:"if-btn",onClick:r,disabled:e.length===0,children:"↓ Download CSV"})]}),f.jsxs("div",{className:"if-card",style:{marginBottom:10},children:[f.jsx("div",{style:{fontSize:12,color:"#f0f6fc",fontWeight:600,marginBottom:4},children:"Backup & Restore"}),f.jsx("div",{style:{fontSize:10,color:"#8b949e",lineHeight:1.7,marginBottom:12},children:"Save the full ContactScout state (officials, scan history, metadata) as JSON. Restore it later or transfer to another device."}),f.jsxs("div",{style:{display:"flex",gap:8,flexWrap:"wrap"},children:[f.jsx("button",{className:"if-btn",onClick:n,disabled:e.length===0,children:"↓ Save Backup"}),f.jsx("button",{className:"if-btn",onClick:l,children:"↑ Restore from Backup"})]})]}),f.jsxs("div",{style:{borderTop:"1px solid #161b22",marginTop:20,paddingTop:16},children:[f.jsx("div",{className:"if-section-label",style:{marginBottom:10},children:"DANGER ZONE"}),f.jsxs("div",{style:{background:"#0d1117",border:"1px solid #da3633",borderRadius:8,padding:"14px 16px"},children:[f.jsx("div",{style:{fontSize:11,color:"#f0f6fc",fontWeight:500,marginBottom:4},children:"Clear all data"}),f.jsx("div",{style:{fontSize:10,color:"#8b949e",lineHeight:1.6,marginBottom:12},children:"Permanently removes all officials, scan history, and results from this browser. Cannot be undone — save a backup first."}),f.jsx("button",{className:"if-btn del",onClick:i,children:"✕ Clear All ContactScout Data"})]})]})]})}function Np(){return f.jsx(_p,{})}function _p(){const e=ap(),[t,n]=U.useState(e.officials),[r,l]=U.useState(e.newOfficials),[i,o]=U.useState(e.scanStatus),[u,s]=U.useState(e.scanMeta),[d,y]=U.useState(0),[h,m]=U.useState("All"),[w,k]=U.useState(new Set),[v,D]=U.useState(!1),[a,c]=U.useState({done:0,total:0}),[p,g]=U.useState([]),[E,j]=U.useState(()=>sessionStorage.getItem(bi)||""),[_,T]=U.useState(fp),[K,I]=U.useState(!1),[ke,He]=U.useState(!1),[jt,Nn]=U.useState(!1),Qt=U.useRef(!1),ot=U.useRef(null),Kt=U.useRef(null);U.useEffect(()=>{cp({officials:t,newOfficials:r,scanStatus:i,scanMeta:u})},[t,r,i,u]);const x=U.useCallback(C=>{const L=new Date().toLocaleTimeString();g(P=>[`[${L}] ${C}`,...P.slice(0,99)])},[]);function z(C){ot.current=C??null,I(!0)}function O(C){j(C),sessionStorage.setItem(bi,C);const L=ot.current;ot.current=null,L==null||L()}async function B(C,L){if(!E)return new Promise((P,F)=>{z(()=>B(C,L).then(P).catch(F))});try{return await vp(E,C,L)}catch(P){throw String(P).includes("Invalid API key")&&j(""),P}}async function Z(C){if(!E){z(()=>Z(C));return}Qt.current=!1,D(!0),c({done:0,total:C.length});for(let L=0;L<C.length&&!Qt.current;L++){const P=C[L],F=t.find(ee=>ee.name===P);if(F){x(`Verifying ${P}…`),n(ee=>ee.map(W=>W.name===P?{...W,status:"checking"}:W));try{const ee=`Verify for 2025-2026:
-Name: ${F.name}
-Title: ${F.title}
-Category: ${F.category}
-District: ${F.district||"N/A"}
-Email: ${F.directEmail||F.officeEmail||"none"}`,W=await B(ip,ee),te=!W.stillInOffice||W.changes!=="No changes detected"||W.flags,Gt=W.stillInOffice?te?"changed":"done":"left_office";n(Tt=>Tt.map(zt=>zt.name===P?{...zt,status:Gt,result:W,directEmail:W.directEmail||zt.directEmail,officeEmail:W.officeEmail||zt.officeEmail,officePhone:W.officePhone||zt.officePhone,title:W.currentTitle||zt.title}:zt)),x(`✓ ${P}: ${W.stillInOffice?"In office":"⚠ LEFT OFFICE"} — ${W.changes}`)}catch(ee){n(W=>W.map(te=>te.name===P?{...te,status:"error",result:{error:String(ee)}}:te)),x(`✗ ${P}: ${String(ee)}`)}c({done:L+1,total:C.length}),L<C.length-1&&await pp(700)}}D(!1),x(`Done — ${C.length} checked.`)}async function Pt(C){var L;if(!v){if(!E){z(()=>Pt(C));return}Qt.current=!1,D(!0),o(P=>({...P,[C]:"scanning"})),x(`Scanning: ${(L=kc.find(P=>P.id===C))==null?void 0:L.label}…`);try{const P=await B(op,hp(_)[C]),F=new Set(t.map(te=>te.name.toLowerCase().trim())),ee=P.officials??[],W=ee.filter(te=>!F.has(te.name.toLowerCase().trim()));o(te=>({...te,[C]:"done"})),s(te=>({...te,[C]:{total:ee.length,confidence:P.confidence,notes:P.notes}})),W.length>0?(l(te=>{const Gt=new Set(te.map(Tt=>Tt.name.toLowerCase()));return[...te,...W.filter(Tt=>!Gt.has(Tt.name.toLowerCase())).map(Tt=>({...Tt,_scanId:C,status:"pending",result:null}))]}),x(`✓ ${W.length} new found.`)):x(`✓ All ${ee.length} already in list.`),P.notes&&x(`ℹ ${P.notes}`)}catch(P){o(F=>({...F,[C]:"error"})),x(`✗ Scan: ${String(P)}`)}D(!1)}}function Ge(C){n(L=>[...L,{...C,status:"pending",result:null}]),l(L=>L.filter(P=>P.name!==C.name)),x(`+ Added ${C.name}`)}function _n(C){l(L=>L.filter(P=>P.name!==C))}function Je(C){k(L=>{const P=new Set(L);return P.has(C)?P.delete(C):P.add(C),P})}function Yt(){const C=prompt("Full name:");if(!C)return;const L=prompt("Title:","")??"",P=prompt("Email:","")??"",F=prompt("Category (US Senate/US House/Executive/State Senate/House/City Council):","City Council")??"City Council";n(ee=>[...ee,{name:C,title:L,directEmail:P,officeEmail:"",officePhone:"",district:"",county:"",category:F,status:"pending",result:null}]),x(`+ Added ${C}`)}function Cc(){const C=t.filter(L=>L.status!=="left_office"&&(L.officeEmail||L.directEmail)).map(mp);if(C.length===0){x("No officials with email to export.");return}ii(new Blob([JSON.stringify({exportedAt:new Date().toISOString(),source:"ContactScout",count:C.length,invitees:C},null,2)],{type:"application/json"}),`contactscout-invitees-${oi()}.json`),x(`✓ Exported ${C.length} invitees for InviteFlow.`)}function Ec(){ii(new Blob([JSON.stringify({exportedAt:new Date().toISOString(),source:"ContactScout",officials:t,newOfficials:r,scanStatus:i,scanMeta:u},null,2)],{type:"application/json"}),`contactscout-backup-${oi()}.json`),x(`✓ Backup saved (${t.length} officials).`)}function Nc(){const C=["Name","Title","District","County","Category","Direct Email","Office Email","Phone","Status","Changes","Flags","Replaced By","Confidence"],L=t.map(F=>{var ee,W,te,Gt;return[F.name,F.title,F.district,F.county,F.category,F.directEmail,F.officeEmail,F.officePhone,F.status,((ee=F.result)==null?void 0:ee.changes)||"",((W=F.result)==null?void 0:W.flags)||"",((te=F.result)==null?void 0:te.replacedBy)||"",((Gt=F.result)==null?void 0:Gt.confidence)||""]}),P=[C,...L].map(F=>F.map(ee=>`"${(ee||"").replace(/"/g,'""')}"`).join(",")).join(`
-`);ii(new Blob([P],{type:"text/csv"}),`contactscout-officials-${oi()}.csv`),x("✓ CSV exported.")}function _c(C){const L=new FileReader;L.onload=()=>{try{const P=JSON.parse(L.result);P.officials&&n(P.officials.map(F=>({...F,result:F.result||null}))),P.newOfficials&&l(P.newOfficials),P.scanStatus&&o(P.scanStatus),P.scanMeta&&s(P.scanMeta),x("✓ Backup loaded.")}catch(P){x(`✗ Import failed: ${String(P)}`)}},L.readAsText(C)}function jc(){confirm("Clear all ContactScout data? This cannot be undone.")&&(n([]),l([]),o({}),s({}),localStorage.removeItem("contactscout_state"),x("✓ All data cleared."))}const Pc=h==="All"?t:t.filter(C=>C.category===h),jn={total:t.length,done:t.filter(C=>C.status==="done").length,changed:t.filter(C=>C.status==="changed").length,left:t.filter(C=>C.status==="left_office").length,ready:t.filter(C=>C.status!=="left_office"&&(C.officeEmail||C.directEmail)).length},Tc=a.total?Math.round(a.done/a.total*100):0,yr=!_.state.trim(),zc=[`DISCOVER${r.length>0?` · ${r.length} NEW`:""}`,`OFFICIALS${t.length>0?` · ${t.length}`:""}`,"EXPORT"],Lc=[["TOTAL",jn.total,"#8b949e"],["VERIFIED",jn.done,"#3fb950"],["CHANGED",jn.changed,"#58a6ff"],["LEFT OFFICE",jn.left,"#f85149"],["READY",jn.ready,"#a371f7"]];return f.jsxs("div",{className:"if-root",children:[f.jsx("input",{ref:Kt,type:"file",accept:".json",style:{display:"none"},onChange:C=>{var P;const L=(P=C.target.files)==null?void 0:P[0];L&&(_c(L),C.target.value="")}}),f.jsxs("div",{style:{padding:"8px 16px",borderBottom:"1px solid #21262d",display:"flex",justifyContent:"space-between",alignItems:"center",gap:8,flexShrink:0},children:[f.jsxs("div",{children:[f.jsx("span",{style:{fontWeight:800,fontSize:15,color:"#f0f6fc",letterSpacing:"-0.02em"},children:"CONTACT SCOUT"}),f.jsx("div",{style:{fontSize:9,color:"#7d8590",letterSpacing:"0.1em",marginTop:1},children:"by Lenya Chan"})]}),f.jsxs("div",{style:{display:"flex",gap:6,alignItems:"center"},children:[f.jsx("button",{className:"if-btn sm if-log-btn",onClick:()=>Nn(C=>!C),children:"☰ Log"}),f.jsxs("button",{className:"if-btn sm",onClick:()=>He(!0),style:yr?{borderColor:"#bb8009",color:"#e3b341"}:{},title:yr?"Configure jurisdiction":`Jurisdiction: ${_.state}`,children:["⊙ ",yr?"Jurisdiction":_.state]}),f.jsxs("button",{className:`if-btn sm${E?" grn":" del"}`,onClick:()=>z(),children:["⚙ Key",E?" ✓":""]}),v&&f.jsx("button",{className:"if-btn sm del",onClick:()=>{Qt.current=!0},children:"■ Stop"})]})]}),!E&&f.jsx("div",{className:"if-banner info",children:f.jsxs("div",{className:"if-banner-body",children:[f.jsx("div",{className:"if-banner-title",children:"Welcome to ContactScout"}),f.jsx("div",{className:"if-banner-text",children:"Add your Claude API key to start discovering officials. Get a free key at console.anthropic.com → API Keys → Create Key."}),f.jsx("button",{className:"if-btn pri sm",onClick:()=>z(),children:"Add Claude API Key"})]})}),E&&yr&&f.jsx("div",{className:"if-banner warn",children:f.jsxs("div",{className:"if-banner-body",children:[f.jsx("div",{className:"if-banner-title",children:"Jurisdiction not configured"}),f.jsx("div",{className:"if-banner-text",children:"Configure your state, counties, and cities so scan prompts target the right officials."}),f.jsx("button",{className:"if-btn sm",style:{borderColor:"#bb8009",color:"#e3b341"},onClick:()=>He(!0),children:"Configure Jurisdiction"})]})}),t.length>0&&f.jsxs("div",{className:"if-stats",children:[Lc.map(([C,L,P])=>f.jsxs("div",{className:"if-stat",children:[f.jsx("span",{className:"if-stat-dot",style:{background:P}}),f.jsx("span",{className:"if-stat-label",children:C}),f.jsx("span",{className:"if-stat-val",style:{color:P},children:L})]},C)),v&&f.jsxs("div",{className:"if-progress-wrap",children:[f.jsx("div",{className:"if-progress-track",children:f.jsx("div",{className:"if-progress-fill",style:{width:`${Tc}%`}})}),f.jsxs("span",{style:{fontSize:10,color:"#58a6ff"},children:[a.done,"/",a.total]})]})]}),f.jsx("div",{className:"if-tab-bar",children:zc.map((C,L)=>f.jsx("button",{className:`if-tab${d===L?" active":""}`,onClick:()=>y(L),children:C},L))}),f.jsxs("div",{className:"if-layout",children:[f.jsxs("div",{className:"if-main",children:[d===0&&f.jsx(xp,{jx:_,scanStatus:i,scanMeta:u,newOfficials:r,running:v,runScan:Pt,addNew:Ge,dismissNew:_n,onConfigureJx:()=>He(!0)}),d===1&&f.jsx(Cp,{officials:Pc,allOfficials:t,activeCat:h,setActiveCat:m,selected:w,toggleSel:Je,setSelected:k,running:v,runVerify:Z,addManually:Yt,onGoDiscover:()=>y(0)}),d===2&&f.jsx(Ep,{officials:t,exportForInviteFlow:Cc,exportBackup:Ec,exportCSV:Nc,importBackup:()=>{var C;return(C=Kt.current)==null?void 0:C.click()},clearAll:jc})]}),f.jsx(wp,{log:p,show:jt,onClose:()=>Nn(!1)})]}),K&&f.jsx(yp,{apiKey:E,onSave:O,onClose:()=>{I(!1),ot.current=null}}),ke&&f.jsx(Sp,{jx:_,onSave:C=>T(C),onClose:()=>He(!1)})]})}Sc(document.getElementById("root")).render(f.jsx(U.StrictMode,{children:f.jsx(Np,{})}));
