--- src/contactscout/built/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>ContactScout</title>
-    <style>
-      * { box-sizing: border-box; margin: 0; padding: 0; }
-      body { background: #080c10; color: #c9d1d9; font-family: 'Courier New', monospace; height: 100dvh; overflow: hidden; }
-      ::-webkit-scrollbar { width: 5px; }
-      ::-webkit-scrollbar-track { background: #0d1117; }
-      ::-webkit-scrollbar-thumb { background: #21262d; border-radius: 3px; }
-    </style>
-    <script type="module" crossorigin src="./assets/index-CzzvB3Ea.js"></script>
-    <link rel="stylesheet" crossorigin href="./assets/index-CU0T3FAk.css">
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
