--- src/contactscout/src/css.ts
@@ -1,2 +1,0 @@
-/* Styles are now imported from styles.css — see src/main.tsx */
-export const CSS = '';
