--- src/contactscout/vite.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-
-export default defineConfig({
-  plugins: [react()],
-  base: './',
-  server: { port: 5174 },
-});
