+++ src/scout/constants.ts
