+++ src/scout/emailPatterns.ts
