+++ src/scout/search.ts
