+++ src/scout/types.ts
