+++ src/scout/utils.ts
