+++ tsconfig.json
@@ -17,5 +17,5 @@
     "noFallthroughCasesInSwitch": true
   },
   "include": ["src"],
-  "exclude": ["src/contact-scout", "src/contactscout"]
+  "exclude": ["src/contact-scout"]
 }
