+++ vite.config.ts
@@ -32,7 +32,6 @@ export default defineConfig(({ mode }) => {
       rollupOptions: {
         input: {
           inviteflow: 'src/inviteflow/index.html',
-          contactscout: 'src/contactscout/index.html',
         },
         output: {
           entryFileNames: 'assets/[name]-[hash].js',
